### 2025-07-14 0a765d5
Change shot domain indices from 1,0,-1 to 1,2,3
