### 2025-07-15 066790e
Fix power_supply_railed check for d3d
