### 2025-07-15 6fdd4ed
Move thresholds to machine-specific config.toml files
