+++ disruption_py/settings/domain_setting.py
@@ -1,410 +1,0 @@
-#!/usr/bin/env python3
-
-"""
-This module provides classes and methods to specify different domain settings
-for the timebase used, including full, flattop, and ramp-up domains.
-"""
-
-from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from typing import Dict, Union
-
-import numpy as np
-from loguru import logger
-from MDSplus import mdsExceptions
-
-from disruption_py.core.physics_method.params import PhysicsMethodParams
-from disruption_py.core.utils.enums import map_string_to_enum
-from disruption_py.core.utils.math import interp1
-from disruption_py.core.utils.misc import shot_msg_patch
-from disruption_py.machine.cmod.physics import CmodPhysicsMethods
-from disruption_py.machine.d3d.util import D3DUtilMethods
-from disruption_py.machine.east.physics import EastPhysicsMethods
-from disruption_py.machine.tokamak import Tokamak
-
-DomainSettingType = Union["DomainSetting", str, Dict[Tokamak, "DomainSettingType"]]
-
-
-@dataclass
-class DomainSettingParams:
-    """
-    Parameters passed to the `get_domain` method.
-
-    Attributes
-    ----------
-    physics_method_params : PhysicsMethodParams
-        The physics method parameters that include timebase information.
-    tokamak : Tokamak
-        The tokamak instance for which the domain is set.
-    logger : Logger
-        Logger instance for logging relevant messages.
-    """
-
-    physics_method_params: PhysicsMethodParams
-    tokamak: Tokamak
-
-    def __post_init__(self):
-        self.logger = shot_msg_patch(logger, self.physics_method_params.shot_id)
-
-
-class DomainSetting(ABC):
-    """
-    Abstract base class for domain settings that modify the timebase
-    domain in physics method parameters.
-    """
-
-    def get_domain(self, params: DomainSettingParams) -> np.ndarray:
-        """
-        Get the domain of the timebase to use.
-
-        Parameters
-        ----------
-        params : DomainSettingParams
-            The parameters including the physics method and tokamak information.
-
-        Returns
-        -------
-        np.ndarray
-            Array representing the timebase domain.
-        """
-        if hasattr(self, "tokamak_overrides"):
-            if params.tokamak in self.tokamak_overrides:
-                return self.tokamak_overrides[params.tokamak](params)
-        return self._get_domain(params)
-
-    @abstractmethod
-    def _get_domain(self, params: DomainSettingParams) -> np.ndarray:
-        """
-        Abstract method to get the modified times attribute of the physics method
-        parameters based on the domain.
-
-        Parameters
-        ----------
-        params : DomainSettingParams
-            The parameters for domain modification.
-
-        Returns
-        -------
-        np.ndarray
-            Array of modified times.
-        """
-
-
-class DomainSettingDict(DomainSetting):
-    """
-    Domain setting class that handles a dictionary of tokamak-domain mappings.
-
-    Parameters
-    ----------
-    domain_setting_dict : dict[Tokamak, DomainSettingType]
-        A dictionary mapping tokamak type strings to the desired domain setting for
-        that tokamak.  E.g. `{'cmod': 'flattop'}`. Any other option passable to the
-        `domain_setting` parameter in `RetrievalSettings` may be used.
-    """
-
-    def __init__(self, domain_setting_dict: Dict[Tokamak, DomainSettingType]):
-        """
-        Initialize DomainSettingDict with a dictionary of tokamak-domain settings.
-
-        Parameters
-        ----------
-        domain_setting_dict : dict
-            Dictionary of tokamak to domain setting mappings.
-        """
[…]
