### 2025-07-16 0676bb5
replace shot domain with time domain
