+++ disruption_py/machine/east/config.toml
@@ -7,7 +7,7 @@ driver = "MySQL"
 host = "202.127.205.10"
 port = 3306
 
-[east.physics.shot_domain_thresholds]
+[east.physics.time_domain_thresholds]
 dipprog_dt = 1e3
 
 [east.tests]
