### 2025-07-16 dcd1b90
replace shot domain with time domain
