### 2025-07-16 6accf07
Adding ability to pull efit equilibrium on C-Mod
