### 2025-07-24 d5e1001
Adding methods to pull mirnov data from D3D
