+++ disruption_py/machine/d3d/mirnov.py
@@ -1,0 +1,184 @@
+#!/usr/bin/env python3
+
+"""
+Module for retrieving and processing EFIT parameters for CMOD.
+"""
+
+import numpy as np
+from MDSplus import mdsExceptions
+from scipy.signal import ShortTimeFFT
+from scipy.signal.windows import kaiser
+import xarray as xr
+
+from disruption_py.core.physics_method.decorator import physics_method, cache_method
+from disruption_py.core.physics_method.params import PhysicsMethodParams
+from disruption_py.core.utils.math import interp1
+from disruption_py.machine.tokamak import Tokamak
+from disruption_py.machine.cmod.mirnov import setup_fft
+
+class D3DMirnovMethods:
+    """
+    Class for retrieving and processing Mirnov-related signals for D3D
+    """
+
+    @staticmethod
+    def get_mirnov_names_and_locations(params: PhysicsMethodParams, debug=False):
+        """Get the names and locations of the Mirnov coils in the ANALYSIS tree.
+
+        Parameters
+        ----------
+        params : PhysicsMethodParams
+            The parameters for the physics method.
+
+        Returns
+        -------
+        all_mirnov_names : list[str]
+            The names of all the Mirnov coils.
+        phi_all : numpy.ndarray
+            The toroidal angles of the Mirnov coils
+        theta_all : numpy.ndarray
+            The poloidal angles of the Mirnov coils
+        theta_pol_all : numpy.ndarray
+            The poloidal offset angles of the Mirnov coils.
+        """
+        
+    @staticmethod
+    def get_mirnov_fft(params: PhysicsMethodParams, mirnov_name: str, freq_resolution: float = 100, max_freq: float = 80e3):
+        """Get and the interpolated fft of a Mirnov coil.
+        
+        This will work best if the params timebase is uniform.
+        """
+        path = r"\magnetics::top.active_mhd.signals"
+        try:
+            mirnov_signal, mirnov_times = params.mds_conn.get_data_with_dims(
+                        path=f"{path}.{mirnov_name}",
+                        tree_name="magnetics",
+                        astype="float32",
+            )
+            # Get the sampling frequency of the Mirnov signal
+            f_mirnov = 1 / np.mean(np.diff(mirnov_times))
+            params.logger.verbose(f"Using Mirnov frequency of {f_mirnov} Hz on {mirnov_name}")
+            if not np.isclose(f_mirnov, 2.5e6, rtol=0.01) and not np.isclose(f_mirnov, 5e6, rtol=0.01):
+                params.logger.warning(f"[Shot {params.shot_id}] Got Mirnov frequency of {f_mirnov} Hz on {mirnov_name}, expected 2.5 MHz or 5 MHz")
+
+            # Set up the fft taking into account the params timebase
+            f_timebase = 1 / np.mean(np.diff(params.times))
+            sft = setup_fft(f_mirnov=f_mirnov, f_target=f_timebase, frequency_resolution=freq_resolution)
+            f_indices = np.where(sft.f < max_freq)
+            freqs = sft.f[f_indices]
+
+            mirnov_fft_full = sft.stft(mirnov_signal)[f_indices]
+
+            # Interpolate the STFFTs to the timebase
+            fft_times = (sft.delta_t * np.arange(mirnov_fft_full.shape[1])) + mirnov_times[0]
+            mirnov_fft_interp = interp1(fft_times, mirnov_fft_full, params.times)
+
+            # Check if the average difference between frequencies is NOT close to the frequency resolution
+            if not np.isclose(np.mean(np.diff(freqs)), freq_resolution, atol=1):
[…]
