+++ disruption_py/machine/d3d/mirnov.py
@@ -16,6 +16,347 @@ from disruption_py.core.utils.math import interp1
 from disruption_py.machine.tokamak import Tokamak
 from disruption_py.machine.cmod.mirnov import setup_fft
 
+# DIII-D Magnetic Sensors Dictionary
+# Data extracted from "Magnetic Diagnostics – Coordinates Of The Sensors" document
+# Ted Strait and S. Munaretto - Last Update 2020-05-05
+D3D_PROBES_BP = {
+    # Bp - 322 degree poloidal array
+    'MPI11M322': {'R': 0.973, 'Z': -0.002, 'phi': 322.5, 'gamma': 89.9, 'L': 0.115, 'W': 0.0556, 'NA': 0.04977, 'type': 'Bp'},
+    'MPI1A322': {'R': 0.974, 'Z': 0.182, 'phi': 322.5, 'gamma': 90.0, 'L': 0.140, 'W': 0.0556, 'NA': 0.05798, 'type': 'Bp'},
+    'MPI2A322': {'R': 0.974, 'Z': 0.512, 'phi': 322.5, 'gamma': 89.7, 'L': 0.140, 'W': 0.0556, 'NA': 0.05929, 'type': 'Bp'},
+    'MPI3A322': {'R': 0.975, 'Z': 0.850, 'phi': 322.5, 'gamma': 90.2, 'L': 0.140, 'W': 0.0556, 'NA': 0.05843, 'type': 'Bp'},
+    'MPI4A322': {'R': 0.972, 'Z': 1.161, 'phi': 322.5, 'gamma': 90.5, 'L': 0.140, 'W': 0.0556, 'NA': 0.05851, 'type': 'Bp'},
+    'MPI5A322': {'R': 1.051, 'Z': 1.330, 'phi': 322.5, 'gamma': 44.6, 'L': 0.141, 'W': 0.0556, 'NA': 0.05891, 'type': 'Bp'},
+    'MPI8A322': {'R': 1.219, 'Z': 1.406, 'phi': 322.5, 'gamma': 0.3, 'L': 0.139, 'W': 0.0556, 'NA': 0.06072, 'type': 'Bp'},
+    'MPI89A322': {'R': 1.402, 'Z': 1.407, 'phi': 322.5, 'gamma': 0.7, 'L': 0.115, 'W': 0.0556, 'NA': 0.04833, 'type': 'Bp'},
+    'MPI9A322': {'R': 1.584, 'Z': 1.408, 'phi': 322.5, 'gamma': 0.6, 'L': 0.140, 'W': 0.0556, 'NA': 0.05917, 'type': 'Bp'},
+    'MPI79FA322': {'R': 1.783, 'Z': 1.323, 'phi': 322.5, 'gamma': -39.3, 'L': 0.141, 'W': 0.0556, 'NA': 0.05963, 'type': 'Bp'},
+    'MPI79NA322': {'R': 1.924, 'Z': 1.206, 'phi': 322.5, 'gamma': -39.2, 'L': 0.114, 'W': 0.0556, 'NA': 0.04792, 'type': 'Bp'},
+    'MPI7FA322': {'R': 2.067, 'Z': 1.090, 'phi': 322.5, 'gamma': -39.4, 'L': 0.140, 'W': 0.0556, 'NA': 0.05700, 'type': 'Bp'},
[…]
