### 2025-07-25 88d8004
Able to get Sxx for a single Mirnov probe
