+++ disruption_py/settings/time_setting.py
@@ -694,7 +694,12 @@ class MirnovTimeSetting(TimeSetting):
             )
 
         max_time = np.max(efit_time)
-        mirnov_time = np.round(np.arange(0, max_time + 1e-4, 1e-4), 4)
+        if params.tokamak == Tokamak.CMOD:
+            mirnov_time = np.round(np.arange(0, max_time + 1e-4, 1e-4), 4)
+        if params.tokamak == Tokamak.D3D:
+            # DIII-D times are always reported in ms and this gets converted elsewhere
+            mirnov_time = np.round(np.arange(0, max_time + 1e-1, 1e-1), 1)
+
         return _postprocess(times=mirnov_time, units=efit_time_unit)
     
 class TMDBTimeSetting(TimeSetting):
