### 2025-07-25 a9470c3
Able to get Sxx for a single Mirnov probe
