### 2025-07-25 5500c44
Better logging in D3D Mirnov methods, fixed TMDB timebase
