### 2025-07-25 9d6604d
Better logging in D3D Mirnov methods, fixed TMDB timebase
