### 2025-07-28 a8f7a27
expanded analysis to disruption table
