### 2025-07-30 73d3a97
read efit21 rather than efit18
