### 2025-07-30 788e1e6
read efit21 rather than efit18
