+++ disruption_py/settings/nickname_setting.py
@@ -256,7 +256,7 @@ class DisruptionNicknameSetting(NicknameSetting):
 
         efit_trees = params.database.query(
             "select tree from code_rundb.dbo.plasmas where "
-            f"shot = {params.shot_id} and runtag = '{runtag}' and deleted = 0 order by idx",
+            f"shot = {params.shot_id} and runtag = 'DISPY' and deleted = 0 order by idx",
             use_pandas=False,
         )
         if len(efit_trees) == 0:
@@ -281,7 +281,7 @@ class DisruptionNicknameSetting(NicknameSetting):
             return "efit21"
         if params.disruption_time is None:
             return DefaultNicknameSetting().get_tree_name(params)
-        return "efit18"
+        return "efit21"
 
     def _get_tree_name(self, params: NicknameSettingParams) -> str:
         """
