### 2025-08-08 3269603
only using gpc2 plut some other outputs
