+++ drafts/test_te_profile_params_ece.py
@@ -45,7 +45,7 @@ def read_dpy(shot_list, local_data_dir, efit_name, bypass=False, save_data=True,
 	data = get_shots_data(
 	    shotlist_setting=shot_list,
 	    retrieval_settings=retrieval_settings,
-	    log_settings=LogSettings(console_level="WARNING"),
+	    log_settings=LogSettings(console_level="DEBUG"),
 	    output_setting="dataframe",
 	 	num_processes = 20,
 	)
@@ -58,41 +58,46 @@ def read_dpy(shot_list, local_data_dir, efit_name, bypass=False, save_data=True,
 
 def manually_check_profiles():
 	
-	shot_list = [1100910023]
-	time_slices = [0.55, 0.88, 1.0]
+	shot_list = [1160930034]
+	time_slices = [0.36, 1.0, 1.32, 1.335]
 	#time_slices = [1.62, 1.64, 1.66]
 	#time_slices = [0.26, 0.28, 0.3]
 	local_data_dir = '/home/henrycw/projects/disruption-py/drafts/local_data/'
-	efit_name='analysis'
+	efit_name='efit21'
 
 	# Pass time slice to be read by physics method
 	np.savetxt(f"{local_data_dir}time_slices.txt", time_slices)
 
 	df = read_dpy(shot_list=shot_list, local_data_dir=local_data_dir, efit_name=efit_name,
 					bypass=True, save_data=False)
+	print(df[['te_width_ece', 'te_core_vs_avg_ece', 'te_edge_vs_avg_ece', 'te0', 'te_width_err']])
 	te_prof = pd.read_csv(f'{local_data_dir}te_prof.csv')
+	print(te_prof)
 
-	fig, axs = plt.subplots(3, 2, sharex='col')
-	axs[0][0].plot(df['time'], df['te_width_ece'], marker='o')
+	fig, axs = plt.subplots(5, 2, sharex='col')
+	axs[0][0].plot(df['time'], df['te_width_ece'], marker='.')
 	axs[0][0].set_ylabel("Te width [m]")
-	axs[1][0].plot(df['time'], df['te_core_vs_avg_ece'], marker='o')
+	axs[1][0].plot(df['time'], df['te_core_vs_avg_ece'], marker='.')
 	axs[1][0].set_ylabel("Core PF")
-	axs[2][0].plot(df['time'], df['te_edge_vs_avg_ece'], marker='o')
+	axs[2][0].plot(df['time'], df['te_edge_vs_avg_ece'], marker='.')
 	axs[2][0].set_ylabel("Edge PF")
 	axs[2][0].set_xlabel("time [s]")
+	axs[3][0].plot(df['time'], df['te0'], marker='.')
+	axs[4][0].plot(df['time'], df['te_width_err'])
 	for i, t in enumerate(time_slices):
 		print(i)
 		prof = te_prof[te_prof['time']==t]
-		axs[i][1].plot(prof['r'], prof['te'], marker='o', c='k', linestyle='none')
-		axs[i][1].plot(prof['r'], prof['te_fit'], c='r', label='Fit')
-		#axs[i][1].plot(prof['r'], prof['te_guess'], c='g')
-		axs[i][1].axvline(prof['r0'].to_numpy()[0], c='gray', label='R0')
-		axs[i][1].axvline(prof['r0'].to_numpy()[0] + prof['a'].to_numpy()[0], c='gray', linestyle='--', label='R0 + a')
-		axs[i][1].set_ylabel("Te [keV]")
-		axs[i][1].set_title(f"t = {t:.3} s")
-		axs[i][1].set_ylim(0, np.max(te_prof['te'].to_numpy())+0.5)
[…]
