+++ disruption_py/machine/cmod/physics.py
@@ -1443,6 +1443,7 @@ class CmodPhysicsMethods:
 
     @staticmethod
     def _get_te_profile_params_ece(
+        params,
         times,
         gpc1_te_data,
         gpc1_te_time,
@@ -1487,7 +1488,7 @@ class CmodPhysicsMethods:
           is much LOWER than Thomson Scattering, for some portion of the LFS profile
           (typically starting around r/a 0.8?). In this case ECE cannot be used." An
           example shot with ECE cutoffs is 1140226024 (Calibration of Thomson density
-          using ECE cutoffs). Because the critical density is proportional to B^2,
+          using ECE cutoffs), 1160930034. Because the critical density is proportional to B^2,
           shots with B = 5.4 T on axis would need to have very high densities to
           experience a cutoff in the profile. We could look for cutoffs by comparing
           the B profile to the ne profile and checking that ne < ncrit throughout the
@@ -1576,20 +1577,20 @@ class CmodPhysicsMethods:
         
 
         # TODO: Delete this preamble code
-        # import pandas as pd 
-        # from disruption_py.core.utils.math import gauss
+        import pandas as pd 
+        from disruption_py.core.utils.math import gauss
         # import matplotlib.pyplot as plt
-        # local_data_dir = '/home/henrycw/projects/disruption-py/drafts/local_data/'
-        # time_slices = np.genfromtxt(f"{local_data_dir}time_slices.txt", dtype=float)
-        # dfs_to_plot = []
+        local_data_dir = '/home/henrycw/projects/disruption-py/drafts/local_data/'
+        time_slices = np.genfromtxt(f"{local_data_dir}time_slices.txt", dtype=float)
+        dfs_to_plot = []
 
         # Constants
         core_bound_factor = 0.2
         edge_bound_factor = 0.8
-        min_okay_channels = 9
+        min_okay_channels = 3
         min_te = 0.02  # [keV]
         min_btor = 4.5  # [T]
-        max_lh_power = 1.0  # [kW]
+        max_lh_power = 10000.0  # [kW]
         min_r_to_avoid_harmonic_overlap = 0.6  # [m]
         rising_tail_factor = 1.2
 
@@ -1604,17 +1605,17 @@ class CmodPhysicsMethods:
         # Interpolate GPC data onto efit timebase. Timebase for radial measurements is
         # slower than efit but radial positions are approx. stable so linear
         # interpolation is safe.
-        n_channels = gpc1_te_data.shape[0]
-        gpc1_te = np.full((n_channels, len(efit_time)), np.nan)
-        gpc1_rad = np.full((n_channels, len(efit_time)), np.nan)
-        for i in range(n_channels):
-            gpc1_te[i, :] = interp1(gpc1_te_time[i, :], gpc1_te_data[i, :], efit_time)
-            if len(gpc1_rad_data[i, :]) > 1:
-                gpc1_rad[i, :] = interp1(
-                    gpc1_rad_time[i, :], gpc1_rad_data[i, :], efit_time
-                )
-            else:
-                gpc1_rad[i, :] = np.full(len(efit_time), np.nan)
+        # n_channels = gpc1_te_data.shape[0]
+        # gpc1_te = np.full((n_channels, len(efit_time)), np.nan)
+        # gpc1_rad = np.full((n_channels, len(efit_time)), np.nan)
+        # for i in range(n_channels):
+        #     gpc1_te[i, :] = interp1(gpc1_te_time[i, :], gpc1_te_data[i, :], efit_time)
+        #     if len(gpc1_rad_data[i, :]) > 1:
+        #         gpc1_rad[i, :] = interp1(
[…]
