### 2025-08-11 15c5ba4
tweak image size
