+++ drafts/test_te_profile_params_ece.py
@@ -4,11 +4,25 @@
 import os
 
 import matplotlib.pyplot as plt
+import matplotlib.colors as colors
 import numpy as np
 import pandas as pd
 
 from disruption_py.settings import LogSettings, RetrievalSettings
 from disruption_py.workflow import get_shots_data
+from disruption_py.core.utils.math import gauss
+
+# TODO: Expand the batch run to full list of shots
+# TODO: Make plots and write up changes
+# TODO: Commit. Then clean and lint code
+# TODO: Make pull request
+
+# Shots to make plots:
+# 1120824024 (example of radiative collapse with edge cooling)
+# 1120808010 (Example of some issues: cutoff)
+# 1151001017 Large sawteeth at 0.665, 0.668 s
+# 1140416021- Shot with LH power and non-thermal emission after LH power turned of-- caught by sanity check
+# 1140725001- Example of issues due to weird tansients in channels [0.4, 0.8, 1.5, 1.62, 1.63]
 
 
 def read_dpy(shot_list, local_data_dir, efit_name, bypass=False, save_data=True, only_columns=None):
@@ -58,8 +72,8 @@ def read_dpy(shot_list, local_data_dir, efit_name, bypass=False, save_data=True,
 
 def manually_check_profiles():
 	
-	shot_list = [1160930034]
-	time_slices = [0.36, 1.0, 1.32, 1.335]
+	shot_list = [1120824024]
+	time_slices = [0.6, 0.614, 0.620, 0.624, 0.625]
 	#time_slices = [1.62, 1.64, 1.66]
 	#time_slices = [0.26, 0.28, 0.3]
 	local_data_dir = '/home/henrycw/projects/disruption-py/drafts/local_data/'
@@ -70,27 +84,37 @@ def manually_check_profiles():
 
 	df = read_dpy(shot_list=shot_list, local_data_dir=local_data_dir, efit_name=efit_name,
 					bypass=True, save_data=False)
-	print(df[['te_width_ece', 'te_core_vs_avg_ece', 'te_edge_vs_avg_ece', 'te0', 'te_width_err']])
+	print(df[['te_width_ece', 'te_core_vs_avg_ece', 'te_edge_vs_avg_ece', 'te0']])
 	te_prof = pd.read_csv(f'{local_data_dir}te_prof.csv')
 	print(te_prof)
 
-	fig, axs = plt.subplots(5, 2, sharex='col')
-	axs[0][0].plot(df['time'], df['te_width_ece'], marker='.')
+	fig, axs = plt.subplots(5, 2, sharex='col', figsize=(14,6))
+	axs[0][0].plot(df['time'], df['te_width_ece'], marker='.', ms=3)
 	axs[0][0].set_ylabel("Te width [m]")
-	axs[1][0].plot(df['time'], df['te_core_vs_avg_ece'], marker='.')
+	axs[1][0].plot(df['time'], df['te_core_vs_avg_ece'], marker='.', ms=2)
 	axs[1][0].set_ylabel("Core PF")
-	axs[2][0].plot(df['time'], df['te_edge_vs_avg_ece'], marker='.')
+	axs[2][0].plot(df['time'], df['te_edge_vs_avg_ece'], marker='.', ms=2)
 	axs[2][0].set_ylabel("Edge PF")
-	axs[2][0].set_xlabel("time [s]")
-	axs[3][0].plot(df['time'], df['te0'], marker='.')
-	axs[4][0].plot(df['time'], df['te_width_err'])
[…]
