+++ disruption_py/machine/cmod/physics.py
@@ -1445,14 +1445,10 @@ class CmodPhysicsMethods:
     def _get_te_profile_params_ece(
         params,
         times,
-        gpc1_te_data,
-        gpc1_te_time,
-        gpc1_rad_data,
-        gpc1_rad_time,
-        gpc2_te_data,
-        gpc2_te_time,
-        gpc2_rad_data,
-        gpc2_rad_time,
+        te,
+        te_time,
+        ece_radii,
+        ece_radii_time,
         te0,
         efit_time,
         r0,
@@ -1565,16 +1561,6 @@ class CmodPhysicsMethods:
 
         Last Major Update: Henry Wietfeldt (08/28/24), (PR: #260)
         """
-        # TODO: To fix
-        # 1) Initial guess variance
-        # 2) Check if standard deviation greater than aminor
-        # 2) 1150826029, 990225017-018, 1000511012, 1100910023, 1160616024: r0 doesn't align with center of Te profile
-        #   - Add check that max of okay indices is within 2 cm of rmagx
-        # 3) 1160826029: "GPC2 grating to 1.85 mm, 0 degrees. Have not yet found and loaded a xcal so Te's will be wrong."
-        # 4) 1090914006: Serious outlier
-        #   - Add check for okay indices that Te < 2 * Te_core
-        # 5) 1110215005: Near end of shot, Btor changed, channels misaligned
-        
 
         # TODO: Delete this preamble code
         import pandas as pd 
@@ -1587,63 +1573,35 @@ class CmodPhysicsMethods:
         # Constants
         core_bound_factor = 0.2
         edge_bound_factor = 0.8
-        min_okay_channels = 3
+        min_okay_channels = 9
         min_te = 0.02  # [keV]
         min_btor = 4.5  # [T]
         max_lh_power = 10000.0  # [kW]
         min_r_to_avoid_harmonic_overlap = 0.6  # [m]
         rising_tail_factor = 1.2
 
-        # Only use EFITs starting after the GPC diagnostic has profiles.
-        if len(gpc1_rad_time) > 0:
-            efit_time = efit_time[
-                efit_time >= max(np.max(gpc1_rad_time[:, 0]), gpc2_rad_time[0])
-            ]
-        else:
-            efit_time = efit_time[efit_time >= gpc2_rad_time[0]]
-
-        # Interpolate GPC data onto efit timebase. Timebase for radial measurements is
-        # slower than efit but radial positions are approx. stable so linear
-        # interpolation is safe.
-        # n_channels = gpc1_te_data.shape[0]
-        # gpc1_te = np.full((n_channels, len(efit_time)), np.nan)
-        # gpc1_rad = np.full((n_channels, len(efit_time)), np.nan)
-        # for i in range(n_channels):
-        #     gpc1_te[i, :] = interp1(gpc1_te_time[i, :], gpc1_te_data[i, :], efit_time)
-        #     if len(gpc1_rad_data[i, :]) > 1:
-        #         gpc1_rad[i, :] = interp1(
-        #             gpc1_rad_time[i, :], gpc1_rad_data[i, :], efit_time
-        #         )
-        #     else:
-        #         gpc1_rad[i, :] = np.full(len(efit_time), np.nan)
-
-        n_channels = gpc2_te_data.shape[0]
-        gpc2_te = np.full((n_channels, len(efit_time)), np.nan)
-        gpc2_rad = np.full((n_channels, len(efit_time)), np.nan)
+        # Interpolate data onto Te timebase. The shape of the plasma (r0 and a)
+        # radial positions of ECE channels (determined by Btor) should be approx. constant
[…]
