### 2025-08-11 a0a6b37
cleanup get_te_profile_params_ece functions
