+++ disruption_py/machine/cmod/physics.py
@@ -1443,13 +1443,11 @@ class CmodPhysicsMethods:
 
     @staticmethod
     def _get_te_profile_params_ece(
-        params,
         times,
         te,
         te_time,
         ece_radii,
         ece_radii_time,
-        te0,
         efit_time,
         r0,
         aminor,
@@ -1514,24 +1512,14 @@ class CmodPhysicsMethods:
         ----------
         times : array_like
             Requested time basis
-        gpc1_te_array: array_like
-            Te measurements from GPC diagnostic
-        gpc1_te_time: array_like
-            Time basis of GPC Te measurements
-        gpc1_rad_data: array_like
-            Radial positions corresponding to GPC channels
-        gpc1_rad_time: array_like
-            Time basis of GPC channel radial positions
-        gpc2_te_array: array_like
+        te: array_like
             Te measurements from GPC2 diagnostic
-        gpc2_te_time: array_like
+        te_time: array_like
             Time basis of GPC2 Te measurements
-        gpc2_rad_data: array_like
+        ece_radii: array_like
             Radial positions corresponding to GPC2 channels
-        gpc2_rad_time: array_like
+        ece_radii_time: array_like
             Time basis of GPC2 channel radial positions
-        te0: array_like
-            Central electron temperature calculated from GPC2 system
         efit_time : array_like
             Time basis of '_efit_tree'
         r0 : array_like
@@ -1561,15 +1549,6 @@ class CmodPhysicsMethods:
 
         Last Major Update: Henry Wietfeldt (08/28/24), (PR: #260)
         """
-
-        # TODO: Delete this preamble code
-        import pandas as pd 
-        from disruption_py.core.utils.math import gauss
-        # import matplotlib.pyplot as plt
-        local_data_dir = '/home/henrycw/projects/disruption-py/drafts/local_data/'
-        time_slices = np.genfromtxt(f"{local_data_dir}time_slices.txt", dtype=float)
-        dfs_to_plot = []
-
         # Constants
         core_bound_factor = 0.2
         edge_bound_factor = 0.8
@@ -1591,7 +1570,7 @@ class CmodPhysicsMethods:
         r0 = interp1(efit_time, r0, te_time)
         aminor = interp1(efit_time, aminor, te_time)
 
-        # Extend the last radii measurement up until the last EFIT. 
+        # Extend the last radii measurement up until the last EFIT.
         # Radii depend on Bt, which should be stable until the current quench.
         indx_last_rad = np.argmax(te_time > ece_radii_time[-1]) - 1
         for i in range(len(radii)):
@@ -1619,11 +1598,6 @@ class CmodPhysicsMethods:
         (okay_time_indices,) = np.where(
             (np.abs(btor) > min_btor) & (lh_power < max_lh_power) & (te_time > ece_radii_time[0])
         )
-
-        # TODO: Delete this temporary snippet for testing
-        idx_time_slices = []
-        for t in time_slices:
-            idx_time_slices.append(np.argmin(np.abs(te_time - t)))
 
         # Main loop for calculations
         te_core_vs_avg = np.full(len(te_time), np.nan)
@@ -1653,7 +1627,6 @@ class CmodPhysicsMethods:
             if np.sum(okay_indices) > min_okay_channels:
                 r = radii[i, okay_indices]
                 y = te[i, okay_indices]
-
                 # Check that there are two raw points in the core and in the edge
                 # so that profile is properly aligned to measure peaking factors
                 # and ensure a reasonable Gaussian fit
@@ -1669,31 +1642,11 @@ class CmodPhysicsMethods:
                 guess = [y.max(), (r.max() - r.min()) / 3]
                 try:
                     pmu = r0[i]
[…]
