+++ drafts/test_te_profile_params_ece.py
@@ -1,174 +1,0 @@
-"""
-"""
-
-import os
-
-import matplotlib.pyplot as plt
-import matplotlib.colors as colors
-import numpy as np
-import pandas as pd
-
-from disruption_py.settings import LogSettings, RetrievalSettings
-from disruption_py.workflow import get_shots_data
-from disruption_py.core.utils.math import gauss
-
-# TODO: Expand the batch run to full list of shots
-# TODO: Make plots and write up changes
-# TODO: Commit. Then clean and lint code
-# TODO: Make pull request
-
-# Shots to make plots:
-# 1120824024 (example of radiative collapse with edge cooling)
-# 1120808010 (Example of some issues: cutoff)
-# 1151001017 Large sawteeth at 0.665, 0.668 s
-# 1140416021- Shot with LH power and non-thermal emission after LH power turned of-- caught by sanity check
-# 1140725001- Example of issues due to weird tansients in channels [0.4, 0.8, 1.5, 1.62, 1.63]
-
-
-def read_dpy(shot_list, local_data_dir, efit_name, bypass=False, save_data=True, only_columns=None):
-	'''
-	saves requested data to dpy_{efit_name}.h5 in the directory specified by local_data_dir
-	  - if data file already exists locally, just pulls data from that local file instead
-      - can bypass reading existing local data if bypass=True
-	  - can skip saving data if save_data = False
-	By default grabs all columns ouptut by disruption_py
-	  - can grab only certain columns by setting 'only_columns' to a list of requested columns
-	returns data as a pandas DataFrame
-	
-	Adapted from A.R. Saperstein
-	'''
-	# need to make sure list is of type 'int' and not 'numpy.int64'
-	shot_list = np.array(shot_list).tolist()
-	
-	data_path = f"{local_data_dir}dpy_{efit_name}.h5"	# location of saved data
-	data_key = "sql_data"
-
-	# check if requested data already stored locally
-	if os.path.exists(data_path) and not bypass:
-		data = pd.read_hdf(data_path, key=data_key)
-		return data
-
-	# default method for pulling disruption-py data
-	retrieval_settings = RetrievalSettings(
-		time_setting="disruption_warning",	# use the set efit's timebase
-		efit_nickname_setting=efit_name,	# set the efit
-		run_methods=None,
-		run_columns=only_columns,
-		only_requested_columns= only_columns is not None
-	)
-	data = get_shots_data(
-	    shotlist_setting=shot_list,
-	    retrieval_settings=retrieval_settings,
-	    log_settings=LogSettings(console_level="DEBUG"),
-	    output_setting="dataframe",
-	 	num_processes = 20,
-	)
-	# Reset the loguru settings
-	if save_data:
-		if not os.path.exists(local_data_dir):
-			os.mkdir(local_data_dir)
-		data.to_hdf(data_path, key=data_key, mode='w')
-	return data
-
-def manually_check_profiles():
-	
-	shot_list = [1120824024]
-	time_slices = [0.6, 0.614, 0.620, 0.624, 0.625]
-	#time_slices = [1.62, 1.64, 1.66]
-	#time_slices = [0.26, 0.28, 0.3]
[…]
