### 2025-08-12 b5d6eeb
fix max_lh_power and minor docstring reformat
