### 2025-08-12 83172ae
minor docstring reformat that I meant to include in prev commit
