### 2025-08-14 e0306c9
Check for sensors with unreasonably large signal
