### 2025-08-14 d449cf6
first pass at pulling required data from newly computed efit
