+++ disruption_py/machine/d3d/efit.py
@@ -5,6 +5,7 @@ Module for retrieving and processing EFIT parameters for DIII-D.
 """
 
 import numpy as np
+import xarray as xr
 
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
@@ -125,3 +126,254 @@ class D3DEfitMethods:
             for param in efit_data:
                 efit_data[param] = interp1(efit_time, efit_data[param], params.times)
         return efit_data
+
+    @staticmethod
+    @physics_method(
+        tokamak=Tokamak.D3D,
+    )
+    def get_geqdsk_parameters(params: PhysicsMethodParams):
+        """
+        Retrieve parameters required to create an GeqdskDataDict https://freeqdsk.readthedocs.io/en/stable/geqdsk.html
+
+        Parameters
+        ----------
+        params : PhysicsMethodParams
+            Parameters containing MDS connection and shot information.
+
+        Returns
+        -------
+        xarray.Dataset
+            A dataset containing the GEQDSK parameters.
+        """
+        # Get EFIT time array
+        efit_time = params.mds_conn.get_data(
+            r"\efit_g_eqdsk:atime", tree_name="_efit_tree"
+        ) / 1.0e3  # [ms] -> [s]
+        
+        # Scalar parameters
+        scalar_params = {
+            # Grid dimensions
+            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),  # Number of horizontal grid points
+            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),  # Number of vertical grid points
+            
+            # Plasma parameters
+            "cpasma": params.mds_conn.get_data(r"\efit_g_eqdsk:cpasma", tree_name="_efit_tree"),  # Plasma current [A]
+            "bcentr": params.mds_conn.get_data(r"\efit_g_eqdsk:bcentr", tree_name="_efit_tree"),  # Vacuum toroidal field [T]
+            "rcentr": params.mds_conn.get_data(r"\efit_g_eqdsk:rcentr", tree_name="_efit_tree"),  # Reference major radius [m]
+            
+            # Grid boundaries
+            "rdim": params.mds_conn.get_data(r"\efit_g_eqdsk:rdim", tree_name="_efit_tree"),    # Horizontal extent of grid [m]
+            "zdim": params.mds_conn.get_data(r"\efit_g_eqdsk:zdim", tree_name="_efit_tree"),    # Vertical extent of grid [m]
+            "rcentr": params.mds_conn.get_data(r"\efit_g_eqdsk:rcentr", tree_name="_efit_tree"), # R at center of grid [m]
+            "rleft": params.mds_conn.get_data(r"\efit_g_eqdsk:rleft", tree_name="_efit_tree"),   # R at left edge of grid [m]
+            "zmid": params.mds_conn.get_data(r"\efit_g_eqdsk:zmid", tree_name="_efit_tree"),     # Z at center of grid [m]
+            
+            # Magnetic axis
+            "rmaxis": params.mds_conn.get_data(r"\efit_g_eqdsk:rmaxis", tree_name="_efit_tree"), # R of magnetic axis [m]
+            "zmaxis": params.mds_conn.get_data(r"\efit_g_eqdsk:zmaxis", tree_name="_efit_tree"), # Z of magnetic axis [m]
+            "simag": params.mds_conn.get_data(r"\efit_g_eqdsk:simag", tree_name="_efit_tree"),   # Flux at magnetic axis [Wb]
+            
+            # Plasma boundary
[…]
