### 2025-08-15 e335abf
Greg's suggested fixes
