+++ disruption_py/machine/cmod/physics.py
@@ -1546,8 +1546,9 @@ class CmodPhysicsMethods:
           electron cyclotron emission cutoff data," Rev. Sci. Instrum., vol. 76, no. 5,
           p. 053506, 2005, doi: 10.1063/1.1899311.
         - https://github.com/MIT-PSFC/disruption-py/pull/260
+        - https://github.com/MIT-PSFC/disruption-py/pull/466
 
-        Last Major Update: Henry Wietfeldt (08/12/25), (PR: #)
+        Last Major Update: Henry Wietfeldt (08/12/25), (PR: #466)
         """
         # Constants
         core_bound_factor = 0.2
@@ -1604,7 +1605,7 @@ class CmodPhysicsMethods:
         # Main loop for calculations
         te_core_vs_avg = np.full(len(te_time), np.nan)
         te_edge_vs_avg = np.full(len(te_time), np.nan)
-        te_width = np.full(len(te_time), np.nan)
+        te_hwm = np.full(len(te_time), np.nan)
         for i in okay_time_indices:
             # Only consider points that are likely to accurately measure Te
             calib_indices = (te[i, :] > min_te) & (radii[i, :] < r0[i] + aminor[i])
@@ -1647,13 +1648,13 @@ class CmodPhysicsMethods:
                 try:
                     pmu = r0[i]
                     _, psigma = gaussian_fit_with_fixed_mean(pmu, r, y, guess)
-                except RuntimeError as e:
-                    if str(e).startswith("Optimal parameters not found"):
+                except RuntimeError as exc:
+                    if str(exc).startswith("Optimal parameters not found"):
                         continue
-                    raise e
+                    raise exc
                 # rescale from sigma to HWHM for sigma values that are physical
                 # https://en.wikipedia.org/wiki/Full_width_at_half_maximum
-                te_width[i] = psigma * np.sqrt(2 * np.log(2))
+                te_hwm[i] = np.abs(psigma) * np.sqrt(2 * np.log(2))
 
                 # Calculate core/edge vs. average using uniformly sampled radial basis
                 r_equal_spaced = np.linspace(r0[i], r0[i] + aminor[i], 100)
@@ -1677,18 +1678,18 @@ class CmodPhysicsMethods:
 
         te_core_vs_avg = interp1(te_time, te_core_vs_avg, times)
         te_edge_vs_avg = interp1(te_time, te_edge_vs_avg, times)
-        te_width = interp1(te_time, te_width, times)
+        te_hwm = interp1(te_time, te_hwm, times)
         # Sanity check: Te width should be positive and less than minor radius
         # Unphysical Te width indicates ECE profile may be bad so set peaking factors to NaN too
         aminor = interp1(te_time, aminor, times)
-        unphysical_mask = (te_width < 0) | (te_width > aminor)
-        te_width[unphysical_mask] = np.nan
+        unphysical_mask = (te_hwm < 0) | (te_hwm > aminor)
+        te_hwm[unphysical_mask] = np.nan
         te_core_vs_avg[unphysical_mask] = np.nan
         te_edge_vs_avg[unphysical_mask] = np.nan
         return {
             "te_core_vs_avg_ece": te_core_vs_avg,
             "te_edge_vs_avg_ece": te_edge_vs_avg,
-            "te_width_ece": te_width,
+            "te_width_ece": te_hwm,
         }
 
     @staticmethod
