### 2025-08-19 50422ff
Checks for outlier channels and density cutoffs
