+++ disruption_py/machine/cmod/physics.py
@@ -7,6 +7,7 @@ Module for retrieving and calculating data for CMOD physics methods.
 import warnings
 
 import numpy as np
+import scipy.constants as const
 from MDSplus import mdsExceptions
 
 from disruption_py.core.physics_method.caching import cache_method
@@ -1448,6 +1449,7 @@ class CmodPhysicsMethods:
         te_time,
         ece_radii,
         ece_radii_time,
+        te0,
         efit_time,
         r0,
         aminor,
@@ -1455,6 +1457,7 @@ class CmodPhysicsMethods:
         t_mag,
         lh_power,
         lh_time,
+        n_e,
     ):
         """
         Calculates Te PF and width from ECE data using the GPC2 diagnostic system.
@@ -1476,7 +1479,6 @@ class CmodPhysicsMethods:
         across the arbitrary core or edge boundary.
 
         ECE as a Te profile diagnostic can suffer from several artifacts:
-        Artifacts currently NOT explicitly checked for
         - Density cutoffs: High ne plasmas (typically H-modes) can have an ECE cutoff.
           According to Amanda Hubbard, "what you wil see is a section of profile which
           is much LOWER than Thomson Scattering, for some portion of the LFS profile
@@ -1484,11 +1486,11 @@ class CmodPhysicsMethods:
           example shot with ECE cutoffs is 1140226024 (Calibration of Thomson density
           using ECE cutoffs), 1160930034. Because the critical density is proportional
           to B^2, shots with B = 5.4 T on axis would need to have very high densities to
-          experience a cutoff in the profile. We could look for cutoffs by comparing
-          the B profile to the ne profile and checking that ne < ncrit throughout the
-          profile; however, a simpler check for now is to ignore shots with B < 4.5 T
-          and assume there are no cutoffs with B >= 4.5 T.
-        Artifacts currently checked for
+          experience a cutoff in the profile. Most cutoffs are caught shots 
+          by ignoring shots with B < 4.5 T. To identify cutoffs on higher field shots,
+          we compare the line averaged density to the critical density for a cutoff
+          at r = a/sqrt(3), which is where n(r) = <n> (line averaged) for parabolic 
+          profiles with moderate peaking. For B=5.4 T, n_crit = 4.2e20 m^-3.
         - Non-aligned grating: The gratings were usually aligned for radial coverage
           assuming Bt=5.4T. For low Bt shots (like 2.8T), sometimes the gratings were
           adjusted, sometimes not. Low Bt shots also tend to have low signal and often
@@ -1520,6 +1522,8 @@ class CmodPhysicsMethods:
             Radial positions corresponding to GPC2 channels
         ece_radii_time: array_like
             Time basis of GPC2 channel radial positions
+        te0: array_like
+            The central electron temperature inferred from the GPC2 system
         efit_time : array_like
             Time basis of '_efit_tree'
         r0 : array_like
@@ -1534,6 +1538,8 @@ class CmodPhysicsMethods:
             Lower hybrid power
         lh_time: array_like
             Time basis of lower hybrid power
+        n_e: array_like
+            Line averaged density with view of plasma core and times as time basis
 
         Returns
         -------
@@ -1559,6 +1565,8 @@ class CmodPhysicsMethods:
         max_lh_power = 1.0  # [kW]
         min_r_to_avoid_harmonic_overlap = 0.6  # [m]
         rising_tail_factor = 1.2
+        max_te_over_te0 = 2.0
+        cmod_maj_rad = 0.68 # [m]
 
[…]
