### 2025-08-20 452a413
minor bug fix for density cutoff check
