### 2025-08-20 fb76ccc
Fix get_te_profile_params_ece docstring
