+++ disruption_py/core/physics_method/runner.py
@@ -11,6 +11,7 @@ from typing import Dict
 import numpy as np
 import xarray as xr
 
+from disruption_py.inout.mds import mdsExceptions
 from disruption_py.core.physics_method.errors import CalculationError
 from disruption_py.core.physics_method.metadata import (
     BoundMethodMetadata,
@@ -175,12 +176,10 @@ def populate_method(
     physics_method_params.logger.trace("Starting method: {name}", name=name)
 
     try:
-
         result = method(params=physics_method_params)
 
     # pylint: disable-next=broad-exception-caught
     except Exception as e:
-
         # log exception
         level = "ERROR"
         if isinstance(e, mdsExceptions.MDSplusERROR):
@@ -242,7 +241,6 @@ def populate_shot(
     start_time = time.time()
     datasets = []
     for bound_method_metadata in run_bound_method_metadata:
-
         # run method
         result = populate_method(
             physics_method_params=physics_method_params,
@@ -251,7 +249,6 @@ def populate_shot(
 
         # convert non-dataset dict to dataset
         if not isinstance(result, (xr.DataArray, xr.Dataset)):
-
             times = physics_method_params.times
 
             # create data_vars dict
