+++ disruption_py/machine/cmod/efit.py
@@ -6,6 +6,7 @@ Module for retrieving and processing EFIT parameters for CMOD.
 
 import numpy as np
 
+from disruption_py.inout.mds import mdsExceptions
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
 from disruption_py.core.utils.math import interp1
