+++ disruption_py/machine/cmod/physics.py
@@ -8,6 +8,7 @@ import warnings
 
 import numpy as np
 
+from disruption_py.inout.mds import mdsExceptions
 from disruption_py.core.physics_method.caching import cache_method
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.errors import CalculationError
