+++ pyproject.toml
@@ -53,6 +53,7 @@ dependencies = [
   "numpy>=1.26.0,<2.0.0",
   "pandas>=2.2.0,<2.3.3",
   "pyodbc>=5.2.0",
+  "s3fs>=2025.7.0",
   "scipy>=1.15.0",
   "SQLAlchemy>=2.0.0",
   "tqdm>=4.67.0",
