+++ disruption_py/machine/mast/config.toml
@@ -1,11 +1,8 @@
-[mast.inout.uda]
-hello = "world"
+[mast.inout.s3]
+endpoint_url = "https://s3.echo.stfc.ac.uk"
 
-[cmod.inout.sql]
-db_name = "something"
-driver = "ODBC"
-host = "server"
-port = 1433
+[mast.inout.sql]
+driver = "dummy"
 
 [mast.tests.shots]
 example = 30420
