+++ .github/workflows/docs.yml
@@ -50,7 +50,7 @@ jobs:
         run: poetry run mkdocs build --site-dir _site
 
       - name: Upload artifact
-        uses: actions/upload-pages-artifact@v3
+        uses: actions/upload-pages-artifact@v4
 
   deploy:
     if: |
