### 2025-09-02 6103f7a
Update deps to Aug 2025 (#469)
* bump actions/checkout to v5

* bump actions/upload-pages-artifact to v4

* pin xarray version to due h5 bug

* lock deps
