### 2025-09-02 56ef7a1
set efit21 as default, but keep efit18 for testing
