### 2025-09-03 711d6bc
default to efit21 for all shots, testing fallback to efit18 only for disruptive shots
