### 2025-09-03 08537db
Refactor C-MOD get_n_equal_1_amplitude (#460)
* Refactor get_n_equal_1_amplitude

* Confirmed python arctan2 = matlab atan2

* Change minimum number of sensors to 3

* Add PR#460 to docstring

* Switch to using str.format() for logger warning

* Replace @ with np.matmul

* fix lazy formatting in logs

* Check for sensors with unreasonably large signal

---------

Co-authored-by: gtrevisan <gtrevisan@users.noreply.github.com>
