### 2025-09-03 461033f
suppress all default MDSplus casting to use native precision
