### 2025-09-03 562e39a
remove several castings to double
