### 2025-09-03 4bb8226
Move threshold value to config
