### 2025-09-04 9a8f006
Remove print statement
