### 2025-09-04 e7bcb73
Change domain settings to physics methods (#433)
* Create CMOD get_shot_domain method

* Update CMOD get_shot_domain

* Add ip_prog to D3D get_ip_parameters output

* Create D3D get_shot_domain method

* Use lowercase names

* Select the longest continous flattop phase as the actual flattop domain

* Add condition to check if a shot only has rampup

* Create EAST get_shot_domain method

* Update docstrings

* minor changes to docstrings

* Implement generic get_shot_domain and infrastructure

* Remove machine-specific get_shot_domain

* Clean up

* Fix isort

* Minor change

* Change shot doma
[…]
