+++ disruption_py/machine/cmod/config.toml
@@ -7,6 +7,10 @@ driver = "ODBC"
 host = "alcdb2"
 port = 1433
 
+[cmod.physics.time_domain_thresholds]
+dipprog_dt = 50e3
+ip_prog = 100e3
+
 [cmod.tests]
 expected_failure_columns = [
   "greenwald_fraction",
