+++ disruption_py/machine/d3d/config.toml
@@ -7,6 +7,11 @@ driver = "FreeTDS"
 host = "d3drdb"
 port = 8001
 
+[d3d.physics.time_domain_thresholds]
+dipprog_dt = 2e3
+ip_prog = 100e3
+power_supply_railed = 1
+
 [d3d.tests]
 expected_failure_columns = [
   "aminor",
