### 2025-09-04 51b150d
make the runtag for DIII-D a config parameter
