### 2025-09-04 42b32da
make the runtag for DIII-D a config parameter
