+++ disruption_py/machine/d3d/config.toml
@@ -1,3 +1,6 @@
+[d3d.efit]
+runtag = "DISPY"
+
 [d3d.inout.mds]
 mdsplus_connection_string = "atlas"
 
