### 2025-09-04 93cd200
add mdsthin to deps
