+++ pyproject.toml
@@ -48,6 +48,7 @@ classifiers = [
 dependencies = [
   "dynaconf>=3.2.0",
   "loguru>=0.7.0",
+  "mdsthin>=1.5.0",
   "netcdf4>=1.7.0",
   "numpy>=1.26.0,<2.0.0",
   "pandas>=2.2.0",
