### 2025-09-04 a072247
try MDSplus first, then fall back onto mdsthin
