### 2025-09-04 5d1dc45
try MDSplus first, then fall back onto mdsthin
