### 2025-09-04 dd32751
harmonize MDSplus imports and exceptions for C-MOD
