### 2025-09-04 1f8b5fd
harmonize MDSplus imports and exceptions for C-MOD
