### 2025-09-04 c9b5561
harmonize MDSplus imports and exceptions for DIII-D
