### 2025-09-04 7c7a91c
harmonize MDSplus imports and exceptions for DIII-D
