### 2025-09-04 4f3dcf0
fixup! harmonize MDSplus imports and exceptions for DIII-D
