### 2025-09-04 9923224
Fix C-MOD te_width_ece unphysical values (#466)
* script for plotting te profile parameters from ece

* Plotting tools

* Check for bad fit of Te ECE profile

* Check profile alignment with magnetic axis and plasma edge

* Remove logging statements from _get_te_profile_params_ece()

* only using gpc2 plut some other outputs

* Code for plots

* cleanup get_te_profile_params_ece functions

* linting

* fix max_lh_power and minor docstring reformat

* minor docstring reformat that I meant to include in prev commit

* linting

* Greg's suggested fixes

* Checks for outlier channels and density cu
[…]
