+++ disruption_py/machine/cmod/physics.py
@@ -7,6 +7,7 @@ Module for retrieving and calculating data for CMOD physics methods.
 import warnings
 
 import numpy as np
+import scipy.constants as const
 from MDSplus import mdsExceptions
 
 from disruption_py.core.physics_method.caching import cache_method
@@ -1478,14 +1479,11 @@ class CmodPhysicsMethods:
     @staticmethod
     def _get_te_profile_params_ece(
         times,
-        gpc1_te_data,
-        gpc1_te_time,
-        gpc1_rad_data,
-        gpc1_rad_time,
-        gpc2_te_data,
-        gpc2_te_time,
-        gpc2_rad_data,
-        gpc2_rad_time,
+        te,
+        te_time,
+        ece_radii,
+        ece_radii_time,
+        te0,
         efit_time,
         r0,
         aminor,
@@ -1493,10 +1491,11 @@ class CmodPhysicsMethods:
         t_mag,
         lh_power,
         lh_time,
+        n_e,
     ):
         """
-        Calculates Te PF and width from ECE data using the two GPC diagnostic systems.
-        GPC diagnostics look at the mid-plane, and each channel detects a different
+        Calculates Te PF and width from ECE data using the GPC2 diagnostic system.
+        GPC2 looks at the mid-plane, and each channel detects a different
         emitted frequency associated with the second harmonic, which depends on B and
         therefore R.
         - te_width is the half-width at half-max of a Gaussian fit of the Te profile
@@ -1514,19 +1513,18 @@ class CmodPhysicsMethods:
         across the arbitrary core or edge boundary.
 
         ECE as a Te profile diagnostic can suffer from several artifacts:
-        Artifacts currently NOT explicitly checked for
         - Density cutoffs: High ne plasmas (typically H-modes) can have an ECE cutoff.
           According to Amanda Hubbard, "what you wil see is a section of profile which
           is much LOWER than Thomson Scattering, for some portion of the LFS profile
           (typically starting around r/a 0.8?). In this case ECE cannot be used." An
           example shot with ECE cutoffs is 1140226024 (Calibration of Thomson density
-          using ECE cutoffs). Because the critical density is proportional to B^2,
-          shots with B = 5.4 T on axis would need to have very high densities to
-          experience a cutoff in the profile. We could look for cutoffs by comparing
-          the B profile to the ne profile and checking that ne < ncrit throughout the
-          profile; however, a simpler check for now is to ignore shots with B < 4.5 T
-          and assume there are no cutoffs with B >= 4.5 T.
-        Artifacts currently checked for
+          using ECE cutoffs), 1160930034. Because the critical density is proportional
+          to B^2, shots with B = 5.4 T on axis would need to have very high densities to
+          experience a cutoff in the profile. Most cutoffs are caught shots
+          by ignoring shots with B < 4.5 T. To identify cutoffs on higher field shots,
+          we compare the line averaged density to the critical density for a cutoff
+          at r = a/sqrt(3), which is where n(r) = <n> (line averaged) for parabolic
+          profiles with moderate peaking. For B=5.4 T, n_crit = 4.2e20 m^-3.
         - Non-aligned grating: The gratings were usually aligned for radial coverage
           assuming Bt=5.4T. For low Bt shots (like 2.8T), sometimes the gratings were
           adjusted, sometimes not. Low Bt shots also tend to have low signal and often
[…]
