### 2025-09-09 030f10a
No longer attempting to read units of time from MDSPlus for the TMDB timebase, determining based on tokamak instead.
