### 2025-09-09 9dd59ab
Added method to retrieve btor
