### 2025-09-09 b6669fa
Fixed earlier refactor to efit_parameter_cols not being copied throughout the file
