### 2025-09-09 bb20229
First pieces of TS implementation
