### 2025-09-09 15aff0b
First pieces of TS implementation
Close to finishing interpolation, but the rhogrid results don't quite make sense

Fixed problem with getting psi_n correct, now gotta do profile fitting

Finished TS profiles with simple interpolation
