+++ disruption_py/machine/cmod/physics.py
@@ -7,7 +7,9 @@ Module for retrieving and calculating data for CMOD physics methods.
 import warnings
 
 import numpy as np
+import xarray as xr
 import scipy.constants as const
+from scipy.interpolate import RegularGridInterpolator, interp1d
 from MDSplus import mdsExceptions
 
 from disruption_py.core.physics_method.caching import cache_method
@@ -21,7 +23,9 @@ from disruption_py.core.utils.math import (
     interp1,
 )
 from disruption_py.machine.cmod.thomson import CmodThomsonDensityMeasure
+from disruption_py.machine.cmod.efit import CmodEfitMethods
 from disruption_py.machine.tokamak import Tokamak
+
 
 
 class CmodPhysicsMethods:
@@ -1291,6 +1295,7 @@ class CmodPhysicsMethods:
         Because the TS chords have uneven spacings, measurements are first interpolated
         to an array of equally spaced vertical positions and then used to calculate
         the peaking factors.
+        TODO(ZanderKeith) I don't think this works if the TS chords are not perfectly aligned in R
 
         Parameters:
         ----------
@@ -1496,6 +1501,165 @@ class CmodPhysicsMethods:
         return CmodPhysicsMethods._get_peaking_factors(
             params.times, ts_time, ts_te, ts_ne, ts_z, efit_time, bminor, z0
         )
+
+    @staticmethod
+    @physics_method(
+        columns=["ne_rho", "te_rho"],
+        tokamak=Tokamak.CMOD,
+    )
+    def get_ts_profiles(params: PhysicsMethodParams):
+        """
+        Retrieve the electron temperature and density profiles from the Thomson
+        scattering diagnostic and interpolate them to normalized minor radius
+        """
+
+        # Ignore shots on blacklist
+        if CmodPhysicsMethods._is_on_blacklist(params.shot_id):
+            raise CalculationError("Shot is on blacklist")
+
+        # Nodes for core and edge TS data
+        # https://cmodwiki.psfc.mit.edu/index.php/Thomson_Scattering
+        if params.shot_id > 1020000000:
+            core_ne_node = ".yag_new.results.profiles.ne_rz"
+            core_Te_node = ".yag_new.results.profiles.te_rz"
+            core_z_node = ".yag_new.results.profiles.z_sorted"
+        else:
+            core_ne_node = ".yag.results.global.profile.ne_rz_t"
+            core_Te_node = ".yag.results.global.profile.te_rz_t"
+            core_z_node = ".yag.results.global.profile.z_sorted"
+        edge_ne_node = ".yag_edgets.results.ne"
+        edge_Te_node = ".yag_edgets.results.te"
+        edge_z_node = ".yag_edgets.data.fiber_z"
+        r_node = ".yag.results.param.r"
+
+        ts_te_core_keV, ts_time = params.mds_conn.get_data_with_dims(core_Te_node, tree_name="electrons") # [keV], [s]
+        ts_te_core = ts_te_core_keV * 1000 # [keV] -> [eV]
+        ts_ne_core = params.mds_conn.get_data(core_ne_node, tree_name="electrons") # [m^-3]
+        ts_z_core = params.mds_conn.get_data(core_z_node, tree_name="electrons") # [m]
+        ts_te_edge = params.mds_conn.get_data(edge_Te_node, tree_name="electrons") # [eV]
+        ts_ne_edge = params.mds_conn.get_data(edge_ne_node, tree_name="electrons") # [m^-3]
+        ts_z_edge = params.mds_conn.get_data(edge_z_node, tree_name="electrons") # [m]
[…]
