### 2025-09-10 7e49e21
Fixed problem with getting psi_n correct, now gotta do profile fitting
