### 2025-09-11 7b51a23
add snippet for HBT-EP
