### 2025-09-11 8b81243
try to fix str conversions for arrays
