### 2025-09-11 7083581
try to fix str conversions for arrays
