### 2025-09-15 7718f4e
add default tree path for HBT-EP
