+++ disruption_py/machine/hbtep/physics.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+
+"""
+Module for retrieving and calculating data for HBTEP physics methods.
+"""
+
+
+class HbtepPhysicsMethods:
+    """
+    This class provides methods to retrieve and calculate physics-related data
+    for HBTEP.
+    """
