### 2025-09-15 977cc28
do not decorate mdsthin exceptions
