### 2025-09-17 0c348bc
Update pyproj.toml, poetry/uv locks
