+++ pyproject.toml
@@ -56,7 +56,8 @@ dependencies = [
   "scipy>=1.15.0",
   "SQLAlchemy>=2.0.0",
   "tqdm>=4.67.0",
-  "xarray>=2025.1.0",
+  "xarray>=2025.1.0,<2025.8.0",
+  "zarr>=3",
 ]
 description = "An open-source physics-based Scientific Framework for Disruption Analysis of Fusion Plasmas for AI/ML applications"
 keywords = [
