### 2025-09-17 c200127
Add test info to config
