### 2025-09-17 9947b43
Refactor s3 -> xarray backend, make more generic
