### 2025-09-17 4b5051c
Refactor s3 -> xarray backend, make more generic
