### 2025-09-17 a7509c1
add verbose statement with MDSplus api package
