+++ disruption_py/workflow.py
@@ -19,7 +19,7 @@ from disruption_py.core.utils.misc import (
     get_elapsed_time,
     without_duplicates,
 )
-from disruption_py.inout.mds import ProcessMDSConnection
+from disruption_py.inout.mds import MDSPLUS_PACKAGE, ProcessMDSConnection
 from disruption_py.inout.sql import ShotDatabase
 from disruption_py.machine.tokamak import Tokamak, resolve_tokamak_from_environment
 from disruption_py.settings import RetrievalSettings
@@ -127,6 +127,7 @@ def get_shots_data(
         log_settings.reset_handlers(num_shots=len(shotlist_list))
 
     # log start
+    logger.debug("MDSplus package: {pkg}", pkg=MDSPLUS_PACKAGE)
     logger.info(
         "Starting workflow: {n:,} shot{s} / {m} process{p}",
         n=len(shotlist_list),
