### 2025-09-17 ae2af93
introduce cmod.efit.tree, reset default to efit18 for now
