+++ disruption_py/machine/d3d/config.toml
@@ -1,5 +1,5 @@
 [d3d.efit]
-runtag = "DISPY"
+runtag = "DIS"
 
 [d3d.inout.mds]
 mdsplus_connection_string = "atlas"
