### 2025-09-18 7b5789a
remove safe_cast routine and mds cast operations
