+++ disruption_py/inout/mds.py
@@ -12,7 +12,7 @@ import numpy as np
 from loguru import logger
 
 from disruption_py.config import config
-from disruption_py.core.utils.misc import safe_cast, shot_msg
+from disruption_py.core.utils.misc import shot_msg
 from disruption_py.core.utils.shared_instance import SharedInstance
 from disruption_py.machine.tokamak import Tokamak
 
@@ -196,7 +196,6 @@ class MDSConnection:
         self,
         path: str,
         tree_name: str = None,
-        astype: str | None = None,
         arguments: Any = None,
     ) -> np.ndarray:
         """
@@ -208,8 +207,6 @@ class MDSConnection:
             MDSplus path to record.
         tree_name : str, optional
             The name of the tree that must be open for retrieval.
-        astype : str, optional, default = None
-            The data type for explicit casting, or None.
         arguments : Any, optional
             Arguments for MDSplus TDI Expression. Default None.
             Please see MDSplus documentation for more information.
@@ -225,8 +222,6 @@ class MDSConnection:
 
         logger.trace(shot_msg("Getting data: {path}"), shot=self.shot_id, path=path)
         data = self.conn.get(f"_sig={path}", arguments).data()
-        if astype:
-            data = safe_cast(data, astype)
 
         return data
 
@@ -236,8 +231,6 @@ class MDSConnection:
         path: str,
         tree_name: str = None,
         dim_nums: List = None,
-        astype: str | None = None,
-        cast_all: bool = False,
     ) -> Tuple:
         """
         Get data and dimension(s) for record at specified path.
@@ -250,10 +243,6 @@ class MDSConnection:
             The name of the tree that must be open for retrieval.
         dim_nums : List, optional
             A list of dimensions that should have their size retrieved. Default [0].
-        astype : str, optional, default = None
-            The data type for explicit casting, or None.
-        cast_all : bool, optional. Default False.
-            Whether to cast both data and dims, or only data.
 
         Returns
         -------
@@ -272,11 +261,6 @@ class MDSConnection:
         data = self.conn.get("_sig=" + path).data()
         dims = [self.conn.get(f"dim_of(_sig,{dim_num})").data() for dim_num in dim_nums]
 
-        if astype:
-            data = safe_cast(data, astype)
-            if cast_all:
-                dims = [safe_cast(dim, astype) for dim in dims]
-
         return data, *dims
 
     @_better_mds_exceptions
@@ -285,7 +269,6 @@ class MDSConnection:
         path: str,
         tree_name: str = None,
         dim_nums: List = None,
-        astype: str | None = None,
     ) -> Tuple:
         """
         Get the specified dimensions for record at specified path.
@@ -298,8 +281,6 @@ class MDSConnection:
             The name of the tree that must be open for retrieval.
         dim_nums : List, optional
             A list of dimensions that should have their size retrieved. Default [0].
-        astype : str, optional, default = None
-            The data type for explicit casting, or None.
 
         Returns
         -------
@@ -314,9 +295,6 @@ class MDSConnection:
 
         logger.trace(shot_msg("Getting dims: {path}"), shot=self.shot_id, path=path)
         dims = [self.conn.get(f"dim_of({path},{d})").data() for d in dim_nums]
-
-        if astype:
-            dims = [safe_cast(dim, astype) for dim in dims]
 
         return dims
 
