### 2025-09-18 a397691
fix cast when reading floats from csv
