### 2025-09-22 9457465
Revert "Remove dipprog_dt_rt from DIII-D XFAIL column (#399)"
This reverts commit 6d0cce98faeed48efdfdb8a7c972ef7ecbc26df4.
