+++ disruption_py/machine/d3d/config.toml
@@ -15,6 +15,7 @@ expected_failure_columns = [
   "beta_p_rt",
   "dbetap_dt",
   "delta",
+  "dipprog_dt_rt",
   "dli_dt",
   "dn_dt",
   "dwmhd_dt",
