### 2025-09-22 a36183a
automatically init DummyDatabase if server is blank
