### 2025-09-22 5ab0db3
drop dummy database doc example
