### 2025-09-22 de00356
fix HBT-EP config structures
