### 2025-09-22 b497d85
add hbtep methods to docs
