### 2025-09-22 75f147e
add MAST methods to docs
