### 2025-09-22 500baad
add MAST docs to menu
