+++ mkdocs.yml
@@ -54,8 +54,12 @@ nav:
                 east_method_reference.md"
               - HBT-EP Built-in Methods: "usage/physics_methods/\
                 hbtep_method_reference.md"
+              - MAST Built-in Methods: "usage/physics_methods/\
+                mast_built_in_method_reference.md"
           - Physics Method Decorators: "usage/physics_methods/\
             decorator_reference.md"
+              - Physics Method Decorators: "usage/physics_methods/\
+                decorator_reference.md"
           - Parameter Reference: "usage/physics_methods/\
             disruption_parameters_reference.md"
       - Data I/O:
