### 2025-09-23 1d175bc
Rough first pass at adding a few parameter groups
