+++ disruption_py/machine/mast/__init__.py
@@ -5,5 +5,6 @@ Module initialization for MAST.
 """
 
 from disruption_py.machine.mast.physics import MastPhysicsMethods
+from disruption_py.machine.mast.efit import MastEfitMethods
 
-METHOD_HOLDERS = [MastPhysicsMethods]
+METHOD_HOLDERS = [MastPhysicsMethods, MastEfitMethods]
