+++ disruption_py/machine/mast/physics.py
@@ -4,6 +4,8 @@
 Physics methods for MAST.
 """
 
+import numpy as np
+from disruption_py.core.physics_method.errors import CalculationError
 import xarray as xr
 from disruption_py.core.utils.math import interp1
 from disruption_py.core.physics_method.decorator import physics_method
@@ -33,3 +35,157 @@ class MastPhysicsMethods:
         times = params.times
         ip = interp1(magtime, ip, times)
         return {"ip": ip}
+
+    @staticmethod
+    @physics_method(
+        columns=["power_nbi", "power_radiated"],
+        tokamak=Tokamak.MAST,
+    )
+    def get_power(params: PhysicsMethodParams):
+        """Get power parameters"""
+        conn: XarrayConnection = params.mds_conn
+        file_name = conn.get_shot_file_path(params.shot_id)
+        ds = xr.open_zarr(file_name, group="summary")
+        power_nbi = ds["power_nbi"].values
+        power_radiated = ds["power_radiated"].values
+        base_time = ds["time"].values
+
+        times = params.times
+        power_nbi = interp1(base_time, power_nbi, times)
+        power_radiated = interp1(base_time, power_radiated, times)
+        return {"power_nbi": power_nbi, "power_radiated": power_radiated}
+
+    @staticmethod
+    @physics_method(
+        columns=["total_injected", "inboard_total", "outboard_total"],
+        tokamak=Tokamak.MAST,
+    )
+    def get_gas(params: PhysicsMethodParams):
+        """Get gas injection parameters"""
+        conn: XarrayConnection = params.mds_conn
+        file_name = conn.get_shot_file_path(params.shot_id)
+        ds = xr.open_zarr(file_name, group="gas_injection")
+        total_injected = ds["total_injected"].values
+        inboard_total = ds["inboard_total"].values
+        outboard_total = ds["outboard_total"].values
+        base_time = ds["time"].values
+
+        times = params.times
+        total_injected = interp1(base_time, total_injected, times)
+        inboard_total = interp1(base_time, inboard_total, times)
+        outboard_total = interp1(base_time, outboard_total, times)
+        return {
+            "total_injected": total_injected,
+            "inboard_total": inboard_total,
+            "outboard_total": outboard_total,
+        }
+
+    @staticmethod
+    @physics_method(
+        columns=["t_e_core", "n_e_core"],
+        tokamak=Tokamak.MAST,
+    )
+    def get_ts_parameters(params: PhysicsMethodParams):
+        """Get Thomson parameters"""
+        conn: XarrayConnection = params.mds_conn
+        file_name = conn.get_shot_file_path(params.shot_id)
+        ds = xr.open_zarr(file_name, group="thomson_scattering")
+        times = params.times
+        base_time = ds["time"].values
+        t_e_core = ds["t_e_core"].values
+        n_e_core = ds["n_e_core"].values
+        t_e_core = interp1(base_time, t_e_core, times)
+        n_e_core = interp1(base_time, n_e_core, times)
+        return {"t_e_core": t_e_core, "n_e_core": n_e_core}
+
+    @staticmethod
+    @physics_method(
+        columns=["n_e", "dn_dt", "greenwald_fraction"],
+        tokamak=Tokamak.MAST,
+    )
+    def get_densities(params: PhysicsMethodParams):
+        r"""
+        Calculate electron density, its time derivative, and the Greenwald fraction.
+
+        The Greenwald fraction is the ratio of the measured electron density $n_e$ and
+        the Greenwald density limit $n_G$ defined as [^1]:
+
+        $$
+        n_G = \frac{I_p}{\pi a^2}
+        $$
+
+        where $n_G$ is given in $10^{20} m^{-3}$ and $I_p$ is in MA.
+
[…]
