### 2025-09-23 13a5e88
cast times to float32 well ahead of xr conversion
