### 2025-09-23 5ac4214
compare only to 0.1 ms absolute precision
