### 2025-09-23 c6fefe2
be resilient for missing libodbc
