### 2025-09-23 f5f1c13
create new REFERENCES page
