+++ REFERENCES.md
@@ -1,0 +1,17 @@
+
+# DisruptionPy References
+
+### Projects
+
+- C Rea, et al. (2023), _"Open and FAIR Fusion for Machine Learning Applications"_, [Project website](https://crea-psfc.github.io/open-fair-fusion/).
+
+### Papers
+
+- GL Trevisan, _et al._ (2025), _"DisruptionPy: An open-source physics-based scientific framework for disruption analysis of fusion plasmas"_, JOSS, [submitted](https://github.com/openjournals/joss-reviews/issues/8538)
+
+### Conferences
+
+- 66th APS-DPP Meeting (2024)
+  - C Rea, et al. (2024), _"Open and FAIR Fusion for Machine Learning Applications"_, [PP12.27](https://meetings.aps.org/Meeting/DPP24/Session/PP12.27)
+  - GL Trevisan, et al. (2024), _"Functional Improvements and Technical Developments of a Community-driven and Physics-informed Numerical Library for Disruption Studies"_, [PP12.9](https://meetings.aps.org/Meeting/DPP24/Session/PP12.9)
+  - Y Wei, et al. (2024), _"Physics validation of parameter methods in DisruptionPy"_, [PP12.10](https://meetings.aps.org/Meeting/DPP24/Session/PP12.10)
