### 2025-09-23 70e007f
add  recent ARS paper
