### 2025-09-23 8e6b2c5
always put et al in italics
