### 2025-09-23 d38abb7
add filter_dict to utils.misc
