### 2025-09-23 5d5af44
do not dump password into config json
