+++ disruption_py/workflow.py
@@ -19,6 +19,7 @@ from tqdm.auto import tqdm
 from disruption_py.config import config
 from disruption_py.core.retrieval_manager import RetrievalManager
 from disruption_py.core.utils.misc import (
+    filter_dict,
     get_elapsed_time,
     get_temporary_folder,
     without_duplicates,
@@ -121,8 +122,8 @@ def get_shots_data(
         raise ModuleNotFoundError("Cannot import MDSplus.")
 
     # dump configuration
-    json_file_path = os.path.join(get_temporary_folder(), "config.json")
-    config_dict = config(tokamak).to_dict()
+    json_file_path = os.path.join(get_temporary_folder(), ".config.json")
+    config_dict = filter_dict(config(tokamak).to_dict(), "_pass")
     with open(json_file_path, "w", encoding="utf8") as f:
         json.dump(config_dict, f, indent=3, sort_keys=True)
     logger.debug("Dumped configuration into: {path}", path=json_file_path)
