### 2025-09-23 d9231fa
Update d3d test_shared_time_setting ground truth
