### 2025-09-25 a0472df
Added IpTimeSetting and get_ip physics method for hbtep
