### 2025-09-27 d60661f
Added a few methods needed for POPSIM transport study
