### 2025-09-27 9500827
Added some debugging, need to figure out the actual time and rho dimensions for zipfits
