+++ .github/workflows/lint.yml
@@ -32,7 +32,7 @@ jobs:
         run: pipx install poetry
 
       - name: Setup Python
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v6
         with:
           python-version-file: 'pyproject.toml'
           cache: 'poetry'
