+++ pyproject.toml
@@ -84,32 +84,11 @@ repository = "https://github.com/MIT-PSFC/disruption-py/"
 [tool.poetry.group.dev]
 optional = true
 
-[tool.poetry.group.dev.dependencies]
-black = "^25.1.0"
-pylint = "^3.3.0"
-pytest = "^8.3.0"
-ruff = "^0.9.0"
-toml-sort = "^0.24.0"
-yamllint = "^1.36.0"
-
 [tool.poetry.group.docs]
 optional = true
 
-[tool.poetry.group.docs.dependencies]
-black = "^25.1.0"
-mkdocs = "^1.6.0"
-mkdocs-literate-nav = "^0.6.0"
-mkdocs-material = "^9.6.0"
-mkdocstrings = {extras = ["python"], version = "^0.29.0"}
-
 [tool.poetry.group.lab]
 optional = true
-
-[tool.poetry.group.lab.dependencies]
-ipykernel = "^6.29.0"
-ipympl = "^0.9.0"
-jupyterlab = "^4.3.0"
-matplotlib = "^3.10.0"
 
 [tool.pylint]
 disable = [
