### 2025-09-29 3f5e723
remove python cap
