+++ pyproject.toml
@@ -70,7 +70,7 @@ maintainers = [
 ]
 name = "disruption-py"
 readme = "README.md"
-requires-python = ">=3.11,<4"
+requires-python = ">=3.11"
 version = "0.12.0-dev"
 
 [project.scripts]
