### 2025-09-29 ce2e48f
Add get_cap_bank_parameters and get_plasma_radii
