+++ disruption_py/machine/hbtep/physics.py
@@ -4,6 +4,10 @@
 Module for retrieving and calculating data for HBTEP physics methods.
 """
 
+import numpy as np
+from scipy.integrate import cumulative_trapezoid
+
+from disruption_py.core.physics_method.caching import cache_method
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
 from disruption_py.core.utils.math import interp1
@@ -24,6 +28,94 @@ class HbtepPhysicsMethods:
         """
         ip, t_ip = params.mds_conn.get_data_with_dims(
             r"\TOP.SENSORS.ROGOWSKIS:IP", tree_name="hbtep2"
-        )
+        )  # [A], [s]
         ip = interp1(t_ip, ip, params.times, "linear")
         return {"ip": ip}
+
+    @staticmethod
+    @physics_method(columns=["i_vfc", "i_ohc"], tokamak=Tokamak.HBTEP)
+    def get_cap_bank_parameters(params: PhysicsMethodParams):
+        """
+        Get the ohmic heating and vertical field coil currents
+        """
+        i_vfc, t_vfc = params.mds_conn.get_data_with_dims(
+            r"\TOP.SENSORS.VF_CURRENT", tree_name="hbtep2"
+        )  # [A], [s]
+        i_ohc, t_ohc = params.mds_conn.get_data_with_dims(
+            r"\TOP.SENSORS.OH_CURRENT", tree_name="hbtep2"
+        )  # [A], [s]
+        i_vfc = interp1(t_vfc, i_vfc, params.times, "linear")
+        i_ohc = interp1(t_ohc, i_ohc, params.times, "linear")
+        return {"i_vfc": i_vfc, "i_ohc": i_ohc}
+
+    @staticmethod
+    @physics_method(columns=["r", "aminor"], tokamak=Tokamak.HBTEP)
+    def get_plasma_radii(params: PhysicsMethodParams):
+        """
+        Get the major & minor radii
+        # TODO: get the updated limitor configuration post ~2023(?) from hbtplot.py
+        """
+        # Determined by Daisuke during copper plasma calibration
+        a = 0.00643005
+        b = -1.10423
+        c = 48.2567
+
+        # Calculated by Jeff, but still has errors
+        vf_pickup = 0.0046315133 * -1e-3
+        oh_pickup = 7.0723416e-08
+
+        # get vf and oh data
+        i_vfc, t_vfc = params.mds_conn.get_data_with_dims(
+            r"\TOP.SENSORS.VF_CURRENT", tree_name="hbtep2"
+        )  # [A], [s]
+        i_ohc, t_ohc = params.mds_conn.get_data_with_dims(
+            r"\TOP.SENSORS.OH_CURRENT", tree_name="hbtep2"
+        )  # [A], [s]
+
+        # get plasma current
+        ip, t_ip = params.mds_conn.get_data_with_dims(
+            r"\TOP.SENSORS.ROGOWSKIS:IP", tree_name="hbtep2"
+        )  # [A], [s]
+        ip *= 1212.3 * 1e-9  # ip gain
+
+        # get cosine Rogowski data
+        cos1_raw, t_cos1_raw = params.mds_conn.get_data_with_dims(
+            r"\TOP.SENSORS.ROGOWSKIS:COS_1:RAW", tree_name="hbtep2"
+        )  # TODO: units?, TODO: check if we need to specify time range
+        # calculate and subtract offset
+        (indices,) = np.where((-1e-3 < t_cos1_raw) & (t_cos1_raw < 0))
+        cos1_raw_offset = np.mean(cos1_raw[indices])
[…]
