### 2025-09-29 3ffb21d
Add get_btor and get_qstar
