### 2025-09-29 b62309a
Add get_v_loop
