### 2025-09-29 e40e769
Add get_h_alpha
