### 2025-09-29 b478843
Add get_sxr_midplane
