### 2025-09-30 8bff629
remove deprecated license classifier
