### 2025-09-30 b4bce82
Implement get_n_mode_data and associated functions
