+++ disruption_py/machine/hbtep/physics.py
@@ -10,7 +10,12 @@ from scipy.integrate import cumulative_trapezoid
 from disruption_py.core.physics_method.caching import cache_method
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
-from disruption_py.core.utils.math import interp1
+from disruption_py.core.utils.math import (
+    butterworth_filter,
+    gaussian_low_pass_filter,
+    interp1,
+)
+from disruption_py.machine.hbtep.util import HbtepUtilMethods
 from disruption_py.machine.tokamak import Tokamak
 
 
@@ -191,3 +196,223 @@ class HbtepPhysicsMethods:
         # interpolate to requested timebase
         sxr = interp1(t_sxr, sxr, params.times, "linear")
         return {"sxr_midplane": sxr}
+
+    @staticmethod
+    @physics_method(
+        columns=[
+            "n_equal_1_mode",
+            "n_equal_1_normalized",
+            "n_equal_1_phase",
+            "n_equal_1_frequency",
+            "n_equal_2_mode",
+            "n_equal_2_normalized",
+            "n_equal_2_phase",
+            "n_equal_2_frequency",
+        ],
+        tokamak=Tokamak.HBTEP,
+    )
+    def get_n_mode_data(params: PhysicsMethodParams):
+        """
+        Get the n=1 and 2 mode amplitdes, phases, and frequencies from the toroidal array (TA)
+        # TODO: implement this function
+        """
+        # Get TA data
+        ta_data = HbtepPhysicsMethods._get_ta_data(params)
+        data = ta_data["ta_pol_data_filt"]
+        time = ta_data["ta_pol_time"]
+        phi = ta_data["ta_pol_phi"]
+        n, m = len(data), len(
+            data[0]
+        )  # TODO: decide if we need to convert data to np.ndarray
+
+        # Construct A matrix and calculate its inversion
+        A = np.zeros((n, 5))
+        A[:, 0] = np.ones(n)
+        A[:, 1] = np.sin(phi)
+        A[:, 2] = np.cos(phi)
+        A[:, 3] = np.sin(2 * phi)
+        A[:, 4] = np.cos(2 * phi)
+        Ainv = np.linalg.pinv(A)
+
+        # Use least squares to solve for coefficients
+        x = np.matmul(Ainv, data)
+        n1_amp = np.sqrt(x[1, :] ** 2 + x[2, :] ** 2)  # [T]
+        n2_amp = np.sqrt(x[1, :] ** 2 + x[2, :] ** 2)
+        n1_phase = -np.arctan2(x[1, :], x[2, :])  # *-1 to corrected the slope of phase
+        n2_phase = -np.arctan2(x[3, :], x[4, :])
+
+        # Calculate mode frequency. Note that gaussian_low_pass_filter is not a causal filter.
+        n1_phase_filt = gaussian_low_pass_filter(
+            HbtepUtilMethods.unwrap_phase(n1_phase), time, 1e-5
+        )
+        n2_phase_filt = gaussian_low_pass_filter(
+            HbtepUtilMethods.unwrap_phase(n2_phase), time, 1e-5
+        )
+        n1_freq = np.gradient(n1_phase_filt) / np.gradient(time) / (2 * np.pi)
+        n2_freq = np.gradient(n2_phase_filt) / np.gradient(time) / (2 * np.pi)
+
+        # Interpolate output data to the requested timebase
+        n1_amp = interp1(time, n1_amp, params.times, "linear")
+        n2_amp = interp1(time, n2_amp, params.times, "linear")
+        n1_phase = interp1(time, n1_phase, params.times, "linear")
[…]
