### 2025-09-30 5568c44
Update deps to Sep 2025 (#475)
* bump actions/setup-python to v6

* poetry now supports PEP 735 dep groups

* remove python cap

* remove deprecated license classifier

* lock deps
