### 2025-09-30 185ea5d
Add get_m_mode_data and associated functions
