+++ disruption_py/machine/hbtep/physics.py
@@ -276,6 +276,98 @@ class HbtepPhysicsMethods:
         }
 
     @staticmethod
+    @physics_method(
+        columns=[
+            "m_equal_2_mode",
+            "m_equal_2_normalized",
+            "m_equal_2_phase",
+            "m_equal_2_frequency",
+            "m_equal_3_mode",
+            "m_equal_3_normalized",
+            "m_equal_3_phase",
+            "m_equal_3_frequency",
+            "m_equal_4_mode",
+            "m_equal_4_normalized",
+            "m_equal_4_phase",
+            "m_equal_4_frequency",
+        ],
+        tokamak=Tokamak.HBTEP,
+    )
+    def get_m_mode_data(params: PhysicsMethodParams):
+        """
+        Get the m=2,3,4 mode amplitdes, phases, and frequencies from the poloidal array (PA1)
+        """
+        # Get PA1 data
+        pa_data = HbtepPhysicsMethods._get_pa_data(params)
+        data = pa_data["pa1_data_filt"]
+        time = pa_data["pa1_time"]
+        theta = pa_data["pa1_theta"]
+        phi = pa_data["pa1_phi"]
+        n, m = len(data), len(data[0])
+        # Construct A matrix and calculate its inversion
+        A = np.zeros((n, 11))
+        A[:, 0] = np.ones(n)
+        for i in range(1, 6):
+            A[:, i * 2 - 1] = np.sin(i * theta - phi)
+            A[:, i * 2] = np.cos(i * theta - phi)
+        Ainv = np.linalg.pinv(A)
+
+        # Use least squares to solve for coefficients
+        x = np.matmul(Ainv, data)
+        m2_amp = np.sqrt(x[3, :] ** 2 + x[4, :] ** 2)  # [T]
+        m3_amp = np.sqrt(x[5, :] ** 2 + x[6, :] ** 2)
+        m4_amp = np.sqrt(x[7, :] ** 2 + x[8, :] ** 2)
+        m2_phase = np.arctan2(x[3, :], x[4, :])
+        m3_phase = np.arctan2(x[5, :], x[6, :])
+        m4_phase = np.arctan2(x[7, :], x[8, :])
+
+        # Calculate mode frequency. Note that gaussian_low_pass_filter is not a causal filter.
+        m2_phase_filt = gaussian_low_pass_filter(
+            HbtepUtilMethods.unwrap_phase(m2_phase), time, 1e-5
+        )
+        m3_phase_filt = gaussian_low_pass_filter(
+            HbtepUtilMethods.unwrap_phase(m3_phase), time, 1e-5
+        )
+        m4_phase_filt = gaussian_low_pass_filter(
+            HbtepUtilMethods.unwrap_phase(m4_phase), time, 1e-5
+        )
+        m2_freq = np.gradient(m2_phase_filt) / np.gradient(time) / (2 * np.pi)
+        m3_freq = np.gradient(m3_phase_filt) / np.gradient(time) / (2 * np.pi)
+        m4_freq = np.gradient(m4_phase_filt) / np.gradient(time) / (2 * np.pi)
+
+        # Interpolate output data to the requested timebase
+        m2_amp = interp1(time, m2_amp, params.times, "linear")
+        m3_amp = interp1(time, m3_amp, params.times, "linear")
+        m4_amp = interp1(time, m4_amp, params.times, "linear")
+        m2_phase = interp1(time, m2_phase, params.times, "linear")
+        m3_phase = interp1(time, m3_phase, params.times, "linear")
+        m4_phase = interp1(time, m4_phase, params.times, "linear")
+        m2_freq = interp1(time, m2_freq, params.times, "linear")
[…]
