### 2025-09-30 cec3986
Remove redundant variables and TODOs
