### 2025-10-01 fb8aa65
Add option to choose the new "disruption" efit21 tree (#470)
* read efit21 rather than efit18

* set efit21 as default, but keep efit18 for testing

* default to efit21 for all shots, testing fallback to efit18 only for disruptive shots

* make the runtag for DIII-D a config parameter

* introduce cmod.efit.tree, reset default to efit18 for now
