### 2025-10-02 1ac2e09
Implement get_sxr_total
