### 2025-10-02 906cf9a
Implement get_euv_total
