+++ disruption_py/machine/hbtep/physics.py
@@ -376,6 +376,52 @@ class HbtepPhysicsMethods:
         return {"sxr_total": sxr_total}
 
     @staticmethod
+    @physics_method(
+        columns=["euv_000_total", "euv_025_total", "euv_090_total", "euv_270_total"],
+        tokamak=Tokamak.HBTEP,
+    )
+    def get_euv_total(params: PhysicsMethodParams):
+        """
+        Get the total emissions from the 4 EUV fan arrays
+        - 000: outboard midplane array, total emission
+        - 025: outboard upper midplane array, emission from top one-third of the plasma
+        - 090: top array, emission from the inboard half of the plasma
+        - 270: bottom array, emission from the outboard half of the plasma
+        """
+        euv_data = HbtepPhysicsMethods._get_euv_data(params)
+        euv_000_total = interp1(
+            euv_data["time"],
+            np.sum(euv_data["euv_000_data"], axis=0),
+            params.times,
+            "linear",
+        )
+        euv_025_total = interp1(
+            euv_data["time"],
+            np.sum(euv_data["euv_025_data"], axis=0),
+            params.times,
+            "linear",
+        )
+        euv_090_total = interp1(
+            euv_data["time"],
+            np.sum(euv_data["euv_090_data"], axis=0),
+            params.times,
+            "linear",
+        )
+        euv_270_total = interp1(
+            euv_data["time"],
+            np.sum(euv_data["euv_270_data"], axis=0),
+            params.times,
+            "linear",
+        )
+
+        return {
+            "euv_000_total": euv_000_total,
+            "euv_025_total": euv_025_total,
+            "euv_090_total": euv_090_total,
+            "euv_270_total": euv_270_total,
+        }
+
+    @staticmethod
     @cache_method
     def _get_ta_data(params: PhysicsMethodParams):
         r"""
@@ -781,4 +827,112 @@ class HbtepPhysicsMethods:
         output["det_ap_pol"] = 0.001270000007934868
         output["det_ap_tor"] = 0.02539999969303608
 
+        return output
+
+    @staticmethod
+    @cache_method
+    def _get_euv_data(params: PhysicsMethodParams):
+        """
+        Get data from the 4 EUV fan array
+        """
+        output = {}
+        output["detectors"] = [0, 25, 90, 270]  # poloidal location of the 4 fan arrays
+
+        time = params.mds_conn.get_dims(
+            r"\TOP.SENSORS.EUV.POL.DET000.channel_01:RAW", tree_name="hbtep2"
+        )
+        output["time"] = time[0]
+
+        for detector in output["detectors"]:
+            data, r, z, gain = [], [], [], []
+            for channel in range(1, 17):
+                address = (
+                    rf"\TOP.SENSORS.EUV.POL.DET{detector:03d}.CHANNEL_{channel:02d}"
+                )
+                data.append(
+                    params.mds_conn.get_data(address + ":RAW", tree_name="hbtep2")
+                )
+                r.append(params.mds_conn.get_data(address + ":R", tree_name="hbtep2"))
+                z.append(params.mds_conn.get_data(address + ":Z", tree_name="hbtep2"))
+                gain.append(
+                    params.mds_conn.get_data(address + ":GAIN", tree_name="hbtep2")
+                )
+            output[f"euv_{detector:03d}_data"] = data
+            output[f"euv_{detector:03d}_r"] = r
+            output[f"euv_{detector:03d}_z"] = z
+            output[f"euv_{detector:03d}_gain"] = gain
+
+        # Aperture location
+        output["det_ap_r"] = [1.160508, 1.090508, 0.907057, 0.929746]
+        output["det_ap_z"] = [0.000000, 0.099975, 0.166773, -0.173440]
+        output["det_ap_pol"] = [0.000635, 0.000635, 0.000635, 0.000635]
[…]
