### 2025-10-03 1612246
added Saperstein conferences to REFERENCES.md
