### 2025-10-06 5dfedd1
tweak references
