### 2025-10-06 a22d894
reorder, add next and previous APS DPP confs
