+++ REFERENCES.md
@@ -12,12 +12,19 @@
 
 ### Conferences
 
+- [67th APS-DPP Meeting (2025)](https://meetings.aps.org/Meeting/DPP25)
+  - ...
+
+- [45th ITPA-MDC Meeting (2025)](https://sites.google.com/pppl.gov/itpa-mdc2025)
+  - AR Saperstein, _et al._ (2025), _"Progress on the development of an off-normal warning system for SPARC"_
+
 - [66th APS-DPP Meeting (2024)](https://meetings.aps.org/Meeting/DPP24)
   - C Rea, _et al._ (2024), _"Open and FAIR Fusion for Machine Learning Applications"_, [PP12.27](https://meetings.aps.org/Meeting/DPP24/Session/PP12.27)
   - GL Trevisan, _et al._ (2024), _"Functional Improvements and Technical Developments of a Community-driven and Physics-informed Numerical Library for Disruption Studies"_, [PP12.9](https://meetings.aps.org/Meeting/DPP24/Session/PP12.9)
   - Y Wei, _et al._ (2024), _"Physics validation of parameter methods in DisruptionPy"_, [PP12.10](https://meetings.aps.org/Meeting/DPP24/Session/PP12.10)
   - AR Saperstein, _et al._ (2024), _"Development and preliminary calibration of an off-normal warning system for SPARC"_, [TO06.9](https://meetings.aps.org/Meeting/DPP24/Session/TO06.9)
+
 - [3rd IAEA Technical Meeting on Plasma Disruptions and their Mitigation (2024)](https://conferences.iaea.org/event/380/)
   - AR Saperstein, _et al._ (2024), _"Development and preliminary calibration of an off-normal warning system for SPARC"_, [32305](https://conferences.iaea.org/event/380/contributions/32305/)
-- [45th ITPA-MDC Meeting (2025)](https://sites.google.com/pppl.gov/itpa-mdc2025)
-  - AR Saperstein, _et al._ (2025), _"Progress on the development of an off-normal warning system for SPARC"_
+
+- [65th APS-DPP Meeting (2023)](https://meetings.aps.org/Meeting/DPP23)
