### 2025-10-06 e875e8f
remove h_alpha from DIII-D xfailures
