+++ disruption_py/machine/d3d/config.toml
@@ -28,7 +28,6 @@ expected_failure_columns = [
   "dn_dt",
   "dwmhd_dt",
   "greenwald_fraction",
-  "h_alpha",
   "kappa",
   "kappa_area",
   "li",
