### 2025-10-07 5a60938
Change xarray to use path. More generic
