### 2025-10-08 00255a4
move mdsthin into dev deps
