+++ pyproject.toml
@@ -5,6 +5,7 @@ requires = ["poetry-core>=2.0"]
 [dependency-groups]
 dev = [
   "black>=25.1.0",
+  "mdsthin>=1.5.0",
   "pylint>=3.3.0",
   "pytest>=8.3.0",
   "ruff>=0.9.0",
@@ -48,7 +49,6 @@ classifiers = [
 dependencies = [
   "dynaconf>=3.2.0",
   "loguru>=0.7.0",
-  "mdsthin>=1.5.0",
   "netcdf4>=1.7.0",
   "numpy>=1.26.0,<2.0.0",
   "pandas>=2.2.0",
