### 2025-10-08 154918c
fixup! add MAST methods to docs
