+++ docs/usage/physics_methods/mast_built_in_method_reference.md
@@ -1,4 +1,4 @@
-## Built-in Methods For CMod { .doc .doc-heading }
+## Built-in Methods For MAST { .doc .doc-heading }
 ::: disruption_py.machine.mast.physics
     handler: python
 	options:
