### 2025-10-08 95aa7cf
add doc reference to MAST physics methods
