### 2025-10-08 30b5451
slight change in exception and logging logic
