### 2025-10-08 73a4e68
fixup! slight change in exception and logging logic
