### 2025-10-09 e290fe8
Auto initialize dummy DB if SQL server is missing (#473)
* automatically init DummyDatabase if server is blank

* drop dummy database doc example

* be resilient for missing libodbc
