### 2025-10-09 7d5d62a
Add mdsthin as fallback from MDSplus (#471)
* add mdsthin to deps

* lock deps

* clearer full-mdsplus imports for compare_nodenames

* try MDSplus first, then fall back onto mdsthin

* harmonize MDSplus imports and exceptions in framework

* harmonize MDSplus imports and exceptions for C-MOD

* harmonize MDSplus imports and exceptions for DIII-D

* harmonize MDSplus imports and exceptions for EAST

* fixup! harmonize MDSplus imports and exceptions for DIII-D

* try to fix str conversions for arrays

* do not decorate mdsthin exceptions

* add verbose statement with MDSplus api package

* slig
[…]
