+++ disruption_py/workflow.py
@@ -5,6 +5,7 @@ main workflow
 """
 
 import argparse
+import sys
 import time
 from itertools import repeat
 from multiprocessing import Pool
@@ -102,6 +103,9 @@ def get_shots_data(
     """
     log_settings = resolve_log_settings(log_settings)
     log_settings.setup_logging()
+
+    if "mdsthin" in sys.modules:
+        logger.warning("Cannot import MDSplus, falling back onto mdsthin!")
 
     tokamak = resolve_tokamak_from_environment(tokamak)
     database = _get_database_instance(tokamak, database_initializer)
