+++ disruption_py/machine/cmod/physics.py
@@ -8,7 +8,6 @@ import warnings
 
 import numpy as np
 import scipy.constants as const
-from MDSplus import mdsExceptions
 
 from disruption_py.core.physics_method.caching import cache_method
 from disruption_py.core.physics_method.decorator import physics_method
@@ -20,6 +19,7 @@ from disruption_py.core.utils.math import (
     gaussian_fit_with_fixed_mean,
     interp1,
 )
+from disruption_py.inout.mds import mdsExceptions
 from disruption_py.machine.cmod.thomson import CmodThomsonDensityMeasure
 from disruption_py.machine.tokamak import Tokamak
 
