+++ pyproject.toml
@@ -5,6 +5,7 @@ requires = ["poetry-core>=2.0"]
 [dependency-groups]
 dev = [
   "black>=25.1.0",
+  "mdsthin>=1.5.0",
   "pylint>=3.3.0",
   "pytest>=8.3.0",
   "ruff>=0.9.0",
