+++ tests/test_time_setting.py
@@ -52,8 +52,8 @@ def run_test_time_setting(
     )
     times = shot_data["time"].to_numpy()
     # Check start, end, and length of time array
-    assert times[0] == pytest.approx(t_start, rel=1e-4)
-    assert times[-1] == pytest.approx(t_stop, rel=1e-4)
+    assert times[0] == pytest.approx(t_start, abs=1e-4)
+    assert times[-1] == pytest.approx(t_stop, abs=1e-4)
     assert len(times) == length
     # Check for duplicated time point
     assert len(times) == len(np.unique(times))
@@ -65,7 +65,7 @@ def test_shared_time_setting(tokamak: Tokamak, test_folder_f: str):
     """
     test_setup = {
         Tokamak.CMOD: ["analysis", 1150805012, 0.0601, 1.2799, 6100],
-        Tokamak.D3D: ["efit01", 161228, 0.1, 5.0395, 9880],
+        Tokamak.D3D: ["efit01", 161228, 0.1, 5.04, 9881],
         Tokamak.EAST: ["efit_east", 55012, 0.301, 5.7, 5401],
     }
     run_test_time_setting(tokamak, "ip_efit", *test_setup[tokamak], test_folder_f)
