+++ disruption_py/settings/time_setting.py
@@ -83,7 +83,7 @@ def _postprocess(times: np.ndarray, units: str = "") -> np.ndarray:
     np.ndarray
         Post-processed array of times.
     """
-    times = np.unique(times)
+    times = np.unique(times).astype("float32")
     units = units.lower().strip()
     if units == "ms":
         return times * 1e-3
@@ -248,10 +248,10 @@ class EfitTimeSetting(TimeSetting):
             Array of times in the timebase.
         """
         (efit_time,) = params.mds_conn.get_dims(
-            r"\efit_aeqdsk:ali", tree_name="_efit_tree", astype="float64"
+            r"\efit_aeqdsk:ali", tree_name="_efit_tree"
         )
         efit_time_unit = params.mds_conn.get_data(
-            r"units_of(dim_of(\efit_aeqdsk:ali))", tree_name="_efit_tree", astype="str"
+            r"units_of(dim_of(\efit_aeqdsk:ali))", tree_name="_efit_tree"
         )
         if efit_time_unit not in {"s", "ms", "us"}:
             params.logger.verbose(
@@ -594,7 +594,7 @@ class SignalTimeSetting(TimeSetting):
         """
         try:
             (signal_time,) = params.mds_conn.get_dims(
-                self.signal_path, tree_name=self.tree_name, astype="float64"
+                self.signal_path, tree_name=self.tree_name
             )
         except mdsExceptions.MdsException:
             params.logger.error(
@@ -603,9 +603,7 @@ class SignalTimeSetting(TimeSetting):
             )
             raise
         signal_unit = params.mds_conn.get_data(
-            f"units_of(dim_of({self.signal_path}))",
-            tree_name=self.tree_name,
-            astype="str",
+            f"units_of(dim_of({self.signal_path}))", tree_name=self.tree_name
         )
         if (
             not signal_unit.strip()
