+++ disruption_py/machine/d3d/config.toml
@@ -23,11 +23,11 @@ expected_failure_columns = [
   "beta_p_rt",
   "dbetap_dt",
   "delta",
+  "dipprog_dt_rt",
   "dli_dt",
   "dn_dt",
   "dwmhd_dt",
   "greenwald_fraction",
-  "h_alpha",
   "kappa",
   "kappa_area",
   "li",

[…]
