### 2025-10-09 38a0979
added wietfeldt references
