+++ REFERENCES.md
@@ -15,6 +15,7 @@
 
 - [67th APS-DPP Meeting (2025)](https://meetings.aps.org/Meeting/DPP25)
   - AD Maris, _et al._ (2025), _"Collisionality scaling of the tokamak density limit: data-driven analysis, cross-device prediction, and real-time avoidance"_
+  - H Wietfeldt, _et al._ (2025), _"Assessing the Risk of UFO Disruptions to SPARC via a Multi-device Study"_
 
 - [6th International Conference on Data-Driven Plasma Science (2025)](https://web.cvent.com/event/7de9d238-e170-4fbf-8de5-20abc5c6eb49/summary)
   - Y Wei, _et al._ (2025), _"DisruptionPy: An open-source physics-based scientific package for disruption studies on magnetic fusion experiment devices"_
@@ -22,6 +23,7 @@
 - [45th ITPA-MDC Meeting (2025)](https://sites.google.com/pppl.gov/itpa-mdc2025)
   - AR Saperstein, _et al._ (2025), _"Progress on the development of an off-normal warning system for SPARC"_
   - AD Maris, _et al._ (2025), _"Cross-device prediction and real-time avoidance of the density limit"_
+  - H Wietfeldt, _et al._ (2025), _"Characterizing UFO Disruptions on Alcator C-Mod"_
 
 - [Open Source Software for Fusion Energy (2025)](https://ossfe.org/OSSFE_2025/#welcome-to-ossfe-2025)
   - Y Wei, _et al._ (2025), _"DisruptionPy: an open-source Python library for disruption study."_, [POS-10](https://ossfe.org/OSSFE_2025/y-disruptionpy)
@@ -32,6 +34,7 @@
   - Y Wei, _et al._ (2024), _"Physics validation of parameter methods in DisruptionPy"_, [PP12.10](https://meetings.aps.org/Meeting/DPP24/Session/PP12.10)
   - AR Saperstein, _et al._ (2024), _"Development and preliminary calibration of an off-normal warning system for SPARC"_, [TO06.9](https://meetings.aps.org/Meeting/DPP24/Session/TO06.9)
   - AD Maris, _et al._ (2024), _"Correlation of the tokamak density limit with edge collisionality"_, **Invited** [BI02.02](https://meetings.aps.org/Meeting/DPP24/Session/BI02.2)
+  - H Wietfeldt, _et al._ (2024), _"Discerning Why Some High-Z UFOs in C-Mod Caused Immediate Disruptions while Others Did Not"_, [NP12.00111](https://meetings.aps.org/Meeting/DPP24/Session/NP12.111)
 
 - [3rd IAEA Technical Meeting on Plasma Disruptions and their Mitigation (2024)](https://conferences.iaea.org/event/380/)
   - AR Saperstein, _et al._ (2024), _"Development and preliminary calibration of an off-normal warning system for SPARC"_, [32305](https://conferences.iaea.org/event/380/contributions/32305/)
