### 2025-10-14 ee81244
fix unsafe casting in get_power_parameters
