### 2025-10-14 b96247b
Fix #170648 | get_ip_parameters: TypeError
