+++ disruption_py/machine/d3d/physics.py
@@ -646,11 +646,16 @@ class D3DPhysicsMethods:
         except mdsExceptions.MdsException as e:
             params.logger.warning("Failed to get ipimode signal. Setting to NaN.")
             params.logger.opt(exception=True).debug(e)
-            ipimode = np.full(len(params.times), np.nan)
-        feedback_on_indices = np.where((ipimode == 0) | (ipimode == 3))
-        ip_error[feedback_on_indices] = (
-            ip[feedback_on_indices] - ip_prog[feedback_on_indices]
-        )
+            ipimode = [np.nan]
+        if (
+            ~np.isnan(ip).all()
+            and ~np.isnan(ip_prog).all()
+            and ~np.isnan(ipimode).all()
+        ):
+            feedback_on_indices = np.where((ipimode == 0) | (ipimode == 3))
+            ip_error[feedback_on_indices] = (
+                ip[feedback_on_indices] - ip_prog[feedback_on_indices]
+            )
         # Finally, get 'epsoff' to determine if/when the E-coil power supplies have railed
         # Times at which power_supply_railed ~=0 (i.e. epsoff ~=0) mean that
         # PCS feedback control of Ip is not being applied.  Therefore the
