### 2025-10-14 24cfb38
Fix #170648 | get_time_domain: TypeError
