### 2025-10-15 e88b897
add HBT-EP methods to menu
