### 2025-10-15 a75203d
cache get_plasma_radii
