### 2025-10-16 26fed03
introduce MDS_HOST, harmonize default_tree_path
