### 2025-10-16 d9dd5d2
Create hbtep_times override for DisruptionTimeSetting
