### 2025-10-16 4b33f4f
Convert all mds addresses to lowercase
