+++ disruption_py/machine/hbtep/physics.py
@@ -32,7 +32,7 @@ class HbtepPhysicsMethods:
         Get the plasma current
         """
         ip, t_ip = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.ROGOWSKIS:IP", tree_name="hbtep2"
+            r"\top.sensors.rogowskis:ip", tree_name="hbtep2"
         )  # [A], [s]
         ip = interp1(t_ip, ip, params.times, "linear")
         return {"ip": ip}
@@ -44,10 +44,10 @@ class HbtepPhysicsMethods:
         Get the ohmic heating and vertical field coil currents
         """
         i_vfc, t_vfc = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.VF_CURRENT", tree_name="hbtep2"
+            r"\top.sensors.vf_current", tree_name="hbtep2"
         )  # [A], [s]
         i_ohc, t_ohc = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.OH_CURRENT", tree_name="hbtep2"
+            r"\top.sensors.oh_current", tree_name="hbtep2"
         )  # [A], [s]
         i_vfc = interp1(t_vfc, i_vfc, params.times, "linear")
         i_ohc = interp1(t_ohc, i_ohc, params.times, "linear")
@@ -71,21 +71,21 @@ class HbtepPhysicsMethods:
 
         # get vf and oh data
         i_vfc, t_vfc = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.VF_CURRENT", tree_name="hbtep2"
+            r"\top.sensors.vf_current", tree_name="hbtep2"
         )  # [A], [s]
         i_ohc, t_ohc = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.OH_CURRENT", tree_name="hbtep2"
+            r"\top.sensors.oh_current", tree_name="hbtep2"
         )  # [A], [s]
 
         # get plasma current
         ip, t_ip = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.ROGOWSKIS:IP", tree_name="hbtep2"
+            r"\top.sensors.rogowskis:ip", tree_name="hbtep2"
         )  # [A], [s]
         ip *= 1212.3 * 1e-9  # ip gain
 
         # get cosine Rogowski data
         cos1_raw, t_cos1_raw = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.ROGOWSKIS:COS_1:RAW", tree_name="hbtep2"
+            r"\top.sensors.rogowskis:cos_1:raw", tree_name="hbtep2"
         )  # TODO: units?, TODO: check if we need to specify time range
         # calculate and subtract offset
         (indices,) = np.where((-1e-3 < t_cos1_raw) & (t_cos1_raw < 0))
@@ -139,7 +139,7 @@ class HbtepPhysicsMethods:
         Calculate B_tor from the TF probe data
         """
         btor, t_btor = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.TF_PROBE", tree_name="hbtep2"
+            r"\top.sensors.tf_probe", tree_name="hbtep2"
         )  # [T], [s]
         btor = interp1(t_btor, btor, params.times, "linear")
 
@@ -170,7 +170,7 @@ class HbtepPhysicsMethods:
         Get v_loop measurement
         """
         v_loop, t_v_loop = params.mds_conn.get_data_with_dims(
-            r"\TOP.SENSORS.LOOP_VOLTAGE", tree_name="hbtep2"
+            r"\top.sensors.loop_voltage", tree_name="hbtep2"
[…]
