### 2025-10-16 c7bec6c
Update test_time_setting
