### 2025-10-16 cd6bec0
Make get_n_mode_data functional in case TA sensors have different length of time
