### 2025-10-16 68cf3ff
Update get_m_mode_data with the same changes in get_n_mode_data
