### 2025-10-17 1951541
Replace all *pi/180 with np.deg2rad
