+++ disruption_py/machine/hbtep/physics.py
@@ -500,46 +500,42 @@ class HbtepPhysicsMethods:
             "TA10_S2R",
         ]
         # Toroidal and poloidal positions of the poloidal sensors
-        output["ta_pol_phi"] = (
-            np.pi
-            / 180
-            * np.array(
-                [
-                    241.5,
-                    250.5,
-                    259.5,
-                    277.5,
-                    286.5,
-                    295.5,
-                    313.5,
-                    322.5,
-                    331.5,
-                    349.5,
-                    358.5,
-                    7.5,
-                    25.5,
-                    34.5,
-                    43.5,
-                    61.5,
-                    70.5,
-                    79.5,
-                    97.5,
-                    106.5,
-                    115.5,
-                    133.5,
-                    142.5,
-                    151.5,
-                    169.5,
-                    178.5,
-                    187.5,
-                    205.5,
-                    214.5,
-                    223.5,
-                ]
-            )
+        output["ta_pol_phi"] = np.deg2rad(
+            [
+                241.5,
+                250.5,
+                259.5,
+                277.5,
+                286.5,
+                295.5,
+                313.5,
+                322.5,
+                331.5,
+                349.5,
+                358.5,
+                7.5,
+                25.5,
+                34.5,
+                43.5,
+                61.5,
+                70.5,
+                79.5,
+                97.5,
+                106.5,
+                115.5,
+                133.5,
+                142.5,
+                151.5,
+                169.5,
+                178.5,
+                187.5,
+                205.5,
+                214.5,
+                223.5,
+            ]
         )
-        output["ta_pol_theta"] = (
-            np.ones(len(output["ta_pol_phi"])) * (189 - 360) * np.pi / 180
+        output["ta_pol_theta"] = np.deg2rad(
+            np.ones(len(output["ta_pol_phi"])) * (189 - 360)
         )
         # Toroidal and poloidal positions of the radial sensors -- not implemented in hbteplib
 
@@ -664,88 +660,80 @@ class HbtepPhysicsMethods:
         ]
 
         # Toroidal and poloidal positions of the sensors
-        output["pa1_theta"] = (
-            np.array(
-                [
-                    -174.74778518,
-                    -164.23392461,
-                    -153.66901098,
-                    -143.01895411,
-                    -132.24974382,
-                    -121.3277924,
-                    -110.22067715,
-                    -98.93591492,
-                    -87.23999699,
-                    -75.60839722,
-                    -63.97679673,
-                    -52.34519359,
-                    -40.71359604,
-                    -29.08199717,
-                    -17.45039318,
-                    -5.81879416,
-                    5.81280487,
-                    17.44440438,
-                    29.07600466,
-                    40.70760263,
-                    52.33920936,
-                    63.97080017,
-                    75.60240749,
-                    87.23400093,
-                    98.93591492,
-                    110.22067715,
-                    121.3277924,
-                    132.24974382,
-                    143.01895411,
-                    153.66901098,
-                    164.23392461,
-                    174.74778518,
-                ]
-            )
-            * np.pi
-            / 180.0
+        output["pa1_theta"] = np.deg2rad(
+            [
+                -174.74778518,
+                -164.23392461,
+                -153.66901098,
+                -143.01895411,
+                -132.24974382,
+                -121.3277924,
+                -110.22067715,
+                -98.93591492,
+                -87.23999699,
+                -75.60839722,
+                -63.97679673,
+                -52.34519359,
+                -40.71359604,
+                -29.08199717,
+                -17.45039318,
+                -5.81879416,
+                5.81280487,
+                17.44440438,
+                29.07600466,
+                40.70760263,
+                52.33920936,
[…]
