### 2025-10-17 52063d1
Use list comprehension
