### 2025-10-17 e21cda7
Add @cache_method decorator to get_btor
