### 2025-10-17 96d5531
Fix TypeError("'numpy.float32' object does not support item assignment")
