### 2025-10-17 bebebcc
Minor fix to d3d get_peaking_factors
