### 2025-10-20 dc22d4a
cache get_ip, called by get_qstar
