### 2025-10-20 d8dfa4d
Minor changes to _check_mds_data
