### 2025-10-20 7e20c33
Improve logic in retrieving density data
