### 2025-10-20 6bb92a8
Fix #170648 | get_rt_ip_parameters: TypeError('only integer scalar arrays can be converted to a scalar index')
