### 2025-10-21 11ff9b6
Use lowercase address
Co-authored-by: Gregorio L. Trevisan <gtrevisan@users.noreply.github.com>
