### 2025-10-21 e756831
Update tests for hbtep
