+++ disruption_py/machine/hbtep/config.toml
@@ -9,4 +9,7 @@ expected_failure_columns = []
 test_columns = []
 
 [hbtep.tests.shots]
-flattop1_fast = 103590
+flattop1_full = 77324
+flattop2_full = 102709
+flattop3_fast = 103590
+flattop4_full = 117676
