### 2025-10-22 64a1e7e
Added references for survival analysis and in-progress TMDB work
