+++ REFERENCES.md
@@ -10,15 +10,18 @@
 - AR Saperstein, _et al._ (2025), _"Design and development of an off-normal warning system for SPARC"_, Nucl. Fusion **65** 116007, DOI: [10.1088/1741-4326/ae074e](https://doi.org/10.1088/1741-4326/ae074e)
 - GL Trevisan, _et al._ (2025), _"DisruptionPy: An open-source physics-based scientific framework for disruption analysis of fusion plasmas"_, JOSS, [submitted](https://github.com/openjournals/joss-reviews/issues/8538)
 - AD Maris, _et al._ (2024), _"Correlation of the L-mode density limit with edge collisionality"_, Nucl. Fusion **65** 016051, DOI: [10.1088/1741-4326/ad90f0](https://doi.org/10.1088/1741-4326/ad90f0)
+- Z Keith, _et al._ (2024), _"Risk-aware framework development for disruption prediction: Alcator C-Mod and DIII-D survival analysis"_, JoFE 43:21, DOI: [10.1007/s10894-024-00413-y](https://doi.org/10.1007/s10894-024-00413-y)
 
 ### Conferences
 
 - [67th APS-DPP Meeting (2025)](https://meetings.aps.org/Meeting/DPP25)
   - AD Maris, _et al._ (2025), _"Collisionality scaling of the tokamak density limit: data-driven analysis, cross-device prediction, and real-time avoidance"_
   - H Wietfeldt, _et al._ (2025), _"Assessing the Risk of UFO Disruptions to SPARC via a Multi-device Study"_
+  - Z Keith, _et al._ (2025), _"Enabling data-driven NTM studies with advanced mode labeling"_
 
 - [6th International Conference on Data-Driven Plasma Science (2025)](https://web.cvent.com/event/7de9d238-e170-4fbf-8de5-20abc5c6eb49/summary)
   - Y Wei, _et al._ (2025), _"DisruptionPy: An open-source physics-based scientific package for disruption studies on magnetic fusion experiment devices"_
+  - Z Keith, _et al._ (2025), _"A tearing mode database for Alcator C-Mod and DIII-D"_
 
 - [45th ITPA-MDC Meeting (2025)](https://sites.google.com/pppl.gov/itpa-mdc2025)
   - AR Saperstein, _et al._ (2025), _"Progress on the development of an off-normal warning system for SPARC"_
