### 2025-10-22 746f1d2
Remove _check_mds_data and its references. Postpone this to a later PR.
