### 2025-10-22 a305865
Fix n2_amp computation
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
