### 2025-10-22 cd69ca2
Implement HBT-EP physics methods (#476)
* add skeleton for HBT-EP

* no SQL server for HBT-EP

* fix HBT-EP config structures

* add hbtep methods to docs

* add HBT-EP to readme

* add HBT-EP docs to menu

* Added IpTimeSetting and get_ip physics method for hbtep

* Add HBTEP to get_method_holders.

* Add get_cap_bank_parameters and get_plasma_radii

* Add get_btor and get_qstar

* Add get_v_loop

* Add get_h_alpha

* Add get_sxr_midplane

* Implement get_n_mode_data and associated functions

* Minor changes

* Add get_m_mode_data and associated functions

* Remove redundant variables and TOD
[…]
