### 2025-10-22 6b0bb98
Fix logic for using alternative density data source
