### 2025-10-22 f7ad91f
cleaner if statement
