### 2025-10-22 ce52fed
Set tree_name=None when fetching data from PTDATA to avoid missing tree error
