+++ disruption_py/machine/d3d/physics.py
@@ -321,15 +321,13 @@ class D3DPhysicsMethods:
         """
         # Get edge loop voltage and smooth it a bit with a median filter
         v_loop, t_v_loop = params.mds_conn.get_data_with_dims(
-            f'ptdata("vloopb", {params.shot_id})', tree_name="d3d"
+            f'ptdata("vloopb", {params.shot_id})'
         )
         t_v_loop /= 1e3  # [ms] -> [s]
         v_loop = scipy.signal.medfilt(v_loop, 11)
         v_loop = interp1(t_v_loop, v_loop, params.times, "linear")
         # Get plasma current
-        ip, t_ip = params.mds_conn.get_data_with_dims(
-            f"ptdata('ip', {params.shot_id})", tree_name="d3d"
-        )
+        ip, t_ip = params.mds_conn.get_data_with_dims(f"ptdata('ip', {params.shot_id})")
         t_ip /= 1e3  # [ms] -> [s]
 
         # Alessandro Pau (JET & AUG) has given Cristina a robust routine that
@@ -447,7 +445,7 @@ class D3DPhysicsMethods:
         )
         try:
             ip, t_ip = params.mds_conn.get_data_with_dims(
-                f"ptdata('ip', {params.shot_id})", tree_name="_efit_tree"
+                f"ptdata('ip', {params.shot_id})"
             )  # [A], [ms]
             t_ip = t_ip / 1.0e3  # [ms] -> [s]
             ipsign = np.sign(np.sum(ip))
@@ -510,7 +508,7 @@ class D3DPhysicsMethods:
         - pull requests: #[251](https://github.com/MIT-PSFC/disruption-py/pull/251)
         """
         ne_rt, t_ne_rt = params.mds_conn.get_data_with_dims(
-            f"ptdata('dssdenest', {params.shot_id})", tree_name="_efit_tree"
+            f"ptdata('dssdenest', {params.shot_id})"
         )  # [10^19 m^-3]
         t_ne_rt = t_ne_rt / 1.0e3  # [ms] to [s]
         ne_rt = ne_rt * 1.0e19  # [10^19 m^-3] -> [m^-3]
@@ -615,7 +613,7 @@ class D3DPhysicsMethods:
         # Get measured plasma current parameters
         try:
             ip, t_ip = params.mds_conn.get_data_with_dims(
-                f"ptdata('ip', {params.shot_id})", tree_name="d3d"
+                f"ptdata('ip', {params.shot_id})"
             )  # [A], [ms]
             t_ip = t_ip / 1.0e3  # [ms] -> [s]
             dip_dt = np.gradient(ip, t_ip)
@@ -627,7 +625,7 @@ class D3DPhysicsMethods:
         # Get programmed plasma current parameters
         try:
             ip_prog, t_ip_prog = params.mds_conn.get_data_with_dims(
-                f"ptdata('iptipp', {params.shot_id})", tree_name="d3d"
+                f"ptdata('iptipp', {params.shot_id})"
             )  # [A], [ms]
             t_ip_prog = t_ip_prog / 1.0e3  # [ms] -> [s]
             polarity = D3DUtilMethods.get_polarity(params)
@@ -649,7 +647,7 @@ class D3DPhysicsMethods:
         # 'ip_prog' signal is irrelevant, and therefore 'ip_error' is not defined.
         try:
             ipimode, t_ipimode = params.mds_conn.get_data_with_dims(
-                f"ptdata('ipimode', {params.shot_id})", tree_name="d3d"
+                f"ptdata('ipimode', {params.shot_id})"
             )
             t_ipimode = t_ipimode / 1.0e3  # [ms] -> [s]
[…]
