### 2025-10-23 a693eee
Mostly done with geqdsk method for C-Mod, must conver to disruption-py standard with time under idx dim
