+++ disruption_py/machine/cmod/efit.py
@@ -7,6 +7,8 @@ Module for retrieving and processing EFIT parameters for CMOD.
 import numpy as np
 from MDSplus import mdsExceptions
 
+import xarray as xr
+
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
 from disruption_py.core.utils.math import interp1
@@ -171,6 +173,303 @@ class CmodEfitMethods:
         """
 
         return CmodEfitMethods._get_efit_equilibrium(params, params.times)
+    
+    @staticmethod
+    @physics_method(
+        tokamak=Tokamak.CMOD,
+    )
+    def get_geqdsk_parameters(params: PhysicsMethodParams):
+        """
+        Retrieve parameters required to create an GeqdskDataDict https://freeqdsk.readthedocs.io/en/stable/geqdsk.html
+        Also see https://fusion.gat.com/theory/Efitgeqdsk
+
+        This requires the following:
+        
+        1. shot
+        2. grid parameters
+        - Either (nx, ny, rdim, zdim, rcenter, rleft, zmid) or (rgrid, zgrid)
+        3. scalar parameters
+        - rmagx, zmagx, simagx, sibdry, bccenter, cpasma, 
+        4. 1D profile parameters
+        - fpol, pres, ffprime, pprime, qpsi
+        5. 2D parameters
+        - psi or f or psirz
+        6. boundary and limiter data (if available)
+        - nbdry, nlim, rbdry, zbdry, rlim, zlim
+
+        Sometimes the C-Mod EFIT uses 'x' and 'y' instead of 'r' and 'z' for parameter names
+        This function will convert them to the standard GEQDSK naming convention.
+
+        Parameters
+        ----------
+        params : PhysicsMethodParams
+            Parameters containing MDS connection and shot information.
+
+        Returns
+        -------
+        xarray.Dataset
+            A dataset containing the GEQDSK parameters.
+        """
+        # Get EFIT time array
+        efit_time = params.mds_conn.get_data(r"\efit_a_eqdsk:atime", tree_name="_efit_tree")
+        
+        # Scalar parameters
+        scalar_params = {
+            # Grid dimensions
+            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),  # Number of horizontal grid points
+            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),  # Number of vertical grid points
+            
+            # Plasma parameters
+            "cpasma": params.mds_conn.get_data(r"\efit_g_eqdsk:cpasma", tree_name="_efit_tree"),  # Plasma current [A]
+            "bcentr": params.mds_conn.get_data(r"\efit_g_eqdsk:bcentr", tree_name="_efit_tree"),  # Vacuum toroidal field [T]
+            
+            # Magnetic axis
+            "rmaxis": params.mds_conn.get_data(r"\efit_g_eqdsk:rmaxis", tree_name="_efit_tree"), # R of magnetic axis [m]
+            "zmaxis": params.mds_conn.get_data(r"\efit_g_eqdsk:zmaxis", tree_name="_efit_tree"), # Z of magnetic axis [m]
+            
+            # Plasma boundary
+            # Note that ssimag is -1 * simagx from the aeqdsk, apparently
+            "ssimag": params.mds_conn.get_data(r"\efit_g_eqdsk:ssimag", tree_name="_efit_tree"),   # Flux at magnetic axis [Wb]
[…]
