### 2025-10-23 4477b9d
Finished first pass at C-Mod geqdsk signals
