+++ disruption_py/machine/cmod/efit.py
@@ -200,6 +200,9 @@ class CmodEfitMethods:
         Sometimes the C-Mod EFIT uses 'x' and 'y' instead of 'r' and 'z' for parameter names
         This function will convert them to the standard GEQDSK naming convention.
 
+        Also, since it doesn't really make sense to get interpolated 'nbbbs', we're not including it in the dataset.
+        Later you can look at the nonzero values in 'rbbbs' or 'zbbbs' to determine how many boundary points there are at each time.
+
         Parameters
         ----------
         params : PhysicsMethodParams
@@ -212,13 +215,19 @@ class CmodEfitMethods:
         """
         # Get EFIT time array
         efit_time = params.mds_conn.get_data(r"\efit_a_eqdsk:atime", tree_name="_efit_tree")
-        
+
+        # Grid parameters (For C-Mod a lot are missing, but can be inferred from rgrid/zgrid)
+        grid_params = {
+            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),        # Number of horizontal grid points
+            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),        # Number of vertical grid points
+            "rdim": params.mds_conn.get_data(r"\efit_g_eqdsk:xdim", tree_name="_efit_tree"),    # Horizontal dimension of grid [m]
+            "zdim": params.mds_conn.get_data(r"\efit_g_eqdsk:zdim", tree_name="_efit_tree"),    # Vertical dimension of grid [m]
+        }
+        rgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:rgrid", tree_name="_efit_tree"),  # R grid [m]
+        zgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:zgrid", tree_name="_efit_tree"),  # Z grid [m]
+
         # Scalar parameters
         scalar_params = {
-            # Grid dimensions
-            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),  # Number of horizontal grid points
-            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),  # Number of vertical grid points
-            
             # Plasma parameters
             "cpasma": params.mds_conn.get_data(r"\efit_g_eqdsk:cpasma", tree_name="_efit_tree"),  # Plasma current [A]
             "bcentr": params.mds_conn.get_data(r"\efit_g_eqdsk:bcentr", tree_name="_efit_tree"),  # Vacuum toroidal field [T]
@@ -231,14 +240,6 @@ class CmodEfitMethods:
             # Note that ssimag is -1 * simagx from the aeqdsk, apparently
             "ssimag": params.mds_conn.get_data(r"\efit_g_eqdsk:ssimag", tree_name="_efit_tree"),   # Flux at magnetic axis [Wb]
             "ssibry": params.mds_conn.get_data(r"\efit_g_eqdsk:ssibry", tree_name="_efit_tree"),   # Flux at plasma boundary [Wb]
-        }
-
-        # Grid boundaries (For C-Mod a lot are missing, but can be inferred from rgrid/zgrid)
-        grid_params = {
-            "rdim": params.mds_conn.get_data(r"\efit_g_eqdsk:xdim", tree_name="_efit_tree"),     # Horizontal dimension of grid [m]
-            "zdim": params.mds_conn.get_data(r"\efit_g_eqdsk:zdim", tree_name="_efit_tree"),     # Vertical dimension of grid [m]
[…]
