### 2025-10-23 c8edcab
Fixed indexing for Mirnov fft methods
