### 2025-10-27 f37edbe
bugfix #165931 get_density_parameters
