### 2025-10-27 db0d0e8
bugfix #174087 | get_rt_ip_parameters: UnboundLocalError("cannot access local variable 'power_supply_railed' where it is not associated with a value")
