### 2025-10-27 9baf4e6
cleanup ~np.isnan(x).all() expressions
