### 2025-10-27 8e0d91c
Temporary fix for #174756 | get_density_parameters: IndexError('index 0 is out of bounds for axis 0 with size 0')
