### 2025-10-27 d9b4c66
Add verbose level across all loggers
