### 2025-10-27 b682067
complete and reorder aps dpp 2025 refs by id
