### 2025-10-28 3cd4be6
Update fsspec deps
