+++ pyproject.toml
@@ -58,8 +58,6 @@ dependencies = [
   "tqdm>=4.67.0",
   "xarray>=2025.1.0,<2025.8.0",
   "zarr>=3",
-  "aiohttp (>=3.12.15,<4.0.0)",
-  "requests (>=2.32.5,<3.0.0)",
 ]
 description = "An open-source physics-based Scientific Framework for Disruption Analysis of Fusion Plasmas for AI/ML applications"
 keywords = [
