### 2025-10-28 9ec8fa8
pin py3.12 for actions
