+++ .github/workflows/build.yml
@@ -37,7 +37,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v6
         with:
-          python-version-file: 'pyproject.toml'
+          python-version: '3.12'
           cache: 'poetry'
 
       - name: Print versions
