### 2025-10-28 e45c633
pin py3.12 for actions (#480)
