+++ .github/workflows/tests.yml
@@ -117,7 +117,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v6
         with:
-          python-version-file: 'pyproject.toml'
+          python-version: '3.12'
           cache: 'poetry'
 
       - name: Install requirements
