### 2025-10-28 6c7a4bc
Fix bug in checking feedback_off_indices logic
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
