### 2025-10-28 21b9c23
Fix DIII-D physics methods' errors (#479)
* fix unsafe casting in get_power_parameters

* Fix #170648 | get_ip_parameters: TypeError

* Fix #170648 | get_time_domain: TypeError

* Fix TypeError("'numpy.float32' object does not support item assignment")

* Minor fix to d3d get_peaking_factors

* Minor changes to _check_mds_data

* pylint

* Improve logic in retrieving density data

* Fix #170648 | get_rt_ip_parameters: TypeError('only integer scalar arrays can be converted to a scalar index')

* Remove _check_mds_data and its references. Postpone this to a later PR.

* Fix logic for using alter
[…]
