+++ disruption_py/machine/d3d/physics.py
@@ -170,7 +170,7 @@ class D3DPhysicsMethods:
                 r"\top.nb:pinj", tree_name="d3d"
             )
             t_nbi /= 1e3  # [ms] -> [s]
-            p_nbi *= 1e3  # [KW] -> [W]
+            p_nbi = p_nbi * 1e3  # [KW] -> [W]
             if len(t_nbi) > 2:
                 p_nbi = interp1(
                     t_nbi,
@@ -321,15 +321,13 @@ class D3DPhysicsMethods:
         """
         # Get edge loop voltage and smooth it a bit with a median filter
         v_loop, t_v_loop = params.mds_conn.get_data_with_dims(
-            f'ptdata("vloopb", {params.shot_id})', tree_name="d3d"
+            f'ptdata("vloopb", {params.shot_id})'
         )
         t_v_loop /= 1e3  # [ms] -> [s]
         v_loop = scipy.signal.medfilt(v_loop, 11)
         v_loop = interp1(t_v_loop, v_loop, params.times, "linear")
         # Get plasma current
-        ip, t_ip = params.mds_conn.get_data_with_dims(
-            f"ptdata('ip', {params.shot_id})", tree_name="d3d"
-        )
+        ip, t_ip = params.mds_conn.get_data_with_dims(f"ptdata('ip', {params.shot_id})")
         t_ip /= 1e3  # [ms] -> [s]
 
         # Alessandro Pau (JET & AUG) has given Cristina a robust routine that
@@ -407,15 +405,28 @@ class D3DPhysicsMethods:
         ption-py/blob/matlab/DIII-D/get_density_parameters.m)
         - pull requests: #[249](https://github.com/MIT-PSFC/disruption-py/pull/249)
         """
-        ne, t_ne = params.mds_conn.get_data_with_dims(
-            r"\density", tree_name="_efit_tree"
-        )
-        # If EFIT disruption tree does not contain density data,
+        try:
+            ne, t_ne = params.mds_conn.get_data_with_dims(
+                r"\density", tree_name="_efit_tree"
+            )
+        except mdsExceptions.MdsException:
+            ne = [np.nan]
+            t_ne = [np.nan]
+        # If EFIT disruption tree does not contain density data or if the data is invalid,
         # then read density from BCI subtree of D3D main tree
-        # TODO: Find a shot to test this logic
-        if len(~np.isnan(ne)) == 0:
+        # Example 1: 170480:
+        #  - "\efit01:density" gives ne = array([], shape=(4, 0), dtype=float64),
+        #  - "\d3d:denv2" gives actual density data
+        # Example 2: 172756:
+        #  - r"\efit01:density" gives ne = array([4.06199e19]), t_ne = array([1800])
+        #  - "\d3d:denv2" gives actual density data
+        if not np.isfinite(ne).any() or len(ne) < 2:
             ne, t_ne = params.mds_conn.get_data_with_dims(r"\denv2", tree_name="d3d")
-
+            tree_name = params.mds_conn.get_tree_name_of_nickname("_efit_tree")
+            params.logger.verbose(
+                rf"density: data from \{tree_name}:density is either empty or invalid."
+                r" Use \d3d:denv2 instead."
+            )
         ne = ne * 1.0e6  # [cm^3] -> [m^3]
         t_ne = t_ne / 1.0e3  # [ms] -> [s]
         dne_dt = np.gradient(ne, t_ne)
@@ -437,7 +448,7 @@ class D3DPhysicsMethods:
         )
         try:
             ip, t_ip = params.mds_conn.get_data_with_dims(
-                f"ptdata('ip', {params.shot_id})", tree_name="_efit_tree"
+                f"ptdata('ip', {params.shot_id})"
             )  # [A], [ms]
[…]
