### 2025-10-28 8325783
include refs in abstract
