### 2025-10-28 6a5b912
explicit citation in readme
