### 2025-10-28 ace312b
add zenodo to publications
