### 2025-10-28 1f79888
add AI4F CRP 2025
