### 2025-10-29 3f2227c
exclude py3.14 for now
