### 2025-10-29 e12402d
add new XFAILs for EAST (#482)
