### 2025-10-29 85fef07
pin pandas and netcdf4 for now
