### 2025-11-03 7ed3c0c
Update deps to Oct 2025 (#483)
* exclude py3.14 for now

* unpin xarray

* pin pandas and netcdf4 for now

* lock deps
