+++ pyproject.toml
@@ -48,14 +48,14 @@ classifiers = [
 dependencies = [
   "dynaconf>=3.2.0",
   "loguru>=0.7.0",
-  "netcdf4>=1.7.0",
+  "netcdf4>=1.7.0,<1.7.3",
   "numpy>=1.26.0,<2.0.0",
-  "pandas>=2.2.0",
+  "pandas>=2.2.0,<2.3.3",
   "pyodbc>=5.2.0",
   "scipy>=1.15.0",
   "SQLAlchemy>=2.0.0",
   "tqdm>=4.67.0",
-  "xarray>=2025.1.0,<2025.8.0",
+  "xarray>=2025.1.0",
 ]
 description = "An open-source physics-based Scientific Framework for Disruption Analysis of Fusion Plasmas for AI/ML applications"
 keywords = [
@@ -70,7 +70,7 @@ maintainers = [
 ]
 name = "disruption-py"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.11,<3.14"
 version = "0.12.0-dev"
 
 [project.scripts]
