### 2025-11-03 a2b7e2a
add release target to makefile
