### 2025-11-03 b36f58f
fixup! add release target to makefile
