### 2025-11-03 a03d52d
Add release target in makefile (#485)
* add release target to makefile

* fixup! add release target to makefile
