+++ disruption_py/machine/d3d/efit.py
@@ -134,6 +134,27 @@ class D3DEfitMethods:
     def get_geqdsk_parameters(params: PhysicsMethodParams):
         """
         Retrieve parameters required to create an GeqdskDataDict https://freeqdsk.readthedocs.io/en/stable/geqdsk.html
+        Also see https://fusion.gat.com/theory/Efitgeqdsk
+
+        This requires the following:
+        
+        1. shot
+        2. grid parameters
+        - Either (nx, ny, rdim, zdim, rcenter, rleft, zmid) or (rgrid, zgrid)
+        3. scalar parameters
+        - rmagx, zmagx, simagx, sibdry, bccenter, cpasma, 
+        4. 1D profile parameters
+        - fpol, pres, ffprime, pprime, qpsi
+        5. 2D parameters
+        - psi or f or psirz
+        6. boundary and limiter data (if available)
+        - nbdry, nlim, rbdry, zbdry, rlim, zlim
+
+        Some signals are missing on DIII-D:
+        - rcentr
+
+        Also, since it doesn't really make sense to get interpolated 'nbbbs', we're not including it in the dataset.
+        Later you can look at the nonzero values in 'rbbbs' or 'zbbbs' to determine how many boundary points there are at each time.
 
         Parameters
         ----------
@@ -147,36 +168,36 @@ class D3DEfitMethods:
         """
         # Get EFIT time array
         efit_time = params.mds_conn.get_data(
-            r"\efit_g_eqdsk:atime", tree_name="_efit_tree"
+            r"\efit_a_eqdsk:atime", tree_name="_efit_tree"
         ) / 1.0e3  # [ms] -> [s]
+
+        # Grid parameters
+        # For DIII-D, rcentr and rleft are missing but can be inferred from rgrid
+        grid_params = {
+            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),        # Number of horizontal grid points
+            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),        # Number of vertical grid points
+            "rdim": params.mds_conn.get_data(r"\efit_g_eqdsk:xdim", tree_name="_efit_tree"),    # Horizontal dimension of grid [m]
+            "zdim": params.mds_conn.get_data(r"\efit_g_eqdsk:zdim", tree_name="_efit_tree"),    # Vertical dimension of grid [m]
+            "zmid": params.mds_conn.get_data(r"\efit_g_eqdsk:zmid", tree_name="_efit_tree"),     # Z at center of grid [m]
+        }
+        rgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:r", tree_name="_efit_tree"),  # R grid [m]
+        zgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:z", tree_name="_efit_tree"),  # Z grid [m]
         
         # Scalar parameters
         scalar_params = {
-            # Grid dimensions
-            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),  # Number of horizontal grid points
-            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),  # Number of vertical grid points
-            
             # Plasma parameters
             "cpasma": params.mds_conn.get_data(r"\efit_g_eqdsk:cpasma", tree_name="_efit_tree"),  # Plasma current [A]
             "bcentr": params.mds_conn.get_data(r"\efit_g_eqdsk:bcentr", tree_name="_efit_tree"),  # Vacuum toroidal field [T]
[…]
