### 2025-11-04 40da194
Changed frequency resolution to avoid non-causal data
