### 2025-11-05 904d339
Added probe type to C-mod, removed theta
