+++ disruption_py/machine/cmod/mirnov.py
@@ -117,7 +117,7 @@ class CmodMirnovMethods:
         # These are the coils without positions in MDSPlus, need to hard-code values.
         # Taken from here: https://cmodwiki.psfc.mit.edu/index.php/FastMagneticsLocations#2010_Locations
         # Might need to include some logic for pre and post 2010... TODO(ZanderKeith)
-        # The only things that change pre and post 2010 are the EF probes, so 
+        # The only things that change pre and post 2010 are the EF probes
 
         mirnov_names_tab = ["BP1T_ABK", "BP2T_ABK", "BP3T_ABK", "BP4T_ABK", "BP5T_ABK", "BP6T_ABK"]
         mirnov_names_tgh = ["BP1T_GHK", "BP2T_GHK", "BP3T_GHK", "BP4T_GHK", "BP5T_GHK", "BP6T_GHK"]
@@ -126,16 +126,16 @@ class CmodMirnovMethods:
 
         mirnov_names_diff = ["BP_EF_TOP", "BP_EF_BOT"]
 
-        phi_tab = [-23.10, -25.50, -27.90, -23.10, -25.50, -27.90]
-        phi_tgh = [-224.40, -226.80, -229.20, -224.40, -226.80, -229.20]
-        phi_top = [-344.80, -10.16, -59.87, -169.55]
-        phi_bot = [-344.80, -59.87, -169.55]
-
-        # Yes, the order of these is different from the c-mod wiki website. I'm putting it like phi, R, Z, theta_pol
+        # Yes, the order of these is different from the c-mod wiki website. I'm putting it like R, phi, Z, theta_pol
         R_tab = [0.9045, 0.9045, 0.9045, 0.9045, 0.9045, 0.9045]
         R_tgh = [0.9042, 0.9042, 0.9042, 0.9042, 0.9042, 0.9042]
         R_top = [0.9126, 0.9126, 0.9146, 0.9126]
         R_bot = [0.9131, 0.9151, 0.9131]
+
+        phi_tab = [-23.10, -25.50, -27.90, -23.10, -25.50, -27.90]
+        phi_tgh = [-224.40, -226.80, -229.20, -224.40, -226.80, -229.20]
+        phi_top = [-344.80, -10.16, -59.87, -169.55]
+        phi_bot = [-344.80, -59.87, -169.55]
 
         Z_tab = [0.1030, 0.1030, 0.1030, -0.1030, -0.1030, -0.1030]
         Z_tgh = [0.1000, 0.1000, 0.1000, -0.1000, -0.1000, -0.1000]
@@ -149,22 +149,18 @@ class CmodMirnovMethods:
 
         all_mirnov_names = mirnov_names_ab + mirnov_names_gh + mirnov_names_k + mirnov_names_tab + mirnov_names_tgh + mirnov_names_top + mirnov_names_bot
 
+        R_all = np.concatenate((R_ab, R_gh, R_k, R_tab, R_tgh, R_top, R_bot))
         phi_all = np.concatenate((phi_ab, phi_gh, phi_k, phi_tab, phi_tgh, phi_top, phi_bot))
         phi_all = np.deg2rad(phi_all)
-        R_all = np.concatenate((R_ab, R_gh, R_k, R_tab, R_tgh, R_top, R_bot))
         Z_all = np.concatenate((Z_ab, Z_gh, Z_k, Z_tab, Z_tgh, Z_top, Z_bot))
-
-        # The angle we care about is when the hypotenuse is the minor radius
-        Rd = R_all - CMOD_R0
-        theta_all = np.arctan2(Z_all, Rd)
 
[…]
