### 2025-11-05 442bb45
Refactored to use 'sensor'
