+++ disruption_py/machine/cmod/mirnov.py
@@ -75,7 +75,7 @@ class CmodMirnovMethods:
         mirnov_names_gh = [f"BP0{p}_GHK" for p in range(1, 10)] + [f"BP{p}_GHK" for p in range(10, 21)]
         mirnov_names_k = [f"BP0{p}_K" for p in range(1, 7)]
 
-        # There's some tomfoolery in here because the probe locations are in a different tree from the signals, so we need to cut off the extra probes
+        # There's some tomfoolery in here because the sensor locations are in a different tree from the signals, so we need to cut off the extra sensors
         # that are NOT being digitized
         phi_ab, _ = params.mds_conn.get_data_with_dims(r"\magnetics::top.rf_lim_coils.phi_AB", tree_name="magnetics")
         phi_gh, _ = params.mds_conn.get_data_with_dims(r"\magnetics::top.rf_lim_coils.phi_GH", tree_name="magnetics")
@@ -94,10 +94,10 @@ class CmodMirnovMethods:
         # theta_pol_k, _ = params.mds_conn.get_data_with_dims(r"\magnetics::top.rf_lim_coils.theta_pol_K", tree_name="magnetics")  # noqa: ERA001
         theta_pol_k = np.empty_like(Z_k)  # Calibration data doesn't exist, so we'll just say NaN and deal with it later
 
-        # For each of the above, we need to cut off the extra probes that are NOT being digitized
-        # This is a bit of a kludge, but I'm expecting probes which aren't digitized to not show up in the analysis tree
-        # TODO(ZanderKeith): Just hardcode all the analysis tree probes
-        # The digitized probes are the first 20 for AB and GH, and the first 6 for K
+        # For each of the above, we need to cut off the extra sensors that are NOT being digitized
+        # This is a bit of a kludge, but I'm expecting sensors which aren't digitized to not show up in the analysis tree
+        # TODO(ZanderKeith): Just hardcode all the analysis tree sensors
+        # The digitized sensors are the first 20 for AB and GH, and the first 6 for K
         phi_ab = phi_ab[:len(mirnov_names_ab)]
         phi_gh = phi_gh[:len(mirnov_names_gh)]
         phi_k = phi_k[:7]
@@ -117,7 +117,7 @@ class CmodMirnovMethods:
         # These are the coils without positions in MDSPlus, need to hard-code values.
         # Taken from here: https://cmodwiki.psfc.mit.edu/index.php/FastMagneticsLocations#2010_Locations
         # Might need to include some logic for pre and post 2010... TODO(ZanderKeith)
-        # The only things that change pre and post 2010 are the EF probes
+        # The only things that change pre and post 2010 are the EF sensors
 
         mirnov_names_tab = ["BP1T_ABK", "BP2T_ABK", "BP3T_ABK", "BP4T_ABK", "BP5T_ABK", "BP6T_ABK"]
         mirnov_names_tgh = ["BP1T_GHK", "BP2T_GHK", "BP3T_GHK", "BP4T_GHK", "BP5T_GHK", "BP6T_GHK"]
@@ -219,8 +219,8 @@ class CmodMirnovMethods:
         -------
         mirnov_ffts : xarray.DataSet
             The FFT of the Mirnov coils.
-            Dimensions are probe, frequency, and time.
-            Coordinates are probe, frequency, time, phi, theta, and theta_pol.
+            Dimensions are sensor, frequency, and time.
+            Coordinates are sensor, frequency, time, phi, theta, and theta_pol.
         """
 
         all_mirnov_names, R_all, phi_all, Z_all, theta_pol_all = CmodMirnovMethods.get_mirnov_names_and_locations(params, debug=False)
[…]
