### 2025-11-06 #8538 @LiamPattinson
Hi @jatkinson1000 and @gtrevisan, I'd be happy to review. I'm working on the [PlasmaFAIR project](https://plasmafair.github.io/) at the University of York Plasma Institute, so the fact I'm already working within fusion might make it easier to arrange access to the private data server.
