### 2025-11-06 #9364 @editorialbot
[REVIEW]: DisruptionPy: An open-source physics-based scientific framework for disruption analysis of fusion plasmas
**Submitting author:** @gtrevisan (<a href="http://orcid.org/0000-0003-4920-6109">Gregorio L. Trevisan</a>)
**Repository:** https://github.com/MIT-PSFC/disruption-py
**Branch with paper.md** (empty if default branch): glt/joss
**Version:** v0.14.0
