### 2025-11-06 #9364 @gtrevisan
Hi all, thanks for helping out with our submission!

I am planning to release [v0.12](https://github.com/MIT-PSFC/disruption-py/pull/455) in the next few days, ahead of our [yearly APS-DPP meeting](https://engage.aps.org/dpp/meetings/annual-meeting), thus I will soon merge the current `dev` branch onto `main`, and I will branch off a new development branch for v0.13.
Due to timing alone, I imagine any proposed feature or bugfix would happen as a new PR onto v0.13-dev.

@jatkinson1000 Please let me know whether JOSS requires any specific development cycle, or enforces any other rule or logic (with impact to our current release workflow) that I should be aware of.

Allow me then to quote [my own comment from the pre-review thread](https://github.com/openjournals/joss-reviews/issues/8538#issuecomment-3427021571) for our reviewers' benefit:

> ### Alcator C-Mod
>
> In order to access data for [Alcator C-Mod](https://en.wikipedia.org/wiki/Alcator_C-Mod) (MIT's own now-decommissioned machine, prototypical for these workflows) one has to _request access_ -- any reviewer might do that easily, and I would make sure any request is pushed forward in a timely fashion. That's because the [MDSplus server](https://www.mdsplus.org/) hosting our C-Mod data is not public-facing and does require authentication and computer access on our [MIT PSFC](https://www.psfc.mit.edu/) systems.
> ### MAST
>
> We currently have UKAEA collaborators helping out and implementing physics methods for [MAST data, which is open access](https://mastapp.site/). Since the S3 endpoints hosting MAST experimental data as Zarr-backed Xarrays are world-readable, anybody can try out our MAST workflows at any time, anywhere.
>
> (Other machines have more stringent requirements for accessing data.)
>
> The framework code is shared among all devices, and only the physics methods will be machine-specific. All these methods are open-source so as to allow anybody to take a look at our physics workflows, but they will pull data from the appropriate authorized sources.
>
> I guess it depends on what level of detail each reviewer might deem appropriate to test our code -- I believe the MAST proof-of-concept might be enough for most purposes but, if not, C-Mod access will be granted easily. How would you like to proceed?

Thank you all!
