### 2025-11-06 #9364 @LiamPattinson
I'll be (extremely) busy until November 17th, and will be able to start my review then. Apologies for not bringing this up in the previous thread, I hope it won't be an issue! In the meantime, please let me know if I need to provide any details in order to gain access to the MDSplus server.
