### 2025-11-06 ff2e8de
replace APS25 epitome with schedule
