+++ REFERENCES.md
@@ -20,14 +20,13 @@ Here follows a non-exhaustive list of projects, publications, and conference con
 ### Conferences
 
 - [67th APS-DPP Meeting (2025)](https://meetings.aps.org/Meeting/DPP25)
-  - AD Maris, _et al._ (2025), _"Collisionality scaling of the tokamak density limit: data-driven analysis, cross-device prediction, and real-time avoidance"_ [BO04.5](https://meetings.aps.org/Meeting/DPP25/Session/BO04.5)
-  - Z Keith, _et al._ (2025), _"Enabling data-driven NTM studies with advanced mode labeling"_, [BP13.167](https://meetings.aps.org/Meeting/DPP25/Session/BP13.167)
-  - C Rea, _et al._ (2025), _"Research in support of the SPARC Off-Normal Warning System"_, [JO04.9](https://meetings.aps.org/Meeting/DPP25/Session/JO04.9)
-  - AR Saperstein, _et al._ (2025), _"Validation of simulated radiative collapse events in TORAX"_, [NP13.161](https://meetings.aps.org/Meeting/DPP25/Session/NP13.161)
-  - H Wietfeldt, _et al._ (2025), _"Characterization of UFOs on Alcator C-Mod and WEST to inform SPARC operation"_ [NP13.163](https://meetings.aps.org/Meeting/DPP25/Session/NP13.163)
-  - EdD Zapata-Cornejo, _et al._ (2025), _"Time series classification algorithms for confinement regime identification in C-Mod"_ [NP13.170](https://meetings.aps.org/Meeting/DPP25/Session/NP13.170)
-  - GL Trevisan, _et al._ (2025), _"A large-scale automated EFIT recomputation workflow for disruption studies at 1 kHz"_, [PP13.93](https://meetings.aps.org/Meeting/DPP25/Session/PP13.93)
-  - Y Wei, _et al._ (2025), _"Scikit-disruption: machine learning toolkit for disruption analysis"_, [PP13.98](https://meetings.aps.org/Meeting/DPP25/Session/PP13.98)
+  - AD Maris, _et al._ (2025), _"Collisionality scaling of the tokamak density limit: data-driven analysis, cross-device prediction, and real-time avoidance"_ [BO04.5](https://schedule.aps.org/dpp/2025/events/BO04/5)
+  - Z Keith, _et al._ (2025), _"Enabling data-driven NTM studies with advanced mode labeling"_, [BP13.167](https://schedule.aps.org/dpp/2025/events/BP13/167)
+  - C Rea, _et al._ (2025), _"Research in support of the SPARC Off-Normal Warning System"_, [JO04.9](https://schedule.aps.org/dpp/2025/events/JO04/9)
+  - AR Saperstein, _et al._ (2025), _"Validation of simulated radiative collapse events in TORAX"_, [NP13.161](https://schedule.aps.org/dpp/2025/events/NP13/161)
+  - H Wietfeldt, _et al._ (2025), _"Characterization of UFOs on Alcator C-Mod and WEST to inform SPARC operation"_ [NP13.163](https://schedule.aps.org/dpp/2025/events/NP13/163)
+  - EdD Zapata-Cornejo, _et al._ (2025), _"Time series classification algorithms for confinement regime identification in C-Mod"_ [NP13.169](https://schedule.aps.org/dpp/2025/events/NP13/169)
+  - GL Trevisan, _et al._ (2025), _"A large-scale automated EFIT recomputation workflow for disruption studies at 1 kHz"_, [PP13.93](https://schedule.aps.org/dpp/2025/events/PP13/93)
 
[…]
