### 2025-11-06 #9364 @gtrevisan
@LiamPattinson @jackleland @rmrsk

You may request an account on our systems [from this webpage](https://account.psfc.mit.edu/).
I'd choose:
- Position Type/Affiliation = Collaborator/s
- Supervisor = myself
- Job Description = testing of the disruption-py workflow with C-mod data for publication review
- Account type = Windows and Linux

You are also free to test out the non-authwalled [MAST workflow](https://github.com/MIT-PSFC/disruption-py/commits/slj/mast/) -- that workflow is the first to rely on Xarray, rather than MDSplus, and as such is still very much under active development.
We will enlist UKAEA's own @samueljackson92 if needed. 🙃
