### 2025-11-07 bc485ed
Create new References page (#474)
* create new REFERENCES page

* add  recent ARS paper

* always put et al in italics

* added Saperstein conferences to REFERENCES.md

* tweak references

* reorder, add next and previous APS DPP confs

* Added Maris references

* Added Wei's talk at ICDDPS-6 & poster at OSSFE-25

* added wietfeldt references

* Added references for survival analysis and in-progress TMDB work

* add MDSplusML paper

* fix JoFE paper

* order aps dpp 2024 refs by id

* complete and reorder aps dpp 2025 refs by id

* reorder by id or alphabetically

* fixup! reorder by id or alp
[…]
