### 2025-11-07 3e6f1ec
Add ip threshold params for MAST
Also fix one or two small inconsistencies
