### 2025-11-07 558e9ba
move install.md back into main folder
