### 2025-11-07 #9364 @gtrevisan
@LiamPattinson @jackleland @rmrsk
If you wish to access Alcator C-Mod data, please read and sign [this Data Agreement](https://drive.google.com/file/d/1YgOJ2mG1syN8U_lAcsR3CnttHA29ySUC) and return it to me via email -- I will then countersign it and collect our Division Head's approval.
