### 2025-11-07 cb5b02c
Minor documentation tweaks (#486)
* move install.md back into main folder

* add shortcut to install in docs

* add references to main menu

* prefer FLOSS to FOSS

* mark HBT-EP as supported, figure by Y Wei
