+++ INSTALL.md
@@ -1,0 +1,185 @@
+
+# Installation
+
+Several _public_ installations are currently maintained automatically:
+
+- on the C-MOD MFE workstations,
+- on the DIII-D `omega` cluster,
+- on the EAST `cs` servers,
+- on the HBT-EP `landau` server.
+
+A _private_ installation is possible on any machine, either on a personal laptop, or on any experimental cluster.
+
+## Public installations
+
+The MIT PSFC Disruption Studies Group hosts a `disruption-py` installation in a NFS folder.
+The following steps apply to both C-MOD workstations and the DIII-D cluster.
+
+Snippets for quick addition to a user's `.bashrc` file are provided in the subsections below.
+
+#### Python
+
+A minimal Python 3.12 distribution is installed in the `"$DISPY_DIR"/cpython/3.12` folder, and can be used in order to create further virtual environments.
+
+#### Dependency management
+
+We use [Poetry](https://python-poetry.org/) and [uv](https://docs.astral.sh/uv/) for dependency management.
+
+The helper scripts automatically activate the appropriate virtual environment, so direct Poetry or uv usage is not needed.
+They may be used directly from the `"$DISPY_DIR"/poetry` or `"$DISPY_DIR"/uv` folders.
+
+#### Branches
+
+Two branches are installed publicly and kept up to date:
+
+- `main`, for stable workflows;
+- `dev`, for fresh features.
+
+The target branch can be controlled through the `DISPY_BRANCH` environment variable. 
+
+#### Virtual environments
+
+For each branch, a Python 3.12 virtual environment is available for usage.
+
+The helper scripts will choose the appropriate virtual environment based on the `DISPY_BRANCH` environment variable.
+It may also be used directly from the `"$DISPY_DIR"/venv` folder.
+
+#### Setup and activation
+
+A setup script will set all the required environment variables to ensure functionality and reproducibility.
+It has to be _sourced_, rather than executed, through:
+
+```bash
+source "$DISPY_DIR"/repo/auto/setup.sh
+```
+
+More often, a user may choose to directly _activate_ the chosen virtual environment, through:
+
+```bash
+source "$DISPY_DIR"/repo/auto/activate.sh
+# or the alias defined in our snippet
+disruption-activate
+```
+
+The helper scripts rely on the user adopting the [Bash shell](https://www.gnu.org/software/bash/).
+
+#### Execution
+
+Even without setup or activation, execution can be invoked through `disruption-python`, which should work as a drop-in replacement for the usual `python` command.
+Since `disruption-python` is an executable, its shorthand execution is made possible by the presence of the `"$DISPY_DIR"/repo/auto` folder within the `PATH` environment variable, as suggested in the installation-specific snippets below.
+
+For example, one may begin a disruption-py interactive session through:
+
+```bash
+disruption-python -ic "import disruption_py"
+```
+
+Or execute a disruption-py-based script through:
+
+```bash
+disruption-python workflow.py
+```
+
+A script can therefore be executed seamlessly through `disruption-python` by specifying a custom shebang:
+
+```
+user@host:~$ head -n1 workflow.py
+#!/usr/bin/env disruption-python
+
+user@host:~$ chmod +x workflow.py
+
+user@host:~$ ./workflow.py
+```
+
+All these helper scripts are subject to change -- we welcome any suggestion to make the process even smoother.
+
+### C-MOD
+
+Suggested snippet to be appended to the user's `~/.bashrc` file:
+
+```bash
+# disruption-py
[…]
