+++ pyproject.toml
@@ -71,7 +71,7 @@ maintainers = [
 name = "disruption-py"
 readme = "README.md"
 requires-python = ">=3.11,<3.14"
-version = "0.12.0-dev"
+version = "0.12.0"
 
 [project.scripts]
 disruption-py = "disruption_py.workflow:cli"
