+++ pyproject.toml
@@ -71,7 +71,7 @@ maintainers = [
 name = "disruption-py"
 readme = "README.md"
 requires-python = ">=3.11,<3.14"
-version = "0.12.0"
+version = "0.13.0-dev"
 
 [project.scripts]
 disruption-py = "disruption_py.workflow:cli"
