### 2025-11-07 #9364 @jatkinson1000
Thanks @kyleniemeyer I was just trying to check this.
My understanding - and I hope @gtrevisan can confirm - is that there is MAST data freely available to test the library with, and this discussion covers additional datasets with other permissions?

Whilst it is not uncommon for data providers to require an account to download data to reduce spamming (e.g. USGS, JAXA, Copernicus Climate Data) these are all simple and automated registration processes. If something requires wet signatures and manual approval then this is not in the spirit of open community software.
