### 2025-11-07 #9364 @gtrevisan
MAST data is [open access](https://mastapp.site/).
Access for the other machines (eg Alcator C-Mod) requires some kind of acknowledgement/authorization in order to make sure that collaborators follow the proper scientific channels (eg: before disseminating previously-unpublished results).

Data coming from ([_very_ expensive](https://www.cfs.energy/news-and-media/commonwealth-fusion-systems-raises-863-million-series-b2-round-to-accelerate-the-commercialization-of-fusion-energy)) fusion devices is often kept under stricter control for even longer, but thankfully that is not our case as [Alcator C-Mod](https://en.wikipedia.org/wiki/Alcator_C-Mod) stopped operating almost 10y ago.
Anybody can be granted access to it freely, all the more if it was publicly-funded research in the first place (eg: [DOE FES](https://www.energy.gov/science/fes/fusion-energy-sciences)).
It is true that our authorization process is clunky, but I don't see whether this should be any different than, say, having a series of [very complicated checkboxes with EULA terms written in tiny font](https://www.huffpost.com/entry/gamestation-grabs-souls-o_n_541549) on a streamlined webpage.

Furthermore, there is a huge difference between open _software_ and open _data_ -- when I submitted in July I did my best to read through all FAQs and I missed any reference to [openness](https://joss.readthedocs.io/en/latest/submitting.html#:~:text=open) of data.

Finally, allow me to mention once again that almost all parts of the framework will be invoked even by our MAST workflow, from the user configuration and settings, to the complicated portions of multiprocessing and physics method caching and invocation, through timebase interpolations and up to final output.
