### 2025-11-08 6000d41
Fixed problem where Mirnov frequency arrays were sometimes inconsistent
