### 2025-11-10 #9364 @jatkinson1000
@gtrevisan Now we've had the weekend to pause, please could you confirm if functionalities of your software package can be fully tested and evaluated using the [open source MAST dataset](https://mastapp.site/) that is freely available to reviewers without any contractual agreements?

@rmrsk my suggestion would be to pause whilst we wait for confirmation, though @kyleniemeyer has authority on this as TEiC.
