### 2025-12-02 3ec160a
Added DIII-D efit method for eqdsk data
