+++ disruption_py/machine/d3d/efit.py
@@ -5,6 +5,7 @@ Module for retrieving and processing EFIT parameters for DIII-D.
 """
 
 import numpy as np
+import xarray as xr
 
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
@@ -128,3 +129,253 @@ class D3DEfitMethods:
             for param in efit_data:
                 efit_data[param] = interp1(efit_time, efit_data[param], params.times)
         return efit_data
+
+    @staticmethod
+    @physics_method(
+        tokamak=Tokamak.D3D,
+    )
+    def get_geqdsk_parameters(params: PhysicsMethodParams):
+        """
+        Retrieve parameters required to create a GEQDSKFile https://freeqdsk.readthedocs.io/en/stable/geqdsk.html
+        Also see https://fusion.gat.com/theory/Efitgeqdsk
+        This requires the following:
+        
+        1. shot number
+        2. grid parameters: (nx, ny, rdim, zdim, rcenter, rleft, zmid) or (rgrid, zgrid)
+        3. scalar parameters: rmagx, zmagx, simagx, sibdry, bccenter, cpasma, etc.
+        4. 1D profile parameters: fpol, pres, ffprime, pprime, qpsi
+        5. 2D parameters: psi or f or psirz
+        6. boundary and limiter data (if available): nbdry, nlim, rbdry, zbdry, rlim, zlim
+
+        Some signals are missing on DIII-D but can be inferred from other signals, when this happens it is noted below.
+
+        Also, it doesn't make sense to get interpolated limiter data ('nbbbs', 'rbbbs', 'zbbbs') because the number of points can change over time.
+        TODO(ZanderKeith): How should we handle this case?
+        
+        Parameters
+        ----------
+        params : PhysicsMethodParams
+            Parameters containing MDS connection and shot information.
+
+        Returns
+        -------
+        xarray.Dataset
+            A dataset containing the GEQDSK parameters.
+
+        """
+        # Get EFIT time array
+        efit_time = params.mds_conn.get_data(
+            r"\efit_a_eqdsk:atime", tree_name="_efit_tree"
+        ) / 1.0e3  # [ms] -> [s]
+
+        # Grid parameters
+        # For DIII-D, rcentr and rleft are missing but can be inferred from rgrid
+        grid_params = {
+            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),        # Number of horizontal grid points
+            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),        # Number of vertical grid points
+            "rdim": params.mds_conn.get_data(r"\efit_g_eqdsk:xdim", tree_name="_efit_tree"),    # Horizontal dimension of grid [m]
+            "zdim": params.mds_conn.get_data(r"\efit_g_eqdsk:zdim", tree_name="_efit_tree"),    # Vertical dimension of grid [m]
+            "zmid": params.mds_conn.get_data(r"\efit_g_eqdsk:zmid", tree_name="_efit_tree"),    # Z at center of grid [m]
+        }
+        rgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:r", tree_name="_efit_tree"),  # R grid [m]
+        zgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:z", tree_name="_efit_tree"),  # Z grid [m]
+
+        # Scalar parameters
+        scalar_params = {
+            # Plasma parameters
[…]
