### 2025-12-02 f0e0b5d
Added uniform 1kHz time setting
