+++ disruption_py/settings/time_setting.py
@@ -700,6 +700,34 @@ class SharedTimeSetting(TimeSetting):
         tmax = np.min([np.max(t) for t in others])
         return times[np.where((times >= tmin) & (times <= tmax))]
 
+class Uniform1kHzTimeSetting(TimeSetting):
+    """
+    Time setting for creating a uniform timebase at 1 kHz, based on the maximum EFIT time.
+    """
+    def _get_times(self, params: TimeSettingParams) -> np.ndarray:
+        """
+        Parameters
+        ----------
+        params : TimeSettingParams
+            Parameters needed to retrieve the timebase.
+
+        Returns
+        -------
+        np.ndarray
+            Array of times in the timebase.
+        """
+        (efit_time,) = params.mds_conn.get_dims(
+            r"\efit_aeqdsk:ali", tree_name="_efit_tree"
+        )
+
+        max_time = np.max(efit_time)
+        if params.tokamak == Tokamak.CMOD:
+            times= np.round(np.arange(0, max_time + 1e-3, 1e-3), 3)
+            efit_time_unit = "s"
+        if params.tokamak == Tokamak.D3D:
+            times = np.round(np.arange(0, max_time + 1, 1), 0)
+            efit_time_unit = "ms"
+        return _postprocess(times=times, units=efit_time_unit)
 
 # --8<-- [start:time_setting_dict]
 _time_setting_mappings: Dict[str, TimeSetting] = {
@@ -712,6 +740,7 @@ _time_setting_mappings: Dict[str, TimeSetting] = {
     },
     "ip": IpTimeSetting(),
     "ip_efit": SharedTimeSetting([IpTimeSetting(), EfitTimeSetting()]),
+    "uniform_1kHz": Uniform1kHzTimeSetting(),
 }
 # --8<-- [end:time_setting_dict]
 
