### 2025-12-02 42927ef
Added C-Mod efit method for eqdsk data
