+++ disruption_py/machine/cmod/efit.py
@@ -5,6 +5,7 @@ Module for retrieving and processing EFIT parameters for CMOD.
 """
 
 import numpy as np
+import xarray as xr
 
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
@@ -98,6 +99,251 @@ class CmodEfitMethods:
                 efit_data[param] = interp1(efit_time, efit_data[param], params.times)
 
         return efit_data
+
+    @staticmethod
+    @physics_method(
+        tokamak=Tokamak.CMOD,
+    )
+    def get_geqdsk_parameters(params: PhysicsMethodParams):
+        """
+        Retrieve parameters required to create a GEQDSKFile https://freeqdsk.readthedocs.io/en/stable/geqdsk.html
+        Also see https://fusion.gat.com/theory/Efitgeqdsk
+        This requires the following:
+        
+        1. shot number
+        2. grid parameters: (nx, ny, rdim, zdim, rcenter, rleft, zmid) or (rgrid, zgrid)
+        3. scalar parameters: rmagx, zmagx, simagx, sibdry, bccenter, cpasma, etc.
+        4. 1D profile parameters: fpol, pres, ffprime, pprime, qpsi
+        5. 2D parameters: psi or f or psirz
+        6. boundary and limiter data (if available): nbdry, nlim, rbdry, zbdry, rlim, zlim
+
+        Sometimes the C-Mod EFIT uses 'x' and 'y' instead of 'r' and 'z' for parameter names
+        This function will convert them to the standard GEQDSK naming convention.
+
+        Also, it doesn't make sense to get interpolated boundary/limiter data because the number of points can change over time.
+        TODO(ZanderKeith): How should we handle this case?
+
+        Parameters
+        ----------
+        params : PhysicsMethodParams
+            Parameters containing MDS connection and shot information.
+        Returns
+        -------
+        xarray.Dataset
+            A dataset containing the GEQDSK parameters.
+        """
+        # Get EFIT time array
+        efit_time = params.mds_conn.get_data(r"\efit_a_eqdsk:atime", tree_name="_efit_tree")
+
+        # Grid parameters (For C-Mod a lot are missing, but can be inferred from rgrid/zgrid)
+        grid_params = {
+            "nw": params.mds_conn.get_data(r"\efit_g_eqdsk:mw", tree_name="_efit_tree"),        # Number of horizontal grid points
+            "nh": params.mds_conn.get_data(r"\efit_g_eqdsk:mh", tree_name="_efit_tree"),        # Number of vertical grid points
+            "rdim": params.mds_conn.get_data(r"\efit_g_eqdsk:xdim", tree_name="_efit_tree"),    # Horizontal dimension of grid [m]
+            "zdim": params.mds_conn.get_data(r"\efit_g_eqdsk:zdim", tree_name="_efit_tree"),    # Vertical dimension of grid [m]
+        }
+        rgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:rgrid", tree_name="_efit_tree")  # R grid [m]
+        zgrid = params.mds_conn.get_data(r"\efit_g_eqdsk:zgrid", tree_name="_efit_tree")  # Z grid [m]
+
+        # Scalar parameters
+        scalar_params = {
+            # Plasma parameters
+            "cpasma": params.mds_conn.get_data(r"\efit_g_eqdsk:cpasma", tree_name="_efit_tree"),  # Plasma current [A]
+            "bcentr": params.mds_conn.get_data(r"\efit_g_eqdsk:bcentr", tree_name="_efit_tree"),  # Vacuum toroidal field [T]
+
[…]
