### 2025-12-02 b7b8c0b
Fixed problem with trailing commas and bad indexing into rgrid and zgrid
