### 2025-12-02 937d7bf
suppress stderr when looking up commit (#492)
