### 2025-12-02 3f4ddd5
use localscratch rather than tmp if available
