### 2025-12-02 42b3f1f
fixup! use localscratch rather than tmp if available
