### 2025-12-04 1cf1271
Remove print statement
