### 2025-12-04 0787b00
Update with more gracefully missing data handling
