+++ disruption_py/machine/mast/config.toml
@@ -13,11 +13,33 @@ ip_prog = 100e3
 [mast.tests]
 expected_failure_columns = []
 test_columns = [
-  "ip",
+  'beta_tor_normal', 
+  'dip_dt', 
+  'dipprog_dt', 
+  'dn_dt', 
+  'elongation', 
+  'elongation_axis', 
+  'greenwald_fraction', 
+  'inboard_total', 
+  'ip', 
+  'ip_prog', 
+  'magnetic_axis_r', 
+  'magnetic_axis_z', 
+  'minor_radius', 
+  'n_e', 
+  'n_e_core', 
+  'outboard_total', 
+  'power_nbi', 
+  'power_radiated', 
+  't_e_core', 
+  'time_domain', 
+  'total_injected', 
+  'triangularity_lower', 
+  'triangularity_upper'
 ]
 
 [mast.tests.shots]
-example = 30421
+example = [30421, 30113, 30014]
 
 [mast.time]
 disruption_time_start = 0.2
