### 2025-12-04 71884f2
Update doc strings
