### 2025-12-04 ada2e99
Small optimisation for opening datatree
