### 2025-12-05 cb94d44
Clean up efit calls
