+++ disruption_py/machine/east/physics.py
@@ -7,7 +7,6 @@ Module for retrieving and calculating data for EAST physics methods.
 import numpy as np
 import scipy
 
-from disruption_py.inout.mds import mdsExceptions
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
 from disruption_py.core.utils.math import interp1, matlab_smooth
