+++ disruption_py/machine/mast/__init__.py
@@ -4,7 +4,7 @@
 Module initialization for MAST.
 """
 
-from disruption_py.machine.mast.physics import MastPhysicsMethods
 from disruption_py.machine.mast.efit import MastEfitMethods
+from disruption_py.machine.mast.physics import MastPhysicsMethods
 
 METHOD_HOLDERS = [MastPhysicsMethods, MastEfitMethods]

[…]
