### 2025-12-05 #9364 @gtrevisan
Great! Thank you for taking the time.

I have listed as main authors the current MIT PSFC people providing major contributions to the project in terms of effort.

Previous contributions to the Matlab scripts, currently frozen in the `matlab` branch, are not part of the current package.
Our non-MIT collaborators who worked on the framework in the past -- very distant past in terms of code -- are credited in the readme and in the commit-centric (not effort-centric) [contribution graph](https://github.com/MIT-PSFC/disruption-py/graphs/contributors).
They are also credited, just like most of our other MIT PSFC contributors (either past or present, eg authors of a physics method or two) in the standard-practice team credit -- "& MIT PSFC Disruptions Group".

Please let me know whether this makes sense to you, and as always I'm open to suggestions.
