+++ REFERENCES.md
@@ -27,6 +27,7 @@ Here follows a non-exhaustive list of projects, publications, and conference con
   - H Wietfeldt, _et al._ (2025), _"Characterization of UFOs on Alcator C-Mod and WEST to inform SPARC operation"_ [NP13.157](https://schedule.aps.org/dpp/2025/events/NP13/157)
   - EdD Zapata-Cornejo, _et al._ (2025), _"Time series classification algorithms for confinement regime identification in C-Mod"_ [NP13.169](https://schedule.aps.org/dpp/2025/events/NP13/169)
   - GL Trevisan, _et al._ (2025), _"A large-scale automated EFIT recomputation workflow for disruption studies at 1 kHz"_, [PP13.88](https://schedule.aps.org/dpp/2025/events/PP13/88)
+  - Y Wei, _et al._ (2025), _"Scikit-disruption: machine learning toolkit for disruption analysis"_, [PP13.93](https://schedule.aps.org/dpp/2025/events/PP13/93)
 
 - [6th International Conference on Data-Driven Plasma Science (2025)](https://web.cvent.com/event/7de9d238-e170-4fbf-8de5-20abc5c6eb49/summary)
   - Z Keith, _et al._ (2025), _"A tearing mode database for Alcator C-Mod and DIII-D"_
