### 2025-12-09 0f07b75
add Amos to explicit author list
