+++ paper.md
@@ -14,6 +14,9 @@ authors:
   - name: Yumou Wei
     orcid: 0000-0002-8868-0017
     affiliation: 1
+  - name: Amos M. Decker
+    orcid: 0009-0004-3728-3259
+    affiliation: 1
   - name: Cristina Rea
     orcid: 0000-0002-9948-2649
     affiliation: 1
