### 2025-12-09 9a3a200
add Josh to explicit author list
