### 2025-12-09 0cdef23
add Bob to explicit author list
