### 2025-12-09 6f57312
Use local scratch rather than tmp for DIII-D (#494)
* use localscratch rather than tmp if available

* fixup! use localscratch rather than tmp if available
