### 2025-12-09 ef50843
add generic methods to docs, tweak menu, rename and tweak others
