+++ mkdocs.yml
@@ -42,18 +42,20 @@ nav:
               - Shotlist Setting: usage/settings/shotlist_setting.md
               - Time Setting: usage/settings/time_setting.md
       - Calculated Parameters:
+          - Overview: usage/physics_methods/physics_method_reference.md
           - Physics Methods:
-              - Overview: usage/physics_methods/physics_method_reference.md
+              - Generic Physics Methods: "usage/physics_methods/\
+                generic_method_reference.md"
               - C-Mod Built-in Methods: "usage/physics_methods/\
-                cmod_built_in_method_reference.md"
-              - D3D Built-in Methods: "usage/physics_methods/\
-                d3d_built_in_method_reference.md"
+                cmod_method_reference.md"
+              - DIII-D Built-in Methods: "usage/physics_methods/\
+                d3d_method_reference.md"
               - EAST Built-in Methods: "usage/physics_methods/\
-                east_built_in_method_reference.md"
+                east_method_reference.md"
               - HBT-EP Built-in Methods: "usage/physics_methods/\
-                hbtep_built_in_method_reference.md"
-              - Physics Method Decorators: "usage/physics_methods/\
-                decorator_reference.md"
+                hbtep_method_reference.md"
+          - Physics Method Decorators: "usage/physics_methods/\
+            decorator_reference.md"
           - Parameter Reference: "usage/physics_methods/\
             disruption_parameters_reference.md"
       - Data I/O:
