### 2025-12-10 #9364 @jatkinson1000
Hi @LiamPattinson, @jackleland, and @rmrsk

Please could you let us know how you are progressing with your reviews?
I see checklists are starting to be filled for all of you, but can't see any linked issues that have been opened.

Please could you aim to finish working through these over the next week and open any issues for the author to address so that this can continue to progress in a timely manner?

@gtrevisan please reflect on the [JOSS authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) and whether any of the individuals discussed above have made _"active [contribution to the] project direction"_ and confirm when you have reached a decision over authors. It remains your decision, but from my perspective both Herbert and Joshua have made significant contributions to this codebase, including the Python implementation.

I'll check back next Friday, but in the meantime please tag me if you need anything, and place any questions or comments for broader discussion in this thread.
