### 2025-12-10 MIT-PSFC/disruption-py#496 @LiamPattinson
Missing dependency when running the code directly from install
Hi, I'm currently reviewing `disruption-py` for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9364). Sorry for the long delay in getting started!

I ran into some trouble running the code using `uvx` or from a fresh install due to a missing dependency:

```console
$ uvx -p 3.12 -- disruption-py --help
Traceback (most recent call last):
  File "/home/<username>/.cache/uv/archive-v0/elx5uXJkIG79UbIHzMXN6/lib/python3.12/site-packages/disruption_py/inout/mds.py", line 25, in <module>
    from mdsthin import MDSplus
ModuleNotFoundError: No module named 'mdsthin'

$ uv venv -p 3.12
$ source .venv/bin/activate
$ uv pip install disruption-py
$ disruption-py --help
Traceback (most recent call last):
  File "/home/<username>/.cache/uv/archive-v0/elx5uXJkIG79UbIHzMXN6/lib/python3.12/site-packages/disruption_py/inout/mds.py", line 25, in <module>
    from mdsthin import MDSplus
ModuleNotFoundError: No module named 'mdsthin'
```

I also get the same issue when doing `[uv] pip install .` from the repo. `mdsthin` is included in the `dev` dependency group, but it looks like it should be a regular dependency to the project.
