### 2025-12-10 #9364 @jatkinson1000
Thanks @LiamPattinson for that feedback.

Yes, the code/version reviewed and published should be identical.

@gtrevisan please confirm if `slj/mast` is to be merged soon and any timescales.
It would not be fair to ask reviewers to inspect code/functionality of branches, and then ask them to check again when it is merged. JOSS expects that code comes to review as ready for publication as the author can make it. If there is still development work to be done, including merging of feature branches, then we would need to look at pausing the review whilst this takes place.
Ideally we'd both want to avoid this as there is no guarantee that reviewers will be available in the future.

Regarding your other points @LiamPattinson - we do not expect you as a reviewer to dedicate significant time to fix failing tests, broken APIs, or unclear install instructions (unless you wish to). The review is intended to check the useability and functionality of the software. Please open these as issues on the repository for the authors to fix, linking back here as described above.
