+++ disruption_py/machine/mast/physics.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=duplicate-code
 
 """
 Physics methods for MAST.
