### 2025-12-10 MIT-PSFC/disruption-py#497 @LiamPattinson
Installation instructions unclear
Part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/9364). Working on the `slj/mast` branch as I lack the access permissions for other tokamaks.

On installing and running the code, I'm getting warnings about a missing SQL server/database:

```console
❯ poetry run python examples/sql.py
2025-12-10 13:49:28.019 | WARNING  | disruption_py.inout.sql:from_config:97 - No SQL server/database!
```

It wasn't clear from the README that this was something I would have to set up, and I wasn't able to resolve it after reading `INSTALL.md` either. I installed FreeTDS, but I don't know what else I need to do resolve the warning. This is also preventing me from successfully running the tests or examples.

It would be helpful if these requirements could be better sign-posted in the README, and if a minimal installation walkthrough could be added to `INSTALL.md`.
