### 2025-12-10 MIT-PSFC/disruption-py#498 @LiamPattinson
More examples?
Part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/9364). Working on the `slj/mast` branch as I lack the access permissions for other tokamaks.

I was able to extract some MAST data by following the README and the `--help` text, but the next step of integrating this into an analysis pipeline is left open in the documentation. I realise that this could be done in any number of ways, but to give new users a jumping-off point and to demonstrate the capabilities of the code, I think it would be helpful to include a short worked example that pulls some data from a public repository and then performs some simple analysis.
