+++ disruption_py/inout/xarray.py
@@ -86,3 +86,6 @@ class XarrayConnection:
             )
 
         return np.array([np.nan])
+
+    def cleanup(self):
+        pass
