+++ disruption_py/inout/mds.py
@@ -19,20 +19,16 @@ from disruption_py.machine.tokamak import Tokamak
 try:
     # first, try full-fledged MDSplus
     import MDSplus
-
 except ModuleNotFoundError:
     try:
         # then, fall back onto mdsthin
         from mdsthin import MDSplus
-
     except ModuleNotFoundError:
         # finally, create dummy class
         class MDSplus:
             """Dummy MDSplus class."""
-
             class Connection:
                 """Dummy Connection class."""
-
             # pylint: disable-next=invalid-name
             class mdsExceptions:
                 """Dummy mdsExceptions class."""
