### 2025-12-10 dfc772e
Add fix linting warnings
