### 2025-12-10 9469aee
prefer MDSplus, warn about mdsthin, disregard if no connection
