+++ disruption_py/workflow.py
@@ -104,10 +104,17 @@ def get_shots_data(
     log_settings = resolve_log_settings(log_settings)
     log_settings.setup_logging()
 
-    if "mdsthin" in sys.modules:
-        logger.warning("Cannot import MDSplus, falling back onto mdsthin!")
-
     tokamak = resolve_tokamak_from_environment(tokamak)
+
+    if "MDSplus" in sys.modules:
+        logger.info("Imported MDSplus.")
+    elif "mdsthin" in sys.modules:
+        logger.warning("Imported mdsthin!")
+    elif not config(tokamak).inout.mds.mdsplus_connection_string:
+        logger.debug("Did not import MDSplus.")
+    else:
+        raise ModuleNotFoundError("Cannot import MDSplus.")
+
     database = _get_database_instance(tokamak, database_initializer)
     # Clean-up parameters
     if retrieval_settings is None:
