### 2025-12-10 MIT-PSFC/disruption-py#498 @AlexSaperstein
I'll try to look into the WM summer school notebooks this week
