### 2025-12-11 dc5532c
Add missing method, missing exception flag
