### 2025-12-11 a6479be
Add generic methods to docs (#495)
* add generic methods to docs, tweak menu, rename and tweak others

* Specify "C-Mod" in the Disruption Parameter Descriptions table header

---------

Co-authored-by: yumouwei <46117079+yumouwei@users.noreply.github.com>
