+++ docs/usage/physics_methods/cmod_method_reference.md
@@ -1,6 +1,5 @@
-
-## Built-in Methods For D3D { .doc .doc-heading }
-::: disruption_py.machine.d3d.physics
+## C-Mod Physics Methods { .doc .doc-heading }
+::: disruption_py.machine.cmod.physics
     handler: python
 	options:
 	  heading_level: 2
