### 2025-12-12 f90b75b
add a couple of generic MDSplus exceptions
