+++ GETTING_STARTED.md
@@ -1,0 +1,289 @@
+# Getting Started with MAST Open Data & DisruptionPy
+
+Welcome to DisruptionPy! This guide will help you install the software and run your first data retrieval workflows.
+
+## What is DisruptionPy?
+
+DisruptionPy is an open-source python package for efficient retrieval of experimental fusion data. It enables fast database preparation for disruption analysis and machine learning applications across multiple tokamak devices including C-Mod, DIII-D, EAST, HBT-EP, and MAST.
+
+In this guide, we will focus on using DisruptionPy with MAST Open Data, which does not require any special permissions to access.
+
+---
+
+## Installation
+
+DisruptionPy is available on [PyPI](https://pypi.org/project/disruption-py/) and can be installed using modern Python package managers.
+
+### Option 1: Installation with uv
+
+[uv](https://docs.astral.sh/uv/) is a fast Python package manager. To install DisruptionPy:
+
+```bash
+uv tool install disruption-py
+```
+
+### Option 2: Installation with Poetry
+
+[Poetry](https://python-poetry.org/) provides robust dependency management:
+
+```bash
+# Add disruption-py
+poetry add disruption-py
+
+# Activate the virtual environment
+poetry run disruption-py
+```
+
+### Option 3: Installation with pip
+
+For a simple pip installation:
+
+```bash
+# Install disruption-py
+pipx install disruption-py
+```
+
+---
+
+## Command Line Interface Examples
+
+DisruptionPy provides a command-line interface for quick data retrieval without writing Python scripts.
+
+### CLI Quick Start
+
+Retrieve data for a MAST shot using the EFIT timebase:
+
+```bash
+disruption-py --tokamak mast --time-base efit --output my_data.csv 30421
+```
+
+**Options explained:**
+
+- `--tokamak mast`: Specifies the MAST tokamak as the data source
+- `--time-base efit`: Uses the EFIT reconstruction timebase for data interpolation
+- `--output my_data.csv`: Saves retrieved data to a CSV file named `my_data.csv`
+- `30421`: The shot number to retrieve data from
+
+
+### CLI Example 1: EFIT Timebase with Specific Methods
+
+Retrieve EFIT parameters on the EFIT timebase:
+
+```bash
+disruption-py \
+  --tokamak mast \
+  --time-base efit \
+  --methods get_efit_parameters \
+  --output mast_efit_data.csv \
+  30421
+```
+
+### CLI Example 2: Disruption Warning Timebase
+
+Use the disruption warning timebase:
+
+```bash
+disruption-py \
+  --tokamak mast \
+  --time-base disruption_warning \
+  --output mast_disruption_warning.csv \
+  30421
+```
+
+### CLI Example 3: Plasma Current (Ip) Timebase
+
+Use the plasma current timebase:
+
+```bash
+disruption-py \
+  --tokamak mast \
+  --time-base ip \
+  --output mast_ip_timebase.csv \
+  30421
+```
+
+### CLI Example 4: Multiple Shots with Parallel Processing
+
+Process multiple MAST shots in parallel:
+
+```bash
+disruption-py \
+  --tokamak mast \
+  --time-base efit \
+  -m get_efit_parameters -m get_ip -m get_density \
+  --processes 2 \
+  --output mast_multiple_shots.csv \
+  30420 30421 30422
+```
+
+ - The `-m`/`--methods` parameter specifies which methods to call on the backend object to retrieve data. Multiple `-m` flags can be used to extract different types of data (e.g., EFIT parameters, plasma current, and density) in a single run.
+
[…]
