### 2025-12-15 eac41dd
Update APS-DPP 2025 IDs (#489)
* update id for Trevisan APS-DPP 2025

* Update wietfeldt APS 2025 poster number

* Add Wei APS-DPP 2025 poster

* Update reference link APS Zapata Cornejo

Link corrected. I accept also your suggestion of removing the hyphen from my second name, if that still OK.

* Updated zkeith APS 2025 poster number

* fixed poster ID for AR Saperstein

* fixed text

---------

Co-authored-by: hwietfeldt <henry.wietfeldt@gmail.com>
Co-authored-by: yumouwei <46117079+yumouwei@users.noreply.github.com>
Co-authored-by: Enrique Zapata <zapata_e@psfc.mit.edu>
Co-authored-by: Zand
[…]
