### 2025-12-17 a560719
First pass at grabbing yag_fit profiles
