### 2025-12-17 694ea61
Made it so shots without profiles don't get put in the dataset (preventing OOM error)
