### 2025-12-18 2adc1d3
Revamp d3d.get_n1_bradial_parameters method
