### 2025-12-18 2fd727c
Added penalties from YAG_FIT
