### 2025-12-18 c25e4fc
Test method on d3d server, then made improvements
