### 2025-12-18 7acc920
Added method to get raw Thomson points
