### 2025-12-18 ba7b1e7
Move get_btor to its own physics method
