### 2025-12-18 562e27e
Fix n_equal_1_mode unit
