### 2025-12-18 fbd4b24
Add docstrings
