+++ disruption_py/machine/cmod/physics.py
@@ -2239,31 +2239,33 @@ class CmodPhysicsMethods:
     def get_thomson_points(params: PhysicsMethodParams):
         """
         Retrieve raw Thomson scattering measurement points for electron temperature and density.
+        Combines both core and edge Thomson data into a single dataset.
 
         A few notes:
-        - Thomson YAG time by default runs from -0.5s to 2.0s (https://cmodwiki.psfc.mit.edu/index.php/Thomson_scattering_programming)  
+        - Thomson YAG time by default runs from -0.5s to 2.0s (https://cmodwiki.psfc.mit.edu/index.php/Thomson_scattering_programming)
         - See https://cmodwiki.psfc.mit.edu/index.php/Thomson_Scattering
+        - See https://cmodwiki.psfc.mit.edu/index.php/Edge_Thomson_Scattering
         """
 
-        data_vars = {}
-
+        # ========== Get Core Thomson Data ==========
         # TODO(ZanderKeith) verify this rho lines up with the R,Z and EFIT rho
-        nodename_rho = f".yag_new.results.profiles.rho_t"
-        rho_data, rho_time = params.mds_conn.get_data_with_dims(nodename_rho, tree_name="electrons")
+        nodename_rho_core = f".yag_new.results.profiles.rho_t"
+        rho_core_data, rho_core_time = params.mds_conn.get_data_with_dims(nodename_rho_core, tree_name="electrons")
         # Cut time to positive values only
-        valid_time_indices = rho_time >= 0
-        rho_data = rho_data[:, valid_time_indices]
-        rho_time = rho_time[valid_time_indices]
+        valid_time_indices = rho_core_time >= 0
+        rho_core_data = rho_core_data[:, valid_time_indices]
+        rho_core_time = rho_core_time[valid_time_indices]
         # Replace -1 with nan
-        rho_data[rho_data == -1] = np.nan
+        rho_core_data[rho_core_data == -1] = np.nan
 
-        rho_data_interp = interp1(
-            rho_time,
-            rho_data,
+        rho_core_data_interp = interp1(
+            rho_core_time,
+            rho_core_data,
             params.times,
             kind="previous",
         )
 
+        core_data = {}
         for quant in ["ne", "te"]:
             nodename_data = f".yag_new.results.profiles.{quant}_rz"
             nodename_error = f".yag_new.results.profiles.{quant}_err"
@@ -2284,9 +2286,9 @@ class CmodPhysicsMethods:
             # Replace 0 with nan
             profile_data[profile_data == 0] = np.nan
             error_data[error_data == 0] = np.nan
-            
+
             # Raise error if profile and rho time do not match
-            if not np.array_equal(profile_time, rho_time):
+            if not np.array_equal(profile_time, rho_core_time):
                 raise ValueError(
                     f"Thomson scattering {quant} profile time does not match rho time."
                 )
@@ -2305,23 +2307,130 @@ class CmodPhysicsMethods:
                 kind="previous",
             )
 
-            # Store profile data and rho coord
-            data_vars[f"{quant}_profile"] = xr.DataArray(
-                profile_data_interp.astype(np.float32).T,
+            core_data[quant] = profile_data_interp.astype(np.float32).T
+            core_data[f"{quant}_error"] = error_data_interp.astype(np.float32).T
+
+        n_core_channels = rho_core_data.shape[0]
+
+        # ========== Get Edge Thomson Data ==========
+        nodename_rho_edge = f".yag_edgets.results.psinorm"
+        rho_edge_data, rho_edge_time = params.mds_conn.get_data_with_dims(nodename_rho_edge, tree_name="electrons")
+        # Cut time to positive values only
[…]
