### 2025-12-19 cb6963d
First pass at adding GP fits that use TS data
