+++ disruption_py/machine/cmod/thomson.py
@@ -5,6 +5,11 @@
 import numpy as np
 import scipy as sp
 
+import jax
+jax.config.update("jax_enable_x64", True) # The data is in f32, but doing computations in f64 improves stability
+import jax.numpy as jnp
+import gpjax as gpx
+
 from disruption_py.core.physics_method.params import PhysicsMethodParams
 from disruption_py.core.utils.math import interp1
 from disruption_py.inout.mds import mdsExceptions
@@ -314,3 +319,208 @@ class CmodThomsonDensityMeasure:
             psi[:, i] = sp.interpolate.griddata(points, values, (r, z), method="cubic")
 
         return psi
+
+
+class CmodThomsonGPProfiles:
+    """
+    Helper class for using Gaussian Processes to fit Thomson scattering profiles.
+    """
+
+    @jax.jit
+    def gp_profile_single_timestep(
+        signal_data,
+        signal_error,
+        signal_rho,
+        result_rho,
+        edge_rho=1.2,
+        edge_value=1e-3,
+        edge_error=0.5,
+        epsilon=1e-3,
+        ):
+        """
+        Fit a single timestep using Gaussian Processes.
+        Assumes there are no missing / nan values in the input data.
+
+        Since this is a 1D profile of a physical plasma quantity, we can make the following assumptions to constrain the fit:
+        1. The signal is always positive.
+        2. The signal has zero gradient at the core (rho=0).
+        3. The signal goes to some relatively known value at the edge (rho=edge_rho).
+        - For ne, Te, J, etc., this value should be 0
+        - For q it could be taken as something like 10
+
+        Implementing these assumptions is done as follows:
+        1. Use a log-transform on the signal values.
+        2. Using JAX, we compute covariences in both the kernel function and its derivative, and add a low-uncertainty data point at (0, signal(0)).
+        3. A synthetic data point is added at (edge_rho, edge_value) with uncertainty edge_error.
+
+        The final piece which makes the profiles 'look good' is fitting in rho^2 space.
+        This essentially spreads out the edge region, allowing the GP to vary more quickly there while remaining smooth in the core.
+
+        Parameters
+        ----------
+        signal_data : jnp.ndarray
+            The signal data to fit.
+        signal_error : jnp.ndarray
+            The standard deviations associated with the signal data.
+        signal_rho : jnp.ndarray
+            The normalized radius (rho) values corresponding to the signal data.
+        result_rho : jnp.ndarray
+            The rho values where the fitted profile should be evaluated.
+        edge_rho : float, optional
+            The rho value at the edge where the signal is expected to approach edge_value (default is 1.2).
+        edge_value : float, optional
+            The expected signal value at edge_rho (default is 1e-3).
+        edge_error : float, optional
+            The uncertainty associated with the edge_value (default is 0.5).
+        epsilon : float, optional
+            A small value added to avoid log(0) issues (default is 1e-3).
+
+        Returns
+        -------
+        jnp.ndarray, jnp.ndarray
+            The fitted signal profile and its standard deviation evaluated at result_rho.
+        """
+        kernel = gpx.kernels.Matern32()  # Kernel that allows some roughness but is still smooth
+
+        # Ensure inputs are jnp arrays with a flat shape. They will be reshaped as needed for given matrix operations.
[…]
