### 2025-12-19 f608a19
Refactored c-mod physics methods to cache Thomson points so they aren't pulled twice and implemented vmapping across timeslices for the gp fitting routine
