+++ disruption_py/machine/cmod/thomson.py
@@ -328,14 +328,18 @@ class CmodThomsonGPProfiles:
 
     @jax.jit
     def gp_profile_single_timestep(
-        signal_data,
-        signal_error,
-        signal_rho,
-        result_rho,
-        edge_rho=1.2,
-        edge_value=1e-3,
-        edge_error=0.5,
-        epsilon=1e-3,
+        signal_data: jnp.ndarray,
+        signal_error: jnp.ndarray,
+        signal_rho: jnp.ndarray,
+        result_rho: jnp.ndarray,
+        kernel: gpx.kernels.AbstractKernel,
+        edge_rho: float,
+        edge_value: float,
+        edge_error: float,
+        epsilon: float,
+        outlier_penalty: float,
+        outlier_cutoff: float,
+        jitter: float,
         ):
         """
         Fit a single timestep using Gaussian Processes.
@@ -366,21 +370,28 @@ class CmodThomsonGPProfiles:
             The normalized radius (rho) values corresponding to the signal data.
         result_rho : jnp.ndarray
             The rho values where the fitted profile should be evaluated.
-        edge_rho : float, optional
-            The rho value at the edge where the signal is expected to approach edge_value (default is 1.2).
-        edge_value : float, optional
-            The expected signal value at edge_rho (default is 1e-3).
-        edge_error : float, optional
-            The uncertainty associated with the edge_value (default is 0.5).
-        epsilon : float, optional
-            A small value added to avoid log(0) issues (default is 1e-3).
+        kernel : gpx.kernels.AbstractKernel
+            The kernel to use for the Gaussian Process. Must be initialized outside this jitted function.
+        edge_rho : float
+            The rho value at the edge where the signal is expected to approach edge_value
+        edge_value : float
+            The expected signal value at edge_rho.
+        edge_error : float
+            The uncertainty associated with the edge_value.
+        epsilon : float
+            A small value added to avoid log(0) issues.
+        outlier_penalty : float
+            The error value to assign to identified outliers to effectively ignore them.
+        outlier_cutoff : float
+            The rho value below which outliers will be identified.
+        jitter : float
+            A small jitter value added to the diagonal of the covariance matrix for numerical stability.
 
         Returns
         -------
         jnp.ndarray, jnp.ndarray
             The fitted signal profile and its standard deviation evaluated at result_rho.
         """
-        kernel = gpx.kernels.Matern32()  # Kernel that allows some roughness but is still smooth
 
         # Ensure inputs are jnp arrays with a flat shape. They will be reshaped as needed for given matrix operations.
         signal_data = jnp.array(signal_data).flatten()
@@ -405,7 +416,7 @@ class CmodThomsonGPProfiles:
         # Perform a quick GP fit without boundary conditions to identify outliers
         K_rough = kernel.gram(rho2.reshape(-1,1)).to_dense()  # K(X, X), prior covariance at observed points
         noise_rough = jnp.diag(errors_log ** 2)  # Observation noise, GPJax uses variance
-        K_total_rough = K_rough + noise_rough + jnp.eye(len(rho2)) * 1e-6  # Add small jitter for numerical stability
+        K_total_rough = K_rough + noise_rough + jnp.eye(len(rho2)) * jitter  # Add small jitter for numerical stability
         # Make predictions at observed points
         alpha = jnp.linalg.solve(K_total_rough, observations_log)
         predictions_rough = K_rough @ alpha
@@ -414,12 +425,12 @@ class CmodThomsonGPProfiles:
         predictions_std = jnp.sqrt(jnp.maximum(predictions_var, 0))
         # Identify outliers as core points where residual > 3 * predicted stddev
[…]
