+++ disruption_py/machine/cmod/physics.py
@@ -21,7 +21,7 @@ from disruption_py.core.utils.math import (
     interp1,
 )
 from disruption_py.inout.mds import mdsExceptions
-from disruption_py.machine.cmod.thomson import CmodThomsonDensityMeasure
+from disruption_py.machine.cmod.thomson import CmodThomsonDensityMeasure, CmodThomsonGPProfiles
 from disruption_py.machine.tokamak import Tokamak
 
 
@@ -2233,204 +2233,243 @@ class CmodPhysicsMethods:
         return profile_ds
     
     @staticmethod
-    @physics_method(
-        tokamak=Tokamak.CMOD,
-    )
-    def get_thomson_points(params: PhysicsMethodParams):
+    @cache_method
+    def _get_thomson_channels_raw(params: PhysicsMethodParams):
         """
-        Retrieve raw Thomson scattering measurement points for electron temperature and density.
+        Retrieve raw Thomson scattering measurement channels for electron temperature and density.
         Combines both core and edge Thomson data into a single dataset.
 
         A few notes:
         - Thomson YAG time by default runs from -0.5s to 2.0s (https://cmodwiki.psfc.mit.edu/index.php/Thomson_scattering_programming)
         - See https://cmodwiki.psfc.mit.edu/index.php/Thomson_Scattering
-        - See https://cmodwiki.psfc.mit.edu/index.php/Edge_Thomson_Scattering
         """
 
-        # ========== Get Core Thomson Data ==========
-        # TODO(ZanderKeith) verify this rho lines up with the R,Z and EFIT rho
-        nodename_rho_core = f".yag_new.results.profiles.rho_t"
-        rho_core_data, rho_core_time = params.mds_conn.get_data_with_dims(nodename_rho_core, tree_name="electrons")
-        # Cut time to positive values only
-        valid_time_indices = rho_core_time >= 0
-        rho_core_data = rho_core_data[:, valid_time_indices]
-        rho_core_time = rho_core_time[valid_time_indices]
-        # Replace -1 with nan
-        rho_core_data[rho_core_data == -1] = np.nan
+        core_nodes = {
+            "rho": ".yag_new.results.profiles.rho_t",
+            "ne": ".yag_new.results.profiles.ne_rz",
+            "ne_error": ".yag_new.results.profiles.ne_err",
+            "te": ".yag_new.results.profiles.te_rz",
+            "te_error": ".yag_new.results.profiles.te_err",
+        }
+        edge_nodes = {
+            "rho": ".yag_edgets.results.psinorm",
+            "ne": ".yag_edgets.results.ne",
+            "ne_error": ".yag_edgets.results.ne.error",
+            "te": ".yag_edgets.results.te",
+            "te_error": ".yag_edgets.results.te.error",
+        }
 
-        rho_core_data_interp = interp1(
-            rho_core_time,
-            rho_core_data,
-            params.times,
-            kind="previous",
-        )
-
-        core_data = {}
-        for quant in ["ne", "te"]:
-            nodename_data = f".yag_new.results.profiles.{quant}_rz"
-            nodename_error = f".yag_new.results.profiles.{quant}_err"
-
-            profile_data, profile_time = params.mds_conn.get_data_with_dims(
-                nodename_data,
-                tree_name="electrons",
-            )
-            error_data = params.mds_conn.get_data(
-                nodename_error,
-                tree_name="electrons",
-            )
-            # Cut time to positive values only
-            valid_time_indices = profile_time >= 0
-            profile_data = profile_data[:, valid_time_indices]
-            error_data = error_data[:, valid_time_indices]
-            profile_time = profile_time[valid_time_indices]
-            # Replace 0 with nan
-            profile_data[profile_data == 0] = np.nan
[…]
