### 2025-12-19 ab1dc02
Fixed problem with NaNs getting into the GP solver
