### 2025-12-19 2e29543
Fixed problem where profiles were not using rectilinear interpolation
