### 2026-01-05 #9364 @rmrsk
Hi @gtrevisan -- I have some comments to the paper itself that should be adressed.

1. I'd like to see the final author list as soon as it is finalized.
2. The Statement of need section describes DisruptionPy quite well, but does not state its relation to other packages/codes. If other related analysis tools exist, they should be briefly discussed and referenced. If not, a statement on the uniqueness of DisruptionPy would be appropriate.
3. The references only contain works from the Disruptions Group (Cristina Rea is an author of every referenced paper, except Stillerman 1997). Fusion plasma disruptions is not my field, so for all I know this may or may not be completely appropriate. But at the very least I'd like you to comment as to why there are no references to works outside of the Disruptions Group, and/or make amendments to the list of references if relevant.
4. The acknowledgement section only mentions the latest funding source, but it seems to me that the software has a longer and perhaps more complicated history? Are there other additional funding sources or projects that should be credited?
