### 2026-01-05 3c519da
remove layouts from readme
