### 2026-01-05 89e1f1d
revamp installation section in readme
