+++ README.md
@@ -28,7 +28,7 @@
 
 DisruptionPy is an open-source Scientific Python package for fast retrieval of experimental Fusion data from [MDSplus](https://www.mdsplus.org/) servers.
 The library allows an efficient database preparation for downstream analysis and/or ML model development for disruption studies.
-Currently supported machines are:
+Currently supported devices are:
 
 - [Alcator C-Mod](https://en.wikipedia.org/wiki/Alcator_C-Mod)
 - [DIII-D](https://en.wikipedia.org/wiki/DIII-D_(tokamak))
@@ -65,11 +65,83 @@ _Figure: Schematic flowchart of a typical DisruptionPy workflow (by: Y Wei)._
 The most recent revamp of DisruptionPy was partially supported by DOE FES under Award DE-SC0024368, ["Open and FAIR Fusion for Machine Learning Applications"](https://crea-psfc.github.io/open-fair-fusion/).
 
 
+## Devices
+
+DisruptionPy supports data extraction workflows on several fusion devices.
+
+### Access
+
+Each device hosts experimental data separately and may require authorization for access.
+
+DisruptionPy itself does not provide access to any of the underlying servers.
+
+To obtain access, please contact your host at the desidered institution.
+
+### Data sources
+
+Most machines require both MDSplus and SQL database access for full functionality.
+
+| Device        | Institution | Data    | Metadata | Access           |
+|---------------|-------------|---------|----------|------------------|
+| Alcator C-Mod | MIT PSFC    | MDSplus | SQL      | :yellow_circle:  |
+| DIII-D        | GA          | MDSplus | SQL      | :yellow_circle:  |
+| EAST          | ASIPP       | MDSplus | SQL      | :red_circle:     |
+| HBT-EP        | Columbia    | MDSplus | -        | :yellow_circle:  |
+
 ## Installation
 
-DisruptionPy is now open-source and [available at PyPI](https://pypi.org/project/disruption-py/)!
+DisruptionPy is [open-source on Github](https://github.com/MIT-PSFC/disruption-py/) and [available at PyPI](https://pypi.org/project/disruption-py/)!
 
-For standard installations, please follow the usual way:
+### Pre-requisites
+
+Depending on the target device, DisruptionPy may need non-Python software to be installed as a pre-requisite.
+For more details, please refer to the Devices section above and our [Installation guide](INSTALL.md).
+
+### Dependencies
+
+DisruptionPy follows best practices for project metadata (eg: [pyproject.toml](https://peps.python.org/pep-0621/) and [dependency groups](https://peps.python.org/pep-0735/)) and can therefore be installed with all its dependencies through any standard Python package manager.
+
+We recommend using either [Poetry](https://python-poetry.org/) or [uv](https://docs.astral.sh/uv/) for a user-friendly virtual environment experience.
+
+### Branches
+
+We suggest to install DisruptionPy:
+- from the stable branch, `main`, for production workflows;
+- from the development branch, `dev`, for development workflows;
+- from any other branch, only for testing workflows.
+
+### A) Run pre-installed
+
+If executing on a supported server, installation might not be necessary at all!
+
+Several public installations are currently maintained automatically, as outlined in our [Installation guide](INSTALL.md).
+
+### B) Install as an application
+
+Most users will want to use DisruptionPy as a standalone application.
+
+DisruptionPy workflows can easily be executed on the fly, without first creating a dummy project.
+
+```bash
+# if you use uv
+uvx disruption-py
+
+# if you use pipx
+pipx install disruption-py
+```
+
+If you use `uv`, please refer to this convenient snippet to execute DisruptionPy from any given branch: 
[…]
