### 2026-01-05 f77576c
fixup! revamp installation section in readme
