### 2026-01-05 d9eac96
add link to contributors page
