### 2026-01-06 dfa18e7
Add a few more physics props
