+++ disruption_py/settings/shotlist_setting.py
@@ -98,12 +98,14 @@ class FileShotlistSetting(ShotlistSetting):
 
     def _get_shotlist(self, params: ShotlistSettingParams) -> List:
         if not self.shotlist:
-            self.kwargs.setdefault("header", "infer")
-            df = pd.read_csv(self.file_path, **self.kwargs)
+            if self.file_path.endswith(".parquet"):
+                df = pd.read_parquet(self.file_path, **self.kwargs)
+            else:
+                self.kwargs.setdefault("header", "infer")
+                df = pd.read_csv(self.file_path, **self.kwargs)
             arr = df.values[:, self.column_index]
             self.shotlist = arr.astype(int).tolist()
         return self.shotlist
-
 
 class IncludedShotlistSetting(FileShotlistSetting):
     """
@@ -179,6 +181,7 @@ _get_shotlist_setting_mappings: Dict[str, ShotlistSetting] = {
 _file_suffix_to_shotlist_setting: Dict[str, Type[ShotlistSetting]] = {
     ".txt": FileShotlistSetting,
     ".csv": FileShotlistSetting,
+    ".parquet": FileShotlistSetting,
 }
 # --8<-- [end:file_suffix_to_shotlist_setting_dict]
 
