### 2026-01-06 c417081
Handle missing xarray data
