### 2026-01-06 c76645a
separate installation from execution in section B
