### 2026-01-06 967c2f3
downgrade warning to debug for empty db config
