### 2026-01-06 fe565ad
drop safe_cast from utils/misc
