### 2026-01-09 6bbb9d6
Revamp installation readmes (#501)
* remove layouts from readme

* revamp installation section in readme

* tweak install

* fixup! revamp installation section in readme

* add link to contributors page

* separate installation from execution in section B

* minor tweaks
