### 2026-01-12 MIT-PSFC/disruption-py#497 @LiamPattinson
I've looked through the updated README on `dev`, and the changes to the installation instructions are a big improvement. The `slj/mast` branch also now works without installing additional dependencies and no longer emits the misleading warning.

On the `dev` branch, I'd argue that this issue is already resolved. For this `slj/mast` branch, there are only really a few small changes I would like to see:

- Merge the README changes to `slj/mast`.
- Add MAST to the currently supported devices and data sources table.
- Make it clear in the text that MAST data can be accessed without additional dependencies, but additional set up is required for other devices.
