### 2026-01-12 fd90a95
Added handling for case where edge TS is not available
