+++ disruption_py/machine/cmod/physics.py
@@ -2261,105 +2261,129 @@ class CmodPhysicsMethods:
 
         data_vars = {}
         for region, nodes in zip(["core", "edge"], [core_nodes, edge_nodes]):
-            # Retrieve rho data
-            rho_data, rho_time = params.mds_conn.get_data_with_dims(
-                nodes["rho"],
-                tree_name="electrons",
-            )
-            # Cut time to positive values only
-            valid_time_indices = rho_time >= 0
-            rho_data = rho_data[:, valid_time_indices]
-            rho_time = rho_time[valid_time_indices]
-            # Replace -1 with nan
-            rho_data[rho_data == -1] = np.nan
-
-            data_vars[f"rho_{region}"] = xr.DataArray(
-                rho_data.astype(np.float32).T,
-                dims=("time", f"{region}_channel"),
-                coords={
-                    "time": rho_time,
-                    f"{region}_channel": np.arange(rho_data.shape[0]),
-                },
-            )
-
-            # Retrieve profiles for ne and te
-            for quant in ["ne", "te"]:
-                profile_data, profile_time = params.mds_conn.get_data_with_dims(
-                    nodes[quant],
-                    tree_name="electrons",
-                )
-                error_data = params.mds_conn.get_data(
-                    nodes[f"{quant}_error"],
+            try:
+                # Retrieve rho data
+                rho_data, rho_time = params.mds_conn.get_data_with_dims(
+                    nodes["rho"],
                     tree_name="electrons",
                 )
                 # Cut time to positive values only
-                valid_time_indices = profile_time >= 0
-                profile_data = profile_data[:, valid_time_indices]
-                error_data = error_data[:, valid_time_indices]
-                profile_time = profile_time[valid_time_indices]
-                # Replace 0 with nan
-                profile_data[profile_data == 0] = np.nan
-                error_data[error_data == 0] = np.nan
+                valid_time_indices = rho_time >= 0
+                rho_data = rho_data[:, valid_time_indices]
+                rho_time = rho_time[valid_time_indices]
+                # Replace -1 with nan
+                rho_data[rho_data == -1] = np.nan
 
-                if region == "edge" and quant == "te":
-                    # Edge Te profile is given in eV, convert to keV
-                    profile_data = profile_data / 1000.0
-                    error_data = error_data / 1000.0
+                data_vars[f"rho_{region}"] = xr.DataArray(
+                    rho_data.astype(np.float32).T,
+                    dims=("time", f"{region}_channel"),
+                    coords={
+                        "time": rho_time,
+                        f"{region}_channel": np.arange(rho_data.shape[0]),
+                    },
+                )
 
-                # Raise error if profile and rho time do not match
-                if not np.array_equal(profile_time, rho_time):
-                    raise ValueError(
-                        f"{region.capitalize()} Thomson scattering {quant} profile time does not match rho time."
+                # Retrieve profiles for ne and te
+                for quant in ["ne", "te"]:
+                    profile_data, profile_time = params.mds_conn.get_data_with_dims(
+                        nodes[quant],
+                        tree_name="electrons",
                     )
+                    error_data = params.mds_conn.get_data(
+                        nodes[f"{quant}_error"],
+                        tree_name="electrons",
+                    )
+                    # Cut time to positive values only
+                    valid_time_indices = profile_time >= 0
+                    profile_data = profile_data[:, valid_time_indices]
+                    error_data = error_data[:, valid_time_indices]
+                    profile_time = profile_time[valid_time_indices]
+                    # Replace 0 with nan
+                    profile_data[profile_data == 0] = np.nan
+                    error_data[error_data == 0] = np.nan
 
[…]
