### 2026-01-12 ef521ed
Generalized constraint definitions
