+++ disruption_py/machine/cmod/thomson.py
@@ -326,6 +326,7 @@ class CmodThomsonGPProfiles:
     Helper class for using Gaussian Processes to fit Thomson scattering profiles.
     """
 
+    @staticmethod
     @jax.jit
     def gp_profile_single_timestep(
         signal_data: jnp.ndarray,
@@ -333,32 +334,29 @@ class CmodThomsonGPProfiles:
         signal_rho: jnp.ndarray,
         result_rho: jnp.ndarray,
         kernel: gpx.kernels.AbstractKernel,
-        edge_rho: float,
-        edge_value: float,
-        edge_error: float,
-        epsilon: float,
+        constraint_val: jnp.ndarray,
+        constraint_grad: jnp.ndarray,
         outlier_penalty: float,
         outlier_cutoff: float,
+        shift: float,
         jitter: float,
         ):
         """
         Fit a single timestep using Gaussian Processes.
         Assumes there are no missing / nan values in the input data.
 
-        Since this is a 1D profile of a physical plasma quantity, we can make the following assumptions to constrain the fit:
+        Since this is a 1D profile of a physical plasma quantity, we can make a few assumptions to improve the fit quality:
         1. The signal is always positive.
-        2. The signal has zero gradient at the core (rho=0).
-        3. The signal goes to some relatively known value at the edge (rho=edge_rho).
-        - For ne, Te, J, etc., this value should be 0
-        - For q it could be taken as something like 10
+        2. The signal has known gradient points (e.g. d(signal)/d(rho) = 0 at rho=0).
+        3. The signal has known edge values (e.g. signal approaches some small value at edge_rho).
 
         Implementing these assumptions is done as follows:
         1. Use a log-transform on the signal values.
-        2. Using JAX, we compute covariences in both the kernel function and its derivative, and add a low-uncertainty data point at (0, signal(0)).
-        3. A synthetic data point is added at (edge_rho, edge_value) with uncertainty edge_error.
+        2. Synthetic data points are added with boundary conditions on the gradients and values.
 
         The final piece which makes the profiles 'look good' is fitting in rho^2 space.
         This essentially spreads out the edge region, allowing the GP to vary more quickly there while remaining smooth in the core.
+        TODO(ZanderKeith): There is a more rigorous way to do this with non-stationary kernels (see Chilenski's thesis) but this is good for now.
 
         Parameters
         ----------
@@ -372,18 +370,16 @@ class CmodThomsonGPProfiles:
             The rho values where the fitted profile should be evaluated.
         kernel : gpx.kernels.AbstractKernel
             The kernel to use for the Gaussian Process. Must be initialized outside this jitted function.
-        edge_rho : float
-            The rho value at the edge where the signal is expected to approach edge_value
-        edge_value : float
-            The expected signal value at edge_rho.
-        edge_error : float
-            The uncertainty associated with the edge_value.
-        epsilon : float
-            A small value added to avoid log(0) issues.
+        constraint_val : jnp.ndarray
+            Nx3 array of added (rho, value, error) points where the signal value is constrained.
+        constraint_grad : jnp.ndarray
+            Nx3 array of added (rho, gradient, error) points where the signal gradient is constrained.
         outlier_penalty : float
[…]
