### 2026-01-13 c23facf
Stopped fitting in log space since the epsilon led to numerical issues
