+++ disruption_py/machine/cmod/thomson.py
@@ -338,7 +338,6 @@ class CmodThomsonGPProfiles:
         constraint_grad: jnp.ndarray,
         outlier_penalty: float,
         outlier_cutoff: float,
-        shift: float,
         jitter: float,
         ):
         """
@@ -346,12 +345,12 @@ class CmodThomsonGPProfiles:
         Assumes there are no missing / nan values in the input data.
 
         Since this is a 1D profile of a physical plasma quantity, we can make a few assumptions to improve the fit quality:
-        1. The signal is always positive.
+        1. The signal is always positive (enforced by clipping predictions).
         2. The signal has known gradient points (e.g. d(signal)/d(rho) = 0 at rho=0).
-        3. The signal has known edge values (e.g. signal approaches some small value at edge_rho).
+        3. The signal has known edge values (e.g. signal approaches zero at the edge).
 
         Implementing these assumptions is done as follows:
-        1. Use a log-transform on the signal values.
+        1. Positivity is enforced by clipping negative predictions to zero.
         2. Synthetic data points are added with boundary conditions on the gradients and values.
 
         The final piece which makes the profiles 'look good' is fitting in rho^2 space.
@@ -378,8 +377,6 @@ class CmodThomsonGPProfiles:
             The error value to assign to identified outliers to effectively ignore them.
         outlier_cutoff : float
             The rho value below which outliers will be identified.
-        shift : float
-            A small shift value added before log-transform to avoid log(0) issues.
         jitter : float
             A small jitter value added to the diagonal of the covariance matrix for numerical stability.
 
@@ -401,37 +398,30 @@ class CmodThomsonGPProfiles:
         signal_error = jnp.where(valid_mask, signal_error, outlier_penalty)
 
         # Normalize inputs by average value to improve numerical stability
-        # Use nanmean to handle any remaining NaNs
         norm_factor = jnp.mean(signal_data)
         observations_norm = signal_data / norm_factor
         errors_norm = signal_error / norm_factor
-
-        # Transform to log-space to enforce positivity
-        # Shift values up by 1 to avoid log(0) issues and make the transformation more stable
-        observations_log = jnp.log(observations_norm + shift)
-        # Error in log space: std_y = |dy/dx| * std_x = (1/(x+shift)) * std_x
-        errors_log = errors_norm / (observations_norm + shift)
 
         # Transform rho to rho^2 space to allow more variation near edge while maintaining smoothness in core
         rho2 = signal_rho ** 2
 
         # Perform a quick GP fit without boundary conditions to identify outliers
         K_rough = kernel.gram(rho2.reshape(-1,1)).to_dense()  # K(X, X), prior covariance at observed points
-        noise_rough = jnp.diag(errors_log ** 2)  # Observation noise, GPJax uses variance
+        noise_rough = jnp.diag(errors_norm ** 2)  # Observation noise, GPJax uses variance
         K_total_rough = K_rough + noise_rough + jnp.eye(len(rho2)) * jitter  # Add small jitter for numerical stability
         # Make predictions at observed points
-        alpha = jnp.linalg.solve(K_total_rough, observations_log)
+        alpha = jnp.linalg.solve(K_total_rough, observations_norm)
         predictions_rough = K_rough @ alpha
         # Compute prediction variance at training points
         predictions_var = jnp.diag(K_rough - K_rough @ jnp.linalg.solve(K_total_rough, K_rough))
         predictions_std = jnp.sqrt(jnp.maximum(predictions_var, 0))
[…]
