### 2026-01-13 49458ad
Outliers now determined by their own error rather than fit error
