### 2026-01-13 78bea2d
Added ability to pass in an arbitrary rho transform function
