### 2026-01-13 32bc840
minimize diff for tokamak resolution
