### 2026-01-13 df1da73
rename builtin to physics methods
