### 2026-01-13 121c167
fix too-many-return-statements in tokamak resolution
