### 2026-01-13 b722a10
drop unnecessary dummy driver
