### 2026-01-13 898020a
add python shebangs
