+++ disruption_py/machine/mast/util.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 """
 Module for helper, not physics, methods.
 """
