+++ tests/test_time_setting.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 """
 Unit tests for the time_setting
 
