### 2026-01-13 7777e00
adapt efit example for MAST
