### 2026-01-13 ceeea51
adapt sql example for MAST
