### 2026-01-13 6af22d3
run linter on MAST config
