### 2026-01-13 b365c5d
Added gpjax and numpy as dependencies
