### 2026-01-14 f8c1a15
Moved rho remapping outside JIT'd function
