### 2026-01-14 3ac8f12
cache MAST get_ip_parameters for generic get_time_domain
