+++ disruption_py/machine/mast/physics.py
@@ -7,6 +7,7 @@ Physics methods for MAST.
 
 import numpy as np
 
+from disruption_py.core.physics_method.caching import cache_method
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.errors import CalculationError
 from disruption_py.core.physics_method.params import PhysicsMethodParams
@@ -22,6 +23,7 @@ class MastPhysicsMethods:
     """
 
     @staticmethod
+    @cache_method
     @physics_method(
         columns=["ip"],
         tokamak=Tokamak.MAST,
