### 2026-01-14 5a519ac
fix mast test shotlist
