### 2026-01-16 c603628
Added method to grab Thomson channels in C-Mod
