+++ disruption_py/machine/cmod/physics.py
@@ -8,6 +8,7 @@ import warnings
 
 import numpy as np
 import scipy.constants as const
+import xarray as xr
 
 from disruption_py.core.physics_method.caching import cache_method
 from disruption_py.core.physics_method.decorator import physics_method
@@ -1471,6 +1472,201 @@ class CmodPhysicsMethods:
         return CmodPhysicsMethods._get_peaking_factors(
             params.times, ts_time, ts_te, ts_ne, ts_z, efit_time, bminor, z0
         )
+
+    @staticmethod
+    @cache_method
+    def _get_thomson_channels_raw(params: PhysicsMethodParams):
+        """
+        Retrieve raw Thomson scattering measurement channels for electron temperature and density.
+        Combines both core and edge Thomson data into a single dataset.
+
+        A few notes:
+        - Thomson YAG time by default runs from -0.5s to 2.0s (https://cmodwiki.psfc.mit.edu/index.php/Thomson_scattering_programming)
+        - See https://cmodwiki.psfc.mit.edu/index.php/Thomson_Scattering
+        """
+
+        core_nodes = {
+            "rho": ".yag_new.results.profiles.rho_t",
+            "ne": ".yag_new.results.profiles.ne_rz",
+            "ne_error": ".yag_new.results.profiles.ne_err",
+            "te": ".yag_new.results.profiles.te_rz",
+            "te_error": ".yag_new.results.profiles.te_err",
+        }
+        edge_nodes = {
+            "rho": ".yag_edgets.results.psinorm",
+            "ne": ".yag_edgets.results.ne",
+            "ne_error": ".yag_edgets.results.ne.error",
+            "te": ".yag_edgets.results.te",
+            "te_error": ".yag_edgets.results.te.error",
+        }
+
+        data_vars = {}
+        for region, nodes in zip(["core", "edge"], [core_nodes, edge_nodes]):
+            try:
+                # Retrieve rho data
+                rho_data, rho_time = params.mds_conn.get_data_with_dims(
+                    nodes["rho"],
+                    tree_name="electrons",
+                )
+                # Cut time to positive values only
+                valid_time_indices = rho_time >= 0
+                rho_data = rho_data[:, valid_time_indices]
+                rho_time = rho_time[valid_time_indices]
+                # Replace -1 with nan
+                rho_data[rho_data == -1] = np.nan
+
+                data_vars[f"rho_{region}"] = xr.DataArray(
+                    rho_data.astype(np.float32).T,
+                    dims=("time", f"{region}_channel"),
+                    coords={
+                        "time": rho_time,
+                        f"{region}_channel": np.arange(rho_data.shape[0]),
+                    },
+                )
+
+                # Retrieve profiles for ne and te
+                for quant in ["ne", "te"]:
+                    profile_data, profile_time = params.mds_conn.get_data_with_dims(
+                        nodes[quant],
+                        tree_name="electrons",
+                    )
+                    error_data = params.mds_conn.get_data(
+                        nodes[f"{quant}_error"],
+                        tree_name="electrons",
+                    )
+                    # Cut time to positive values only
+                    valid_time_indices = profile_time >= 0
+                    profile_data = profile_data[:, valid_time_indices]
+                    error_data = error_data[:, valid_time_indices]
+                    profile_time = profile_time[valid_time_indices]
+                    # Replace 0 with nan
+                    profile_data[profile_data == 0] = np.nan
+                    error_data[error_data == 0] = np.nan
+
+                    if region == "edge" and quant == "te":
+                        # Edge Te profile is given in eV, convert to keV
+                        profile_data = profile_data / 1000.0
+                        error_data = error_data / 1000.0
+
+                    # Raise error if profile and rho time do not match
[…]
