### 2026-01-16 c058349
Bumped numpy version
