+++ pyproject.toml
@@ -49,7 +49,7 @@ dependencies = [
   "dynaconf>=3.2.0",
   "loguru>=0.7.0",
   "netcdf4>=1.7.0,<1.7.3",
-  "numpy>=1.26.0,<2.0.0",
+  "numpy>=2",
   "pandas>=2.2.0,<2.3.3",
   "pyodbc>=5.2.0",
   "scipy>=1.15.0",
