### 2026-01-16 d5cd0d2
Added method to get btor
