+++ disruption_py/machine/cmod/config.toml
@@ -1,5 +1,5 @@
 [cmod.efit]
-tree = "efit18"
+tree = "efit21"
 
 [cmod.inout.mds]
 mdsplus_connection_string = "alcdata-archives"
