### 2026-01-19 a2fe2a4
Minor adjustments to make Disruption-Py pull data on omega
