### 2026-01-20 0160588
set explicit compat for xr.merge
