### 2026-01-23 #9364 @jatkinson1000
Hi All, I'm back from leave so just checking in on this.

Thank you @LiamPattinson, @jackleland, and @rmrsk for your comprehensive reviews and for persevering with evaluating this software.

From reading through the comments and issues for this review I am wondering if this software is really in a state where it can be easily reviewed and is easily usable as open-source software by other researchers. Particular points are:

- use of code in an external branch, not the main package as installed from PyPI,
- various issues with underlying core functionality in databases, dependencies, etc. and
- lack of clear instructions for installation and usage

as raised by multiple reviewers.
This is in addition to other paper-related issues such as protracted authorship debate and reference queries.

It does not seem to me that these are likely to be resolved in on the timescales expected of a JOSS review
See [JOSS Submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements)
> JOSS requires that software should be feature-complete (i.e., no half-baked solutions), [and] packaged appropriately according to common community standards for the programming language being used (e.g., [Python](https://packaging.python.org/), [R](https://r-pkgs.org/index.html))

There would be no issues with withdrawal and resubmission to JOSS at a later date ([noting that the scope has been updated](https://github.com/openjournals/joss-reviews/issues/9364#issuecomment-3719089734)) when all of the the relevant code is in the main branch, installable from PyPI, development has reached a stable point for usability and release, and installation/usage issues have been ironed out. Doing so would be a much more effective use of time for our volunteer reviewers who have already done more here than would usually be expected.

I will not take any action at this moment, but would appreciate if the reviewers could please summarise their thoughts on the above having looked into the software in depth.

Ref:

- [JOSS Author Guidelines](https://joss.readthedocs.io/en/latest/submitting.html#)
- [JOSS Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
