### 2026-01-23 #9364 @LiamPattinson
Thanks @jatkinson1000, I would personally like to see the `slj/mast` branch merged into `main` and a new release to be made before continuing, as without that I'm not able to verify the software's core functionality. I've just checked in on how it's doing, and it looks like it is now passing tests, but I can't say how close it is to being feature-complete.

I think a more detailed worked example or two in the docs would strengthen the project, as while I am able to piece together how things work using the `examples/` scripts and the API docs, there's not a lot to go on as a new user from outside the project. I don't think it would need to be all that complicated; just a light tutorial on how to write a script using `get_shots_data` that pulls some MAST data and then uses it to make a couple of plots relevant to disruptions. It would be useful if it could explain what each of the inputs are and what options are available, especially for things like `RetrievalSettings` which is a bit mysterious to me.
