### 2026-01-24 b585f73
Add hbtep fastcam shotlist and descriptions
