### 2026-01-26 e492f0f
Added label_thermal_quench_onset_time()  method
