+++ disruption_py/machine/cmod/physics.py
@@ -2115,6 +2115,147 @@ class CmodPhysicsMethods:
         v_surf = interp1(efit_time, v_surf, params.times)
 
         return {"v_surf": v_surf}
+    
+    @staticmethod
+    @physics_method(columns=["thermal_quench_time"], tokamak=Tokamak.CMOD)
+    def get_thermal_quench_onset_time(params: PhysicsMethodParams):
+        """
+        Labels the onset time of the thermal quench for a given shot (NaN for non-disruptive shots)
+        using a vertical SXR array due to its off-axis views and robustness across shots,
+        as opposed to ECE. The labeling method is non-causal (i.e. post-shot processing).
+        The TQ is found by searching for min(dSXR/dt) in a time window prior to the CQ.
+        There is a tension between using longer windows to find the first TQ in a multi-stage TQ
+        versus using a shorter window to avoid labeling sawtooth crashes.
+        Thus, for shots with multi-stage thermal quenches, (see shots 1050830034 and 1120717002),
+        this algorithm struggles to select the first thermal quench. Based on manual testing
+        of 120 shots, about 5% of flattop disruptions on C-Mod feature multi-stage thermal quenches.
+        This algorithm has only been tested on flattop disruptions.
+        ----------
+        params: PhysicsMethodParams
+            The parameters storing the requested time base, shot id, etc
+        Returns
+        ----------
+        thermal_quench_time : array_like
+            time of thermal quench onset for the shot, identical values at each time-slice
+
+        Last Major Update: Henry Wietfeldt (1/26/26)
+        """
+        from scipy.signal import butter, filtfilt, resample_poly
+        thermal_quench_time = np.full(len(params.times), np.nan)
+
+        # Get current data for obtaining start of current quench
+        ip, magtime = params.mds_conn.get_data_with_dims(
+            r"\ip", tree_name="magnetics"
+        )
+        ip = np.abs(ip)
+
+        # Get SXR chords
+        n_chords = 38
+        array_path = r"\top.brightnesses.array_1"
+        try:
+            chord_01, t_sxr = params.mds_conn.get_data_with_dims(
+                array_path + ":chord_01",
+                tree_name="xtomo",
+            ) # Units: W/m^2, s
+        except mdsExceptions.MdsException:
+            print(params.shot_id)
+            params.logger.debug("Failed to get SXR " + array_path + " data")
+            return {"thermal_quench_time": np.full(len(params.times), np.nan)}
+        valid_times = (t_sxr > 0) & (t_sxr < 2.0)
+        t_sxr = t_sxr[valid_times]
+        sxr = np.zeros(shape=(n_chords, len(t_sxr)))
+        sxr[0] = chord_01[valid_times]
+        # Get all other SXR chords
+        for i in range(1, n_chords):
+            try:
+                chord, t_chord = params.mds_conn.get_data_with_dims(
+                    array_path + ":chord_" + f"{i+1:02}",
+                    tree_name="xtomo",
+                )
+            except mdsExceptions.MdsException:
+                params.logger.debug("Failed to get SXR " + array_path + " chord " + str(i+1) + " data")
+                sxr[i] = 0.
+                continue
+            # Subtract constant background
+            chord = chord - np.mean(chord[t_chord < 0.])
+            # Occasionally the time bases of a chord are of a different length
[…]
