### 2026-01-27 9ec25e9
Gracefully return NaN for non-disruptive shots
