### 2026-01-29 751dfcc
revert logger changes for better debugging
