### 2026-01-29 7eab839
add MAST to remote tests
