### 2026-01-29 ed51e4e
tweak new dependencies
