+++ pyproject.toml
@@ -51,13 +51,13 @@ dependencies = [
   "loguru>=0.7.0",
   "netcdf4>=1.7.0,<1.7.3",
   "numpy>=1.26.0,<2.0.0",
-  "pandas[pyarrow]>=2.2.0,<2.3.3",
+  "pandas[parquet]>=2.2.0,<2.3.3",
   "pyodbc>=5.2.0",
   "scipy>=1.15.0",
   "SQLAlchemy>=2.0.0",
   "tqdm>=4.67.0",
-  "xarray>=2025.1.0,<2025.8.0",
-  "zarr>=3",
+  "xarray>=2025.1.0",
+  "zarr>=3.0.0",
 ]
 description = "An open-source physics-based Scientific Framework for Disruption Analysis of Fusion Plasmas for AI/ML applications"
 keywords = [
