### 2026-01-29 ed97f2c
add Sam to explicit author list
