### 2026-01-29 db2c74f
readd EAST return statement
