+++ disruption_py/machine/method_holders.py
@@ -24,6 +24,7 @@ def get_method_holders(tokamak: Tokamak):
     if tokamak is Tokamak.EAST:
         from disruption_py.machine.east import METHOD_HOLDERS
 
+        return GENERIC + METHOD_HOLDERS
     if tokamak is Tokamak.MAST:
         from disruption_py.machine.mast import METHOD_HOLDERS
 
