### 2026-01-30 04fd8ae
Fix copilot code review issues.
