+++ disruption_py/machine/mast/efit.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 
 """
-Module for retrieving and processing EFIT parameters for CMOD.
+Module for retrieving and processing EFIT parameters for MAST.
 """
 
 from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.params import PhysicsMethodParams
-from disruption_py.inout.xarray import XarrayConnection
+from disruption_py.inout.xarray_backend import XarrayConnection
 from disruption_py.machine.mast.physics import MastPhysicsMethods
 
 
