### 2026-01-30 fdd1567
Add publications that used the hbt fast camera shotlist
