### 2026-01-30 c5a99d0
Add CMOD attributes
