+++ disruption_py/machine/cmod/config.toml
@@ -10,6 +10,341 @@ driver = "ODBC"
 host = "alcdb2"
 port = 1433
 
+[cmod.physics.attributes.a_minor]
+description = "Minor radius (EFIT)."
+imas = "/equilibrium/time_slice(itime)/boundary/minor_radius"
+imas_description = "Minor radius of the plasma boundary (defined as (Rmax-Rmin) / 2 of the boundary)."
+unit = "dimensionless"
+
+[cmod.physics.attributes.beta_n]
+description = "Normalized plasma beta computed from EFIT."
+imas = "/summary/global_quantities/beta_tor_norm"
+imas_description = "Normalized toroidal beta, defined as 100 * beta_tor * a[m] * B0 [T] / ip [MA]."
+unit = "dimensionless"
+validity = [0, 2]
+
+[cmod.physics.attributes.beta_p]
+description = "Beta poloidal (EFIT)."
+imas = "/equilibrium/time_slice(itime)/global_quantities/beta_pol"
+imas_description = "Poloidal beta. Defined as betap = 4 int(p dV) / (R_0 * mu_0 * Ip^2)."
+unit = "dimensionless"
+validity = [0, 1.5]
+
+[cmod.physics.attributes.bt]
+description = "Toroidal magnetic field strength."
+imas = "/summary/global_quantities/b0"
+imas_description = "Vacuum toroidal field at R0. Positive sign means anti-clockwise when viewed from above. The product R0B0 must be consistent with the b_tor_vacuum_r field of the tf IDS."
+unit = "T"
+
+[cmod.physics.attributes.chisq]
+description = "Magnetic chi^2 (EFIT)."
+unit = "dimensionless"
+validity = [0.1, 40]
+
+[cmod.physics.attributes.dbetap_dt]
+description = "Time derivative of the poloidal beta (EFIT)."
+unit = "s^-1"
+
+[cmod.physics.attributes.dip_dt]
+description = "Time derivative of the Actual plasma current."
+unit = "A/s"
+
+[cmod.physics.attributes.dip_smoothed]
+description = "Smoothed time derivative of the measured plasma current."
+unit = "A/s"
+
+[cmod.physics.attributes.dipprog_dt]
+description = "Time derivative of the programmed plasma current."
+unit = "A"
+validity = [-5.0e6, 5.0e6]
+
+[cmod.physics.attributes.dli_dt]
+description = "Time derivative of the internal inductance (EFIT)."
+unit = "s^-1"
+
+[cmod.physics.attributes.dn_dt]
+description = "Time derivative of the measured line-averaged electron density."
+imas = "/summary/line_averaged/dn_e_dt"
+imas_description = "Time derivative of the line averaged electron density."
+unit = "m^-3*s^-1"
+
+[cmod.physics.attributes.dprad_dt]
+description = "Time derivative of the measured radiated power."
+unit = "W/s"
+
+[cmod.physics.attributes.dwmhd_dt]
+description = "Time derivative of the total plasma energy (EFIT)."
+unit = "J/s"
+
+[cmod.physics.attributes.greenwald_fraction]
+description = "Greenwald fraction."
+imas = "/summary/global_quantities/greenwald_fraction"
+imas_description = "Greenwald fraction =line_average/n_e/value divided by (global_quantities/ip/value *1e6 * pi * minor_radius^2)."
+unit = "dimensionless"
+validity = [0, 1.5]
+
+[cmod.physics.attributes.i_efc]
+description = "Error field correction (EFC) current."
+unit = "A"
+
+[cmod.physics.attributes.ip]
+description = "Actual plasma current."
+imas = "/summary/global_quantities/ip"
[…]
