+++ disruption_py/machine/mast/physics.py
@@ -12,7 +12,7 @@ from disruption_py.core.physics_method.decorator import physics_method
 from disruption_py.core.physics_method.errors import CalculationError
 from disruption_py.core.physics_method.params import PhysicsMethodParams
 from disruption_py.core.utils.math import interp1
-from disruption_py.inout.xarray_backend import XarrayConnection
+from disruption_py.inout.xr import XarrayConnection
 from disruption_py.machine.tokamak import Tokamak
 
 
