### 2026-02-02 5d15507
add get_metadata utility function
