+++ disruption_py/core/utils/misc.py
@@ -4,10 +4,11 @@
 Module for utility functions related to class instantiation, data manipulation, and version control.
 """
 
+import importlib.metadata
 import os
 import subprocess
 import sys
-import time
+from datetime import datetime
 from functools import lru_cache
 from pathlib import Path
 from tempfile import mkdtemp
@@ -59,6 +60,41 @@ def get_commit_hash() -> str:
 
 
 @lru_cache
+def get_metadata() -> Dict[str, str]:
+    """
+    Gather workflow metadata.
+
+    Returns
+    -------
+    Dict[str, str]
+        The workflow metadata.
+    """
+
+    package, *_ = __name__.split(".")
+    version = importlib.metadata.version(package)
+    tag = "v" + ".".join(version.split(".")[:2])
+    repo = "https://github.com/MIT-PSFC/disruption-py"
+    commit = get_commit_hash()
+    if commit:
+        source = f"{repo}/tree/{commit}"
+    else:
+        source = f"{repo}/releases/tag/{tag}"
+
+    metadata = {
+        "user": os.getenv("USER"),
+        "host": os.uname().nodename,
+        "time": datetime.now().isoformat(),
+        "package": package,
+        "version": version,
+        "commit": commit,
+        "source": source,
+    }
+    if not commit:
+        metadata.pop("commit")
+    return metadata
+
+
+@lru_cache
 def get_temporary_folder() -> str:
     """
     Create and return a temporary folder.
@@ -75,12 +111,12 @@ def get_temporary_folder() -> str:
         os.getenv("LOCALSCRATCH", "/tmp"),
         os.getenv("USER"),
         "disruption-py",
-        ("." if "pytest" in sys.modules else "") + time.strftime("%Y-%m-%d"),
+        ("." if "pytest" in sys.modules else "") + datetime.now().strftime("%Y-%m-%d"),
     )
     Path(top).mkdir(parents=True, exist_ok=True)
 
     # create temporary sub folder
-    return mkdtemp(dir=top, prefix=time.strftime("%H.%M.%S-"))
+    return mkdtemp(dir=top, prefix=datetime.now().strftime("%H.%M.%S-"))
 
 
 def shot_msg(message: str) -> str:
