### 2026-02-02 e717bb4
add metadata as dataset attributes
