+++ disruption_py/core/physics_method/runner.py
@@ -11,6 +11,7 @@ from typing import Dict
 import numpy as np
 import xarray as xr
 
+from disruption_py.config import config
 from disruption_py.core.physics_method.errors import CalculationError
 from disruption_py.core.physics_method.metadata import (
     BoundMethodMetadata,
@@ -18,7 +19,7 @@ from disruption_py.core.physics_method.metadata import (
     is_physics_method,
 )
 from disruption_py.core.physics_method.params import PhysicsMethodParams
-from disruption_py.core.utils.misc import get_elapsed_time, to_tuple
+from disruption_py.core.utils.misc import get_elapsed_time, get_metadata, to_tuple
 from disruption_py.inout.mds import mdsExceptions
 from disruption_py.machine.method_holders import get_method_holders
 from disruption_py.settings.retrieval_settings import RetrievalSettings
@@ -282,6 +283,10 @@ def populate_shot(
 
     # merge dataarrays/datasets into dataset
     dataset = xr.merge(datasets, compat="no_conflicts")
+
+    # set attributes for dataset
+    dataset.attrs.update({"tokamak": physics_method_params.tokamak.value})
+    dataset.attrs.update(get_metadata())
 
     # log statistics
     if dataset:
