### 2026-02-02 85592d3
catch AttributeError, too
