### 2026-02-02 34b1169
fixup! catch AttributeError, too
