### 2026-02-02 2ce13c8
prefer dict.get over try-except
