### 2026-02-02 4d8f809
build imas url as derived attribute
