### 2026-02-02 031050b
add shortcut for HBT-EP shotlist
