### 2026-02-02 8051371
tweak HBT-EP publications
