### 2026-02-02 4d4a376
Added methods to get zipfits and wmhdf, updated config to get 1 kHz EFIT results
