### 2026-02-02 59faf2c
unpin pandas and netcdf4
