### 2026-02-03 a0b0a6d
Small updates from code review
