+++ disruption_py/machine/mast/physics.py
@@ -13,6 +13,7 @@ from disruption_py.core.physics_method.errors import CalculationError
 from disruption_py.core.physics_method.params import PhysicsMethodParams
 from disruption_py.core.utils.math import interp1
 from disruption_py.inout.xr import XarrayConnection
+from disruption_py.machine.mast.util import MastUtilMethods
 from disruption_py.machine.tokamak import Tokamak
 
 
@@ -25,7 +26,7 @@ class MastPhysicsMethods:
     @staticmethod
     @cache_method
     @physics_method(
-        columns=["ip"],
+        columns=["ip", "dip_dt", "ip_prog", "dipprog_dt"],
         tokamak=Tokamak.MAST,
     )
     def get_ip_parameters(params: PhysicsMethodParams):
@@ -53,10 +54,10 @@ class MastPhysicsMethods:
 
         times = params.times
 
-        ip = MastPhysicsMethods.interpolate_1d(magtime, ip, times)
-        ip_prog = MastPhysicsMethods.interpolate_1d(ip_prog_time, ip_prog, times)
-        dip_dt = MastPhysicsMethods.interpolate_1d(magtime, dip_dt, times)
-        dipprog_dt = MastPhysicsMethods.interpolate_1d(ip_prog_time, dipprog_dt, times)
+        ip = MastUtilMethods.interpolate_1d(magtime, ip, times)
+        ip_prog = MastUtilMethods.interpolate_1d(ip_prog_time, ip_prog, times)
+        dip_dt = MastUtilMethods.interpolate_1d(magtime, dip_dt, times)
+        dipprog_dt = MastUtilMethods.interpolate_1d(ip_prog_time, dipprog_dt, times)
 
         return {
             "ip": ip,
@@ -91,8 +92,8 @@ class MastPhysicsMethods:
         base_time = conn.get_data(params.shot_id, "summary/time")
 
         times = params.times
-        power_nbi = MastPhysicsMethods.interpolate_1d(base_time, power_nbi, times)
-        power_radiated = MastPhysicsMethods.interpolate_1d(
+        power_nbi = MastUtilMethods.interpolate_1d(base_time, power_nbi, times)
+        power_radiated = MastUtilMethods.interpolate_1d(
             base_time, power_radiated, times
         )
         return {"power_nbi": power_nbi, "power_radiated": power_radiated}
@@ -124,13 +125,11 @@ class MastPhysicsMethods:
         base_time = conn.get_data(params.shot_id, "gas_injection/time")
 
         times = params.times
-        total_injected = MastPhysicsMethods.interpolate_1d(
+        total_injected = MastUtilMethods.interpolate_1d(
             base_time, total_injected, times
         )
-        inboard_total = MastPhysicsMethods.interpolate_1d(
-            base_time, inboard_total, times
-        )
-        outboard_total = MastPhysicsMethods.interpolate_1d(
+        inboard_total = MastUtilMethods.interpolate_1d(base_time, inboard_total, times)
+        outboard_total = MastUtilMethods.interpolate_1d(
             base_time, outboard_total, times
         )
         return {
@@ -165,8 +164,8 @@ class MastPhysicsMethods:
         n_e_core = conn.get_data(params.shot_id, "thomson_scattering/n_e_core")
         base_time = conn.get_data(params.shot_id, "thomson_scattering/time")
 
-        t_e_core = MastPhysicsMethods.interpolate_1d(base_time, t_e_core, times)
-        n_e_core = MastPhysicsMethods.interpolate_1d(base_time, n_e_core, times)
+        t_e_core = MastUtilMethods.interpolate_1d(base_time, t_e_core, times)
[…]
