+++ disruption_py/machine/mast/config.toml
@@ -14,9 +14,14 @@ expected_failure_columns = []
 test_columns = []
 
 [mast.tests.shots]
-a_fast = 30421
-b_full = 30113
-c_full = 30014
+disruption_fast = 30421
+no_disruption_fast = 29666
+disruption1_full = 30113
+disruption2_full = 30014
+no_disruption1_full = 29165
+no_disruption2_full = 28864
+
+
 
 [mast.time]
 disruption_time_start = 0.2
