### 2026-02-03 0f81b5b
Update deps to Jan 2026 (#514)
* unpin pandas and netcdf4

* lock deps

* black! disruption_py/core/physics_method/params.py

* black! disruption_py/machine/cmod/physics.py

* black! disruption_py/machine/generic/physics.py

* black! disruption_py/settings/shotlist_setting.py

* black! docs/examples/custom_time_setting.py
