### 2026-02-03 f811279
Add HBT-EP shotlist (#508)
* Add hbtep fastcam shotlist and descriptions

* Update disruption_py/data/README.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update readme

* Add publications that used the hbt fast camera shotlist

* remove BOM bits

* add shortcut for HBT-EP shotlist

* tweak HBT-EP publications

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
Co-authored-by: gtrevisan <gtrevisan@users.noreply.github.com>
