### 2026-02-03 def2a52
drop imas description
