+++ disruption_py/machine/cmod/config.toml
@@ -13,27 +13,23 @@ port = 1433
 [cmod.physics.attributes.a_minor]
 description = "Minor radius (EFIT)."
 imas = "/equilibrium/time_slice(itime)/boundary/minor_radius"
-imas_description = "Minor radius of the plasma boundary (defined as (Rmax-Rmin) / 2 of the boundary)."
 unit = "dimensionless"
 
 [cmod.physics.attributes.beta_n]
 description = "Normalized plasma beta computed from EFIT."
 imas = "/summary/global_quantities/beta_tor_norm"
-imas_description = "Normalized toroidal beta, defined as 100 * beta_tor * a[m] * B0 [T] / ip [MA]."
 unit = "dimensionless"
 validity = [0, 2]
 
 [cmod.physics.attributes.beta_p]
 description = "Beta poloidal (EFIT)."
 imas = "/equilibrium/time_slice(itime)/global_quantities/beta_pol"
-imas_description = "Poloidal beta. Defined as betap = 4 int(p dV) / (R_0 * mu_0 * Ip^2)."
 unit = "dimensionless"
 validity = [0, 1.5]
 
 [cmod.physics.attributes.bt]
 description = "Toroidal magnetic field strength."
 imas = "/summary/global_quantities/b0"
-imas_description = "Vacuum toroidal field at R0. Positive sign means anti-clockwise when viewed from above. The product R0B0 must be consistent with the b_tor_vacuum_r field of the tf IDS."
 unit = "T"
 
 [cmod.physics.attributes.chisq]
@@ -65,7 +61,6 @@ unit = "s^-1"
 [cmod.physics.attributes.dn_dt]
 description = "Time derivative of the measured line-averaged electron density."
 imas = "/summary/line_averaged/dn_e_dt"
-imas_description = "Time derivative of the line averaged electron density."
 unit = "m^-3*s^-1"
 
 [cmod.physics.attributes.dprad_dt]
@@ -79,7 +74,6 @@ unit = "J/s"
 [cmod.physics.attributes.greenwald_fraction]
 description = "Greenwald fraction."
 imas = "/summary/global_quantities/greenwald_fraction"
-imas_description = "Greenwald fraction =line_average/n_e/value divided by (global_quantities/ip/value *1e6 * pi * minor_radius^2)."
 unit = "dimensionless"
 validity = [0, 1.5]
 
@@ -90,7 +84,6 @@ unit = "A"
 [cmod.physics.attributes.ip]
 description = "Actual plasma current."
 imas = "/summary/global_quantities/ip"
-imas_description = "Total plasma current (toroidal component). Positive sign means anti-clockwise when viewed from above."
 unit = "A"
 
 [cmod.physics.attributes.ip_error]
@@ -100,48 +93,41 @@ unit = "A"
 [cmod.physics.attributes.ip_prog]
 description = "Programmed plasma current."
 imas = "/pulse_schedule/flux_control/plasma_current/reference"
-imas_description = "Reference waveform. Caution : error bars of the reference/data node are not used in the usual sense, instead they are used to describe the control envelope, with a meaning depending on the chosen envelope_type option."
 unit = "A"
 
 [cmod.physics.attributes.kappa]
 description = "Elongation at plasma boundary (EFIT)."
 imas = "/equilibrium/time_slice(itime)/global_quantities/elongation"
-imas_description = "Elongation."
 unit = "dimensionless"
 validity = [0.8, 2]
 
 [cmod.physics.attributes.kappa_area]
 description = "Elongation calculated using the plasma area and minor radius from EFIT."
 imas = "/summary/boundary/elongation"
-imas_description = "Elongation of the plasma boundary."
 unit = "dimensionless"
 validity = [0.8, 2]
 
[…]
