### 2026-02-03 7a758cc
black! disruption_py/settings/shotlist_setting.py
