### 2026-02-03 282633b
Revert "black! disruption_py/settings/shotlist_setting.py"
This reverts commit 7a758cce4f654a7cae0c664cd43527a290a23ce3.
