### 2026-02-03 96a2045
Add attributes, rename variables
