+++ disruption_py/machine/mast/efit.py
@@ -15,26 +15,26 @@ class MastEfitMethods:
     Class for retrieving and processing EFIT parameters for MAST.
     """
 
-    efit_properties = [
-        "beta_tor_normal",
-        "beta_tor",
-        "beta_pol",
-        "elongation",
-        "elongation_axis",
-        "magnetic_axis_r",
-        "magnetic_axis_z",
-        "triangularity_lower",
-        "triangularity_upper",
-        "minor_radius",
-        "bvac_rmag",
-        "bphi_rmag",
-        "li",
-        "vloop_static",
-        "q95",
-    ]
+    efit_properties = {
+        "beta_n": "beta_tor_normal",
+        "beta_t": "beta_tor",
+        "beta_p": "beta_pol",
+        "kappa": "elongation",
+        "rmagx": "magnetic_axis_r",
+        "rmagz": "magnetic_axis_z",
+        "tribot": "triangularity_lower",
+        "tritop": "triangularity_upper",
+        "a_minor": "minor_radius",
+        "bvac_rmag": "bvac_rmag",
+        "bphi_rmag": "bphi_rmag",
+        "li": "li",
+        "v_loop_static": "vloop_static",
+        "v_loop_dynamic": "vloop_dynamic",
+        "q95": "q95",
+    }
 
     @staticmethod
-    @physics_method(columns=efit_properties)
+    @physics_method(columns=list(efit_properties.keys()))
     def get_efit_parameters(params: PhysicsMethodParams):
         """
         Retrieve EFIT parameters for MAST.
@@ -54,9 +54,9 @@ class MastEfitMethods:
         times = params.times
 
         outputs = {}
-        for prop in MastEfitMethods.efit_properties:
+        for key, prop in MastEfitMethods.efit_properties.items():
             signal = conn.get_data(params.shot_id, f"equilibrium/{prop}")
             item = MastUtilMethods.interpolate_1d(eq_time, signal, times)
-            outputs[prop] = item
+            outputs[key] = item
 
         return outputs
