+++ disruption_py/machine/mast/config.toml
@@ -29,3 +29,148 @@ duration_before_disruption = 0.25
 end_of_current_threshold = 200e3
 minimum_duration = 0.1
 minimum_ip = 200e3
+
+
+[mast.physics.attributes.beta_p]
+description = "Beta poloidal (EFIT)."
+imas = "/equilibrium/time_slice(itime)/global_quantities/beta_pol"
+unitss = "dimensionless"
+
+[mast.physics.attributes.beta_t]
+description = "Beta toroidal (EFIT)."
+imas = "/equilibrium/time_slice(itime)/global_quantities/beta_tor"
+units = "dimensionless"
+
+[mast.physics.attributes.beta_n]
+description = "Toroidal beta, defined as the volume-averaged total perpendicular."
+imas = "/equilibrium/time_slice(itime)/global_quantities/beta_tor_normal"
+units = "dimensionless"
+
+[mast.physics.attributes.bphi_rmag]
+description = "Bphi at rmag"
+units = "T"
+
+[mast.physics.attributes.bvac_rmag]
+description = "Bvac at rmag"
+units = "T"
+
+[mast.physics.attributes.dalpha]
+description = "D-alpha camera signal"
+units = "V"
+
+[mast.physics.attributes.dip_dt]
+description = "Time derivative of the Actual plasma current."
+units = "A/s"
+
+[mast.physics.attributes.dipprog_dt]
+description = "Time derivative of the programmed plasma current."
+units = "A"
+
+[mast.physics.attributes.dn_dt]
+description = "Time derivative of the measured line-averaged electron density."
+imas = "/summary/line_averaged/dn_e_dt"
+units = "m^-3*s^-1"
+
+[mast.physics.attributes.kappa]
+description = "Elongation at plasma boundary (EFIT)."
+imas = "/equilibrium/time_slice(itime)/global_quantities/elongation"
+units = "dimensionless"
+
+[mast.physics.attributes.gas_inboard_total]
+description = "Gas injected by the inboard valves"
+units = "cps"
+
+[mast.physics.attributes.gas_outboard_total]
+description = "Gas injected by the outboard valves"
+units = "cps"
+
+[mast.physics.attributes.gas_total_injected]
+description = "Total cumulative injected gas"
+units = "counts"
+
+[mast.physics.attributes.greenwald_fraction]
+description = "Greenwald fraction."
+imas = "/summary/global_quantities/greenwald_fraction"
+units = "dimensionless"
+
+[mast.physics.attributes.ip]
+description = "Actual plasma current."
+imas = "/summary/global_quantities/ip"
+units = "A"
+
+[mast.physics.attributes.ip_prog]
+description = "Programmed plasma current."
+imas = "/pulse_schedule/flux_control/plasma_current/reference"
+units = "A"
+
+[mast.physics.attributes.li]
+description = "Internal inductance (EFIT)."
+imas = "/equilibrium/time_slice(itime)/global_quantities/li_3"
+units = "dimensionless"
+
+[mast.physics.attributes.rmagx]
+description = "Major radius of the magnetic axis"
+imas = "/equilibrium/time_slice(itime)/global_quantities/magnetic_axis_r"
+units = "m"
+
+[mast.physics.attributes.rmagz]
+description = "Height of the magnetic axis"
+imas = "/equilibrium/time_slice(itime)/global_quantities/magnetic_axis_z"
+units = "m"
+
+[mast.physics.attributes.a_minor]
+description = "Minor radius of the plasma boundary (defined as (Rmax-Rmin) / 2 of the boundary)"
+imas = "equilibrium/time_slice(itime)/boundary/minor_radius" 
+units = "m"
+
[…]
