### 2026-02-04 aa9538e
Add metadata to xarray outputs (#512)
* add get_metadata utility function

* use get_metadata in logging header

* add metadata as dataset attributes

* add data vars attributes from config

* explicit attrs logic in xr.concat

* catch AttributeError, too

* fixup! catch AttributeError, too

* prefer dict.get over try-except
