+++ disruption_py/machine/cmod/config.toml
@@ -13,304 +13,304 @@ port = 1433
 [cmod.physics.attributes.a_minor]
 description = "Minor radius (EFIT)."
 imas = "/equilibrium/time_slice(itime)/boundary/minor_radius"
-unit = "dimensionless"
+units = "dimensionless"
 
 [cmod.physics.attributes.beta_n]
 description = "Normalized plasma beta computed from EFIT."
 imas = "/summary/global_quantities/beta_tor_norm"
-unit = "dimensionless"
+units = "dimensionless"
 validity = [0, 2]
 
 [cmod.physics.attributes.beta_p]
 description = "Beta poloidal (EFIT)."
 imas = "/equilibrium/time_slice(itime)/global_quantities/beta_pol"
-unit = "dimensionless"
+units = "dimensionless"
 validity = [0, 1.5]
 
 [cmod.physics.attributes.bt]
 description = "Toroidal magnetic field strength."
 imas = "/summary/global_quantities/b0"
-unit = "T"
+units = "T"
 
 [cmod.physics.attributes.chisq]
 description = "Magnetic chi^2 (EFIT)."
-unit = "dimensionless"
+units = "dimensionless"
 validity = [0.1, 40]
 
 [cmod.physics.attributes.dbetap_dt]
 description = "Time derivative of the poloidal beta (EFIT)."
-unit = "s^-1"
+units = "s^-1"
 
 [cmod.physics.attributes.dip_dt]
 description = "Time derivative of the Actual plasma current."
-unit = "A/s"
+units = "A/s"
 
 [cmod.physics.attributes.dip_smoothed]
 description = "Smoothed time derivative of the measured plasma current."
-unit = "A/s"
+units = "A/s"
 
 [cmod.physics.attributes.dipprog_dt]
 description = "Time derivative of the programmed plasma current."
-unit = "A"
+units = "A"
 validity = [-5.0e6, 5.0e6]
 
 [cmod.physics.attributes.dli_dt]
 description = "Time derivative of the internal inductance (EFIT)."
-unit = "s^-1"
+units = "s^-1"
 
 [cmod.physics.attributes.dn_dt]
 description = "Time derivative of the measured line-averaged electron density."
 imas = "/summary/line_averaged/dn_e_dt"
-unit = "m^-3*s^-1"
+units = "m^-3*s^-1"
 
 [cmod.physics.attributes.dprad_dt]
 description = "Time derivative of the measured radiated power."
-unit = "W/s"
+units = "W/s"
 
 [cmod.physics.attributes.dwmhd_dt]
 description = "Time derivative of the total plasma energy (EFIT)."
-unit = "J/s"
+units = "J/s"
 
 [cmod.physics.attributes.greenwald_fraction]
 description = "Greenwald fraction."
 imas = "/summary/global_quantities/greenwald_fraction"
-unit = "dimensionless"
+units = "dimensionless"
 validity = [0, 1.5]
 
 [cmod.physics.attributes.i_efc]
 description = "Error field correction (EFC) current."
-unit = "A"
+units = "A"
 
 [cmod.physics.attributes.ip]
 description = "Actual plasma current."
 imas = "/summary/global_quantities/ip"
-unit = "A"
+units = "A"
 
 [cmod.physics.attributes.ip_error]
 description = "Difference between the actual and programmed plasma current."
-unit = "A"
+units = "A"
 
 [cmod.physics.attributes.ip_prog]
 description = "Programmed plasma current."
 imas = "/pulse_schedule/flux_control/plasma_current/reference"
-unit = "A"
+units = "A"
 
 [cmod.physics.attributes.kappa]
 description = "Elongation at plasma boundary (EFIT)."
 imas = "/equilibrium/time_slice(itime)/global_quantities/elongation"
-unit = "dimensionless"
[…]
