### 2026-02-04 a415f4e
Apply suggestion from @gtrevisan
Co-authored-by: Gregorio L. Trevisan <gtrevisan@users.noreply.github.com>
