### 2026-02-04 a85e50a
Rename dalpha to d_alpha to make it consistent with D3D h_alpha
