### 2026-02-04 c635b4a
Unlocked numpy version, added additional signals for transport predictor
