+++ disruption_py/machine/mast/config.toml
@@ -22,7 +22,7 @@ no_disruption2_full = 28864
 no_disruption_fast = 29666
 
 [mast.time]
-disruption_time_start = 0.2
+disruption_time_start = 0.0
 disruption_time_step = 0.001
 dt_before_disruption = 0.010
 duration_before_disruption = 0.25
