### 2026-02-05 788e6e8
downgrade opening data message to trace
