### 2026-02-05 02af96e
reviewed and updated cmod's config.toml file
