+++ disruption_py/machine/cmod/config.toml
@@ -46,7 +46,7 @@ description = "Time derivative of the Actual plasma current."
 units = "A/s"
 
 [cmod.physics.attributes.dip_smoothed]
-description = "Smoothed time derivative of the measured plasma current."
+description = "Time derivative of the smoothed measured plasma current."
 units = "A/s"
 
 [cmod.physics.attributes.dipprog_dt]
@@ -64,7 +64,7 @@ imas = "/summary/line_averaged/dn_e_dt"
 units = "m^-3*s^-1"
 
 [cmod.physics.attributes.dprad_dt]
-description = "Time derivative of the measured radiated power."
+description = "Time derivative of the measured total radiated power."
 units = "W/s"
 
 [cmod.physics.attributes.dwmhd_dt]
@@ -78,16 +78,16 @@ units = "dimensionless"
 validity = [0, 1.5]
 
 [cmod.physics.attributes.i_efc]
-description = "Error field correction (EFC) current."
+description = "Error field correction (EFC) coil current."
 units = "A"
 
 [cmod.physics.attributes.ip]
-description = "Actual plasma current."
+description = "Measured plasma current (Rogowski coil)."
 imas = "/summary/global_quantities/ip"
 units = "A"
 
 [cmod.physics.attributes.ip_error]
-description = "Difference between the actual and programmed plasma current."
+description = "Difference between the measured and programmed plasma current."
 units = "A"
 
 [cmod.physics.attributes.ip_prog]
@@ -108,13 +108,13 @@ units = "dimensionless"
 validity = [0.8, 2]
 
 [cmod.physics.attributes.li]
-description = "Internal inductance (EFIT)."
+description = "Normalized internal inductance, li(3) (EFIT)."
 imas = "/equilibrium/time_slice(itime)/global_quantities/li_3"
 units = "dimensionless"
 validity = [0.2, 4.5]
 
 [cmod.physics.attributes.lower_gap]
-description = "Bottom gap (EFIT)."
+description = "Bottom gap between x-point and wall (EFIT)."
 imas = "/equilibrium/time_slice(itime)/boundary/gap(i1)/value"
 units = "m"
 validity = [0.025, 0.25]
@@ -149,7 +149,7 @@ description = "Peaking factor of the electron density profile measured by the Th
 units = "dimensionless"
 
 [cmod.physics.attributes.p_icrf]
-description = "Ion cyclotron resonant heating power."
+description = "Total ion cyclotron resonant heating power."
 imas = "/summary/heating_current_drive/power_launched_ic"
 units = "W"
 validity = [0, inf]
@@ -159,19 +159,19 @@ description = "Total input power (ohmic + LH + ICRF)."
 units = "W"
 
 [cmod.physics.attributes.p_lh]
-description = "Lower hybrid heating power."
+description = "Total lower hybrid heating power."
 imas = "/summary/heating_current_drive/power_launched_lh"
 units = "W"
 validity = [0, inf]
 
 [cmod.physics.attributes.p_oh]
-description = "Ohmic heating power."
+description = "Total Ohmic heating power."
 imas = "/summary/global_quantities/power_ohm"
 units = "W"
 validity = [0, 5.0e6]
 
 [cmod.physics.attributes.p_rad]
-description = "Radiated power measured by the 2pi diode."
+description = "Total radiated power measured by the 2pi diode."
 imas = "/bolometer/power_radiated_total"
 units = "W"
 validity = [0, inf]
@@ -185,13 +185,13 @@ description = "Peaking factor of the electron temperature profile measured by th
 units = "dimensionless"
 
 [cmod.physics.attributes.q0]
-description = "q at magnetic axis (EFIT)."
+description = "Safety factor at magnetic axis (EFIT)."
[…]
