### 2026-02-05 182fa2d
fixup! add script to check imas urls
