### 2026-02-05 3cc86f2
Fix another typo found by copilot
