### 2026-02-06 #9364 @jatkinson1000
Thanks @rmrsk and @LiamPattinson for your feedback, it is very helpful.
@jackleland did you have any thoughts? I can see you are not as far through the checklist as Robert and Liam.

@gtrevisan before I make a decision on this did you have any comments on the feedback above or timescales for a reviewable code?

I'll look to make a decision on this next week.
