### 2026-02-06 456c238
update pulse schedule ip reference following C-MOD
https://github.com/MIT-PSFC/disruption-py/commit/3eb275a169d33a1c06bc9e3308e2e1808f233bcc
