### 2026-02-06 c9cff65
use shot_msg when reporting path failure, drop PID
