### 2026-02-06 2493137
sort efit properties, readd wmhd
