### 2026-02-06 #9364 @gtrevisan
we are almost done with the review of our MAST branch.
I expect to merge `slj/mast` onto `dev` within the next couple of (business) days -- let's say by next Tue.

I will then take another couple of days to prepare our new release, verify changes, carry out on-premises tests, unleash full-scale database workflows, and finally publish v0.13 -- let's say by next Thu.

in other words, if executing from `slj/mast` or `dev` is undesirable, then rest assured that **all reviewers will be able to continue their evaluation from a new `main` release within the next week**.

that said, there won't be any publication until you all reviewed and green-lit our code, so our minor finishing touches on the manuscript are going to be tackled last.
I believe reviewers should be able to do the same, and indeed the "Software paper" section in the checklist is placed last.

let me know whether I missed anything! and feel free to re/open any issue on our repo as needed.
