### 2026-02-06 #9364 @rmrsk
> in other words, if executing from `slj/mast` or `dev` is undesirable, then rest assured that **all reviewers will be able to continue their evaluation from a new `main` release within the next week**.

Great! Please tag me when this is done :)

> that said, there won't be any publication until you all reviewed and green-lit our code, so our minor finishing touches on the manuscript are going to be tackled last. I believe reviewers should be able to do the same, and indeed the "Software paper" section in the checklist is placed last.

I'd like @jatkinson1000 or @kyleniemeyer to comment on whether or not this is acceptable JOSS review practice. I don't think it's fair to ask reviewers to perform partial reviews, especially since the author list is not finalized.  Personally, I'd like to review the complete submission, including the paper, but I will nonetheless accept the recommendations of the editors.
