### 2026-02-06 #9364 @jatkinson1000
Thanks @rmrsk and you are correct to ask.
JOSS reviewers are reviewing the paper and the software sustainability and functionality. They are NOT developers or source reviewers.

@gtrevisan Please do **not** ask our **_volunteer_** JOSS submission reviewers to be reviewing core development of your code.

As I said above, JOSS expects software to come to review in a functional and useable state ready for review.
One of these expectations is packaging and distribution as appropriate for the project - for Python this would be installation from PyPI or Conda, not a clone and dev install.

The same is true of the paper - this should also be ready for review at time of submission to JOSS.

Based on your outline above I will ask the reviewers to revisit the software and paper on **Friday the 13th March** (to account for Thursday timezones). If at that time they still feel that there are significant barriers to being able to review it in the manner of a regular JOSS submission I will be suggesting withdrawal and resubmission at a later date.

I will let @kyleniemeyer add any further thoughts as an editor in chief, but we have already given this submission a lot of leeway with the reviewers showing great patience and engagement.

CC @LiamPattinson @jackleland
