### 2026-02-06 #9364 @gtrevisan
the manuscript has been ready for review since Jul 2025.
I was referring to the minor comments raised by @rmrsk last month, which are going to be implemented as soon as possible.
also, the author list is final from my perspective.

I don't think I've received any other comment on the manuscript from @LiamPattinson or @jackleland at this time.
