### 2026-02-09 ea25bdf
log resolved tokamak as info
