### 2026-02-09 8f313ee
fixup! log resolved tokamak as info
