### 2026-02-09 e0205d0
Introduce MAST workflows (#509)
* lock deps

* try MDSplus first, then fall back onto mdsthin

* harmonize MDSplus imports and exceptions for C-MOD

* harmonize MDSplus imports and exceptions for DIII-D

* fixup! harmonize MDSplus imports and exceptions for DIII-D

* try to fix str conversions for arrays

* add skeleton of MAST tokamak

* Add compatibility with mdsthin

* Update with first data from FAIR MAST

* Remove print statement

* Re-add mast as a method

* Update pyproj.toml, poetry/uv locks

* Add support for different timebases to MAST

* Add test info to config

* Refactor s3 -> xar
[…]
