+++ docs/usage/physics_methods/mast_method_reference.md
@@ -1,0 +1,12 @@
+## Built-in Methods For MAST { .doc .doc-heading }
+::: disruption_py.machine.mast.physics
+    handler: python
+	options:
+	  heading_level: 2
+	  show_source: true
+	  show_if_no_docstring: true
+	  show_root_heading: false
+	  show_root_toc_entry: false
+	  show_root_members_full_path: false
+	  filters:
+		- "!^_"

[…]
