### 2026-02-09 0c4fcbe
Implement ts_data_type = ptdata
