+++ disruption_py/machine/d3d/physics.py
@@ -995,7 +995,7 @@ class D3DPhysicsMethods:
         - issues: #[261](https://github.com/MIT-PSFC/disruption-py/issues/261)
         """
         ## Thomson parameters
-        ts_data_type = "blessed"  # either 'blessed', 'unblessed', or 'ptdata'
+        ts_data_type = "blessed"  # either 'blessed', 'unblessed', or 'ptdata'   # NOTE: debug
         # metric to use for core/edge binning (either 'psin' or 'rhovn')
         ts_radius = "rhovn"
         # ts_radius value defining boundary of 'core' region (between 0 and 1)
@@ -1471,61 +1471,110 @@ class D3DPhysicsMethods:
         elif data_source == "unblessed":
             mds_path = r"\top.ts.revisions.revision00."
         elif data_source == "ptdata":
-            mds_path = r"\top.ts.blessed."  # Don't ask...I don't have the answer
-            raise NotImplementedError("ptdata case not fully implemented yet")  # TODO
+            # Get the Thomson data that are available in the real-time system
+            mds_path = r"\top.ts.blessed."  # Source for the radial & vertical chord positions
+            suffix = {"core": "cor", "tangential": "tan"}
+            # Account for pointname formatting change in 2017
+            if params.shot_id < 172749:  # First shot on Sep 19, 2017
+                suffix["tangential"] = "hor"
+                # TODO: figure out how 'hor' system was implemented. 
+                # 161228: tsshorne1 to tsshorne6 -- check MDSplus if there are only 6 chords
         else:
             raise CalculationError(f"Invalid data_source: {data_source}")
-
-        # Account for pointname formatting change in 2017 (however using ptdata is unimplemented)
-        # NOTE: "suffix" is only used if data_source="ptdata" which isn't implemented yet
-        suffix = {"core": "cor", "tangential": "tan"}
-        if params.shot_id < 172749:  # First shot on Sep 19, 2017
-            suffix["tangential"] = "hor"
 
         lasers = {}
         for laser in ts_systems:
             lasers[laser] = {}
             sub_tree = f"{mds_path}{laser}"
-            try:
-                (t_sub_tree,) = params.mds_conn.get_dims(
-                    f"{sub_tree}:temp", tree_name="electrons"
-                )
-                # lasers[laser]['time'] gets overwritten in the loop later
-                lasers[laser]["time"] = t_sub_tree / 1.0e3  # [ms] -> [s]
-            except mdsExceptions.MdsException as e:
-                lasers[laser] = None
-                params.logger.warning(
-                    "Failed to get {laser} time. Setting laser data to None.",
-                    laser=laser,
-                )
-                params.logger.opt(exception=True).debug(e)
-                continue
-            child_nodes = {
-                "r": "r",
-                "z": "z",
-                "te": "temp",
-                "ne": "density",
-                "time": "time",
-                "te_error": "temp_e",
-                "ne_error": "density_e",
-            }
-            for node, name in child_nodes.items():
+            if data_source == 'ptdata':
+                print('Using PTDATA for Thomson data source (experimental)')
+                
                 try:
-                    lasers[laser][node] = params.mds_conn.get_data(
-                        f"{sub_tree}:{name}", tree_name="electrons"
+                    lasers[laser]["time"] = params.mds_conn.get_dims(f"ptdata('tss{suffix[laser]}te00', {params.shot_id})")[0]
+                    # lasers[laser]["time"] /= 1.0e3 # [ms] -> [s]  TODO: uncomment this after fixing the 'time' overwritten bug
+                except:
+                    pass
+                
+                # Get r, z from MDSplus
+                # TODO: make this look better
+                child_nodes = {
+                    "r": "r",
+                    "z": "z",
+                }
+                for node, name in child_nodes.items():
+                    try:
[…]
