### 2026-02-11 #9364 @gtrevisan
version 0.13 is now officially released:
- https://github.com/MIT-PSFC/disruption-py/releases/tag/v0.13
- https://pypi.org/project/disruption-py/
- https://zenodo.org/records/18612738
