### 2026-02-11 #9364 @gtrevisan
also, the manuscript is now updated according to @rmrsk's suggestions:
- https://github.com/MIT-PSFC/disruption-py/tree/glt/joss
- https://github.com/MIT-PSFC/disruption-py/actions/runs/21919837714/artifacts/5471367339

further comments:

> 2. The Statement of need section describes DisruptionPy quite well, but does not state its relation to other packages/codes. If other related analysis tools exist, they should be briefly discussed and referenced. If not, a statement on the uniqueness of DisruptionPy would be appropriate.

we added references to both TokSearch and DEFUSE, which to our knowledge are the only somewhat similar tools.

https://github.com/MIT-PSFC/disruption-py/blob/695c72d4325cc98453e40cd93fafdf3e7bcb932f/paper.md?plain=1#L65-L71

> 3. The references only contain works from the Disruptions Group (Cristina Rea is an author of every referenced paper, except Stillerman 1997). Fusion plasma disruptions is not my field, so for all I know this may or may not be completely appropriate. But at the very least I'd like you to comment as to why there are no references to works outside of the Disruptions Group, and/or make amendments to the list of references if relevant.

that was the point of our last sentence:

https://github.com/MIT-PSFC/disruption-py/blob/695c72d4325cc98453e40cd93fafdf3e7bcb932f/paper.md?plain=1#L74

we expect that further visibility in the community through dissemination at conferences and relevant fora will raise the profile of our tool and facilitate broader adoption, all the more now that we support MAST open workflows.

> 4. The acknowledgement section only mentions the latest funding source, but it seems to me that the software has a longer and perhaps more complicated history? Are there other additional funding sources or projects that should be credited?

we added our ongoing MIT PSFC MFE award:

https://github.com/MIT-PSFC/disruption-py/blob/695c72d4325cc98453e40cd93fafdf3e7bcb932f/paper.md?plain=1#L78
