### 2026-03-01 #9364 @jatkinson1000
@rmrsk, @LiamPattinson, and @jackleland

Since you have now had some time to evaluate at the new release please could you confirm whether the required additions have been made and if you feel it is in a state in which you are able to review it for JOSS?

If so, I'd like to proceed as close to the proposed JOSS timeline as we can from now, given this has dragged on (through no fault of the reviewers).
