### 2026-03-02 #9364 @rmrsk
@jatkinson1000 @gtrevisan Yeah, I only have some minor comments at this point.

1. The paper looks all right, but I'd suggest refactoring the summary section. Presently, it doesn't actually describe the high-level functionality of the software, as required by the JOSS guidelines. See https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain. I suggest moving the present summary into the statement of need section, and moving paragraph 1 & 3 from the statement section into the summary.

2. The automated tests are all live tests that connect to servers with secret tokens. Is there any way the tests can be verified locally?

3. Consider including pytest as a dependency, as examples/efit.py relies on it.

Finally, these two points can be improved:

The current examples seem comparatively shallow, and could be extended to "solve real-world analysis problems". Internal Python functions have proper docstrings, but some of the functionality requires users to dig through the source code. It would be beneficial - although not mandatory - to build a full Python API documentation.
