### 2026-03-02 #9364 @gtrevisan
@rmrsk quick reply to part of your questions.

2. as discussed, [all machines apart from MAST](https://github.com/MIT-PSFC/disruption-py?tab=readme-ov-file#devices) require authentication.
 tests are trivial for any user if ran from those on-premise computers, but do require opening tunnels otherwise.
 in the spirit of CI/CD, I've set up remote testing to allow me to verify our workflows for all machines in parallel at every commit/PR.
 in any case, you are welcome to run tests for MAST locally -- please see below.

3. `pytest` is indeed a dependency [in the `dev` group](https://github.com/MIT-PSFC/disruption-py/blob/c0c9242183a09f124fb94e14f904552a38265781/pyproject.toml#L6-L14).
 tests or examples are not part of the PyPI release, as usual, but they are part of the repo and are useful starting points for users/developers -- please see below.

- **functionality documentation**: the docstrings for each physics method are already parsed and offered as an easily-browsable dynamically-generated webpage -- is this what you would like to see?
 https://mit-psfc.github.io/disruption-py/usage/physics_methods/mast_method_reference/#disruption_py.machine.mast.physics.MastPhysicsMethods.get_densities
 please correct me if I misunderstood.

### test execution

if someone clones our repo, local tests are as easy as:
```
make install test-fast test
```
which is equivalent to:
```bash
# install with dev deps
poetry install --with dev

# run fast tests
GITHUB_ACTIONS=true poetry run pytest -v tests

# run full tests
poetry run pytest -v tests
```
or any other command to a similar effect, eg:
```
uv sync --group dev
uv run pytest
```
