### 2026-03-03 #9364 @rmrsk
@gtrevisan Thanks for answering that right away!

Ok, I was able to run all the tests locally as well, so consider that box ticked. For the documentation, I was referring to some files that could not be found in the webpages, particularly some of those that lie under disruption_py/core. Since this doesn't mean anything substantial for the review, I will tick this box anyways. If these files were left out of the generated documentation by accident, at least you now know about it...

The only major item left on my list is the improvement on the Summary section, which should be a quick fix. A better example that provides a bit more depth than the current ones would be beneficial, but not a mandatory requirement according to JOSS guidelines.
