### 2026-03-03 #9364 @jatkinson1000
Glad to see this is progressing, just a note @gtrevisan to say that if @rmrsk needed to ask for these instructions they need providing clearly as part of the documentation, not just on this thread, before the box can truly be ticked.
Assuming they're not already there and @rmrsk just missed them I'd suggest opening an issue (linking back to this review) to log this.
