### 2026-03-03 MIT-PSFC/disruption-py#524 @rmrsk
Run tests locally (JOSS review)
To comply with JOSS guidelines, disruption-py should include instructions for how to run the test suite locally. See https://github.com/openjournals/joss-reviews/issues/9364#issuecomment-3989280218
