### 2026-03-03 #9364 @LiamPattinson
I've ticked all but two points on my list, with the remaining work being this point raised by @rmrsk:

> The paper looks all right, but I'd suggest refactoring the summary section. Presently, it doesn't actually describe the high-level functionality of the software, as required by the JOSS guidelines. See https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain. I suggest moving the present summary into the statement of need section, and moving paragraph 1 & 3 from the statement section into the summary.

I agree that the summary paragraph would be better suited in the statement of need, and that the functionality of the software should be described in the summary, but otherwise the paper itself looks good to me.

I was able to get the tests to run just fine using `uv`, but was also a little surprised to see `pytest` wasn't installed by `uv sync`. I'd normally expect the `dev` group to be switched on by default and for `uv sync --no-dev` to be necessary if users want to skip it. It might be worth mentioning in the README that developers should run `uv sync --group=dev` to help out those who aren't so familiar with this stuff, but that's a _very_ minor point.

As stated before, I would personally like to see a more in-depth example/tutorial, but I was able to test the basics of the software from the README alone, and I can piece together how to do more complicated things by looking at the `examples/` directory and the API docs, so I'll consider that box ticked.

In summary, beyond the issue with the paper, all looks ready to me.
