### 2026-03-06 #9364 @gtrevisan
@rmrsk @LiamPattinson
I think my latest addition to the readme (#525) satisfies your comments about dev/deps, but please let me know otherwise.

@jatkinson1000
I already adjusted the paper, but I am currently waiting for coauthors' approval, so I will push the final version into the JOSS branch early next week at the latest.
also, we still haven't heard back from @jackleland...
