### 2026-03-09 #9364 @LiamPattinson
I've just read the new draft of the paper, and it's all looking solid to me, so that's also my checklist finished.

Thanks to @gtrevisan for responding to all of our feedback, and to @jatkinson1000 and @rmrsk for your work on the review!
