### 2026-03-11 #9364 @jatkinson1000
Hi @jackleland and @gtrevisan

JOSS requires 2 reviews to pass, so at this point you do not need to complete a review unless you want to @jackleland
3 reviews is preferred - as more comprehensive and in case people disappear mid-review - but not required.

Please let me know how you wish to proceed.

@gtrevisan I will work through this on Friday when I have time set aside for detailed JOSS work (editing is also voluntary and I cannot do everything as soon as they become available from authors/reviewers). However, you may wish to read through the post-review steps for authors (creating a release and archive etc.) to ensure that you are ready to do this when required.
