### 2026-03-11 #9364 @gtrevisan
@jackleland any further review is welcome, but please let me know your intentions that so I can hold off any post-review action.

@jatkinson1000 I can easily publish a new release in the next couple of days, as usual.

a couple of questions, feel free to answer when you have some spare time:
- may I assume that the DOI assigned to this review is final? `10.21105/joss.09364`
- may I push some README changes about the (imminent) acceptance of the JOSS paper in the repo itself just before tagging the new release?
- should I do anything at all about the `glt/joss` paper branch? should I rebase it off of the new `main` once I release the final version?
- when providing the archive DOI, may I use the version-agnostic one, or should I point at the upcoming release?
