### 2026-03-11 #9364 @jatkinson1000
> * may I assume that the DOI assigned to this review is final? `10.21105/joss.09364`

I think it should be, but cannot guarantee for sure. General advice would be to wait before using, and make any addition of badges/references a follow-up commit post-publication

> * may I push some README changes about the (imminent) acceptance of the JOSS paper in the repo itself just before tagging the new release?

Provided they are not major or code changes that would be fine - we accept that sometimes there are additions to be made to README/CHANGELOGs etc. before a final release can be created and tagged for the codebase.

> * should I do anything at all about the `glt/joss` paper branch? should I rebase it off of the new `main` once I release the final version?

This is somewhat up to you.
Some authors want the paper to appear in the repository so merge it, others leave it in a branch.
The only thing to say is that if you do choose to merge it it will cause some confusion in the review thread and I'll need to update the "branch with paper", so let me know if you do this.

To clarify here - the reviewers have been reviewing the main branch available from PyPI, so this is what the release archive should reflect. If you wish to rebase the paper branch so that those changes also appear in the paper branch that is fine (and I would agree, sensible).

> * when providing the archive DOI, may I use the version-agnostic one, or should I point at the upcoming release?

The archive for JOSS must be the specific version/release that was reviewed and passed by the reviewers including any additions they made/suggested to meet the review criteria.
The author list should match that on the paper.
See [here](https://joss.readthedocs.io/en/latest/editing.html#post-review) for more information.

The easiest way, if acceptable to you, is to link your repository to zenodo so that each release automatically creates a new archive that can be pointed to specifically, but is also tied to future releases (so somewhat "version-agnostic").
If using other approaches you may need to create a specific instance for the reviewed version of the code.

Note that there may be a couple of minor alterations as part of the editorial/typing process.
