### 2026-03-11 #9364 @gtrevisan
**wrt the paper branch:**
I do _not_ wish to merge `glt/joss`, but I can easily rebase it off of the upcoming new `main` so that the code within that branch is updated following what will be archived on Zenodo.
still, what happens to that branch after publication? may it be deleted at some point, since it will stay unmerged forever?

**wrt the Zenodo DOI:**
what I mean is this, taken from our current Zenodo release versioning:

at the time of publication, the "cite all versions" DOI [10.5281/zenodo.13935223](https://doi.org/10.5281/zenodo.13935223) _will automatically redirect_ to v0.14, and it will keep redirecting there until I release v0.15.
normally I very much prefer the version-agnostic DOI, so that new users will be redirected to the updated and current version, even if they are visiting from a very old link -- sometimes people might miss the top banner that Zenodo puts in place for outdated versions.
if required, of course, I will use for JOSS the version-specific DOI that will be assigned to me upon release.

**wrt the third review:**
I'm happy to expedite this towards acceptance and publication, given my Jul 2025 submission, but once again -- @jackleland please stop me and let me know if you would like to contribute your review!
