### 2026-03-11 #9364 @gtrevisan
(apologies for the double post.)

out of clarity, I plan to merge v0.14:
- https://github.com/MIT-PSFC/disruption-py/pull/517

which includes a tiny documentation PR following the above discussions:
- https://github.com/MIT-PSFC/disruption-py/pull/525

but which also includes a handful of bug fixes and one small feature improvement:
- https://github.com/MIT-PSFC/disruption-py/pull/520

alternatively, I could merge **only the tiny documentation PR as a patch release**, ie v0.13.1 -- please LMK in case this is preferred to minimize diffs between already-reviewed and to-be-archived code.
