### 2026-03-13 #9364 @jatkinson1000
Thanks @jackleland are you still on track to complete a review?
I'd like to move this forward next week so if you could complete it by Monday that would be great.

If that's too soon let me know, we'll proceed with 2 reviewers and you can devote efforts towards a future JOSS paper!
