### 2026-03-16 #9364 @jackleland
Okay, I've finished my review. Overall everything looks great and I was able to get the tests running on MAST data. I only have some very minor points:
1. I spent a bit of time trying to install this and get it working on my Mac, but realised that MDSPlus doesn't support it. It doesn't actually say anywhere which OSs are supported so it might be worth explicitly mentioning that somewhere.
2. Installing MDSPlus is a little more involved (I don't have access to a system with it on), so a dockerfile/devcontainer with dependencies pre-installed would make local testing and contributing easier. This could also be distributed as a singularity container if HPC is also the main usage scenario.
3. There's not much info availble—bar 'make a PR'—if you want to contribute, so might be good to have a bit more detail in a CONTRIBUTING.md about how to get set up and what code standards to adhere to, though the extensive CI does get around the latter point. In addition to this, having a PULL_REQUEST_TEMPLATE.md would also standardise contributions.

That said I know I have got to this late so these are all pretty minor points, and should be viewed more as suggestions for future improvements than requirements for JOSS approval.
