### 2026-03-16 #9364 @gtrevisan
@jackleland let me try and reply quickly here:

1. MDSplus is a _pre-requisite_ for MDSplus machines, so I cannot spend time documenting its installation:

    - MDSplus rarely needs to be _installed_ as MDSplus workflows should execute on the machines themselves, very close to the data, for obvious performance reasons;
    - MDSplus has extensive support, including for older systems, but as for any client-server software infrastructure some care needs to be placed in making sure the client and server versions are compatible, which might be tricky if running a freshly-installed client version with some server which hasn't been updated since the 90s;
    - MDSplus supports generic macOS machines but yes, AFAIK [compatibility issues are expected for Apple chips](https://www.mdsplus.org/index.php/Latest_Macintosh_Distributions);
    - for cases like this I've included a python-only MDSplus driver, called `mdsthin`, in the `dev` dependency group, but I am not publicizing it as I don't think it's as robust as a full-fledged MDSplus installation -- in a sense, that solves any installation problem without the user being aware of it.

2. due to the above reasons, I don't believe that containers would be a good investment right now:

    - as mentioned, the proper way to satisfy the MDSplus prerequisite for MDSplus machines is using whatever is made available by IT on such systems;
    - HPC systems are routinely used for hardcore computations, but are rarely deploying full-scale access to raw data capable of supporting truly HPC I/O operations -- with the only exception of the mentioned [TokSearch project](https://ga-fdp.github.io/toksearch/) which AFAIK has some serious hardware behind it _only_ on the DIII-D cluster;
    - according to the systems I can access, so far containerization of HPC workflows has spotty support and even rarer adoption.

3. that is something that I already evaluated, but decided that it's much better to wait and see what new contributions might pop up first, before thinking about standardizing them. due to the smallish scale of our project, I imagine any possible PR will be carefully reviewed separately in an iterative fashion, as we normally do. once I find myself asking the same questions in either issues or PRs, yes, I will definitely push a template to guide discussion from the start.

@jatkinson1000 LMK if you agree that I may release v0.14.0 according to [my plan from a few messages above](https://github.com/openjournals/joss-reviews/issues/9364#issuecomment-4040783381)! thanks.
