### 2026-03-16 #9364 @jatkinson1000
@gtrevisan and @jackleland

**MDSplus**
I think the suggestion of noting that there are issues with running the code and its dependencies on AppleSilicon is good. It is not unlikely that some users will have this operating system/hardware so noting up front that they should seek other resources to run it will help. E.g. it would have prevented @jackleland from trying to install it as described above. Whilst not publiscising `mdsthin` to generic users is fine, making it clear to MacOS AppliSilicon users that they can use the software with this would be useful. Perhaps an "operating systems" page can be added to the documentation, and/or a note to the README. Once this is done @jackleland can check to see that they are happy with this by following these instructions to get it to run on their machine.

**Containers**
Containers might be a nice solution to some workflows/issues, but I agree is not a requirement for publication here and now.

**Contribution Guidelines**
JOSS requires submissions to be freely open for contribution, and for there to be guidelines for users to do so. All three reviewers have checked off **Community guidelines** on the checklists, so this is not a barrier to publication.
As an advocate for Open-Source I would say that adding detailed guidelines can occur early in the project and always be changed/adapted later. It may make it easier for users to contribute. I reiterate that this is a personal note and not a blocker to publication.
