### 2026-03-16 #9364 @gtrevisan
the way I see it, this is more like a project depending on CUDA not making it clear for users that a NVIDIA GPU is required... which could be true, but would probably be self-evident for most people.
then again, I've used a Macbook for years during my postdoc in Fusion, and of course I've met countless others, but I do not recall ever seeing anyone trying to install MDSplus, precisely because people just relied on the preinstalled server/cluster packages.
in any case, I believe the new "devices table" clearly shows that the generic user will **_not_** require MDSplus -- only someone who is authorized to connect to MDSplus for some device will choose that route, and will probably be knowledgeable enough to simply skip any instruction presented in our repo.

all the other points are duly noted, but at this time I am sure they will not prevent anybody from contributing!
... complication of the subject matter might! 😄

unless further actions are required to complete the review, I would be happier to push this forward as soon as possible.
LMK! thanks!
