### 2026-03-18 #9364 @jatkinson1000
@gtrevisan CUDA is a widespread and well-known libraryand I agree when listed as a dependency users will usually know whether they are able to use software on their machine or not. I do not feel it is reasonable to compare knowledge of whether your machine supports CUDA to knowing whether your machine will support MDSPlus.

Further, libraries for which CUDA is optional will clearly outline alternative installation procedures for those without the ability to use CUDA on their system, e.g. PyTorch.
You have indicated that there is a workaround in the MDSPlus case by using `mdsthin`.

@jackleland encountered an issue whilst attempting to evaluate the software according to its installation instructions and experienced difficulty. You responded to this to say that this is a known limitation, and that there is a workaround implemented.
The core of the JOSS review is to assess the usability of the software, that it does what it claims to, and make suggestions to improve the experience for other users.
It has been suggested that you add documentation to highlight known incompatibilities and possible workarounds so that future users do not encounter the same issues as @jackleland.

I do not think this is a particularly strenuous or Herculean request.
Please make the changes and ask @jackleland for feedback as to whether they feel it resolves the difficulties they encountered. Once they are happy then we can proceed with wrapping this up.
