### 2026-03-18 #9364 @gtrevisan
@jackleland
may I ask why you felt the need to install MDSplus in the first place?
didn't the [Devices table](https://github.com/MIT-PSFC/disruption-py?tab=readme-ov-file#devices), wrt both the Data and Access columns, make it clear that for MAST -- which is the only open-access workflow available -- the MDSplus data access was not necessary?

in any case, please LMK whether this is sufficient:
- https://github.com/MIT-PSFC/disruption-py/pull/529
