### 2026-03-19 #9364 @jackleland
A fair question, I was going off of information in INSTALL.md:

> A private installation is possible on any machine, either on a personal laptop, or on any experimental cluster.

And then further down

> As Free/Libre and Open-Source Software (FLOSS), disruption-py can also be installed on any machine. We currently provide an installation guide for Ubuntu-based boxes, but generic Unix machines or Windows systems should support similar or equivalent steps.

So no I didn't need MDSPlus to use the open dataset in this case, but a user who did want to use one of the MDSPlus datasets from an apple silicon Mac would have encountered this problem.

Extending the PR to include clarifications in INSTALL.md as well would be good.
