### 2026-03-19 027734e
rephrase instructions about prereq in INSTALL.md
