### 2026-03-19 #9364 @gtrevisan
@jackleland LMK if this works for you: [whole PR](https://github.com/MIT-PSFC/disruption-py/pull/529/changes) / [changes to INSTALL](https://github.com/MIT-PSFC/disruption-py/pull/529/commits/027734eea42d713755c8a5834492eea5e89e06b8)
