### 2026-03-20 #9364 @gtrevisan
as discussed, I've merged and released v0.14:
- https://github.com/MIT-PSFC/disruption-py/pull/517

version-agnostic Zenodo DOI is still:
- https://doi.org/10.5281/zenodo.13935223

while the specific DOI for v0.14 is:
- https://doi.org/10.5281/zenodo.19051387

@jatkinson1000 please LMK whether I can expedite further steps...
maybe by providing these copy/paste commands for you, with @ signs removed?

```
editorialbot set v0.14.0 as version
```
```
editorialbot set 10.5281/zenodo.19051387 as archive
```
