### 2026-03-21 #9364 @jatkinson1000
Thank you @jackleland

@gtrevisan as previously discussed, editorial and review roles for JOSS are entirely voluntary.
Now that all reviewers have confirmed that they are happy I will work through the acceptance process but am comfortable with the procedures and use of editorialbot, thank you, and there are now some editorial checks before that point.
My last comment on this review was 3 days ago. Your patience is appreciated, particularly given that we had to wait longer than usual early in the process for the authors to bring the software up to the standard expected for a JOSS submission.
