### 2026-03-21 #9364 @jatkinson1000
@gtrevisan I have read the paper and have a number of corrections that need making, listed below. Please make these changes in your paper branch, re-render the paper, and let me know when this is done so I can check them off.

It is now the weekend and I will then be away at an event early next week, so will aim to revisit this after that once the changes have been made.

For line references I am referring to the latest render of the paper [here](https://github.com/openjournals/joss-papers/blob/joss.09364/joss.09364/10.21105.joss.09364.pdf).

- [x] Several citations use Rea, Cristina rather than Rea, C. and this is reflected in the in-text citations. Please fix all occurrences.
- [x] There is a mixture of name formatting throughout the references that is reflected as inconsistencies throughout the text - e.g. "Stillerman et al., (2025)" is rendering as "J. Stillerman et al., (2025)", similarly Stillerman et al (1997) renders with a "J. A. Stillerman".
- [x] Similar issues exist with "C. Rea" and "Cristina Rea" appearing in text rather than "Rea".
- [x] Pau et al 2023 should read "EUROfusion". Please also check the rest of reference.
- [x] Montes et al. (2019) the final D in "DIII-D" should be capitalised. And please check the rest of this reference, sp "Alcator C-Mod".
- [x] MDSPlus is currently rendered `<Scp>MDS</scp>plus` in Stillerman (1997). In BibTeX you need `{MDS}plus` to render this correctly.
- Given the multiple issues above, and not being a domain specialist, please double all references to ensure that they have correct formatting and capitalisation.

- [x] Line 12 abbreviations AI and ML need defining.
- [x] Line 48 "Fusion Community" should not be capitalised, unless it refers to a specific organisation.
- [x] line 60 should read "examples", not "example".
- [x] Line 67, I think "developed" should be removed.
