### 2026-03-21 #9364 @gtrevisan
I am not sure how some of this citation formatting troubles happened.

for example, Humphreys 2020 has full names for all authors in the bib file yet they are correctly truncated to "last, f." in the pdf.
how come Rea's weren't? maybe the style in use allows them expanded for only 2 authors (Rea & Granetz)? or maybe it chooses to use the first initial with short last names (Rea, Zhu)?
I looked again for references style guides but I [could not find much](https://joss.readthedocs.io/en/latest/paper.html#citations)...

some other issues for sure came from my use of: https://doi2bib.org ... for example, that `<scp>`.
anyways, I hope I've fixed all of them.

> Several citations use Rea, Cristina rather than Rea, C. and this is reflected in the in-text citations. Please fix all occurrences.

https://github.com/MIT-PSFC/disruption-py/commit/7ed9031ff6e1a4e726c93f7680ef7e52e9ee0f77

>  There is a mixture of name formatting throughout the references that is reflected as inconsistencies throughout the text - e.g. "Stillerman et al., (2025)" is rendering as "J. Stillerman et al., (2025)", similarly Stillerman et al (1997) renders with a "J. A. Stillerman".

https://github.com/MIT-PSFC/disruption-py/commit/cee9fd511681c8fef1d497898bca9cc1bcdad34d

> Pau et al 2023 should read "EUROfusion". Please also check the rest of reference.

https://github.com/MIT-PSFC/disruption-py/commit/66fff2557314752ade258db32ade70ebf3e66077

>  Montes et al. (2019) the final D in "DIII-D" should be capitalised. And please check the rest of this reference, sp "Alcator C-Mod".

https://github.com/MIT-PSFC/disruption-py/commit/54f06a5504efb0980fce4b2029bd72dc7195203f

> MDSPlus is currently rendered <Scp>MDS</scp>plus in Stillerman (1997). In BibTeX you need {MDS}plus to render this correctly.

https://github.com/MIT-PSFC/disruption-py/commit/e50c53f9caff04d64885da6042ea027890793e68

> Line 12 abbreviations AI and ML need defining.

https://github.com/MIT-PSFC/disruption-py/commit/269ee5fc4e33d39c87312435a5b49deecbd7fe98

> Line 48 "Fusion Community" should not be capitalised, unless it refers to a specific organisation.

https://github.com/MIT-PSFC/disruption-py/commit/098bf1f680dc0b94a01b81f097a666d40c9c6528

> line 60 should read "examples", not "example".

absolutely, not sure how I missed that. thanks!
https://github.com/MIT-PSFC/disruption-py/commit/ddd8f3e4e3c00341f9228e177d3bcc5be62a21c8

> Line 67, I think "developed" should be removed.

absolutely, not sure how I missed that. thanks!
https://github.com/MIT-PSFC/disruption-py/commit/4a55381bd13f5f86a99eb78f5b09aab51d42a2e1
