### 2026-03-21 #9364 @gtrevisan
it used to be "Stillerman, J. A.": after I removed the middle initial it became "Stillerman, Joshua", not "Stillerman, J.", so I've pushed another commit. not sure how to proceed.
@jatkinson1000 as a nuclear option, would you like for me to remove all first names?
