### 2026-03-21 #9364 @jatkinson1000
@gtrevisan I have made an effort to fix this locally for you and opened a pull request at https://github.com/MIT-PSFC/disruption-py/pull/532

There are currently some conflicts as you made the above changes after I forked, so you may need to cherry-pick, rebase, or reset to get my changes.

The issue, I think, is to do with the fact that you had inconsistent representation of the same author across multiple citations who was then used multiple times in the same citation (e.g. Stillerman now renders like the attached image for me instead of a combination of J A and J).
$\LaTeX$/pandoc does this, I believe, as there could be different authors with the same initials and it wants to avoid mixing them.
