### 2026-03-21 #9364 @gtrevisan
sorry, just to be clear -- are you aware that those naming differences are actually due to the original publications, not me botching up initials?
I don't think I'm allowed to "fix metadata" just because it's prettier... am I?

anyways, I appreciate the help! I'll definitely have to wait to get back to the office for conflict resolution between the different bib files, as I have exhausted my weekend time for this, too. next week, then, I'll follow your cues. thanks!
