### 2026-03-22 #9364 @gtrevisan
thank you for your PR! I've pushed another couple of commits on top of that:
- `reapply fixes to paper.md` @ https://github.com/MIT-PSFC/disruption-py/commit/d55d3572747d1776b35c4596dfdcb7be1fccc978
- `reapply fixes to paper.bib` @ https://github.com/MIT-PSFC/disruption-py/commit/bd66048e8eedc518b4a3d880f3c8c854035c3ce3

hopefully all of your checklist items will be fixed for good! 💪
