### 2026-03-27 #9364 @jatkinson1000
> sorry, just to be clear -- are you aware that those naming differences are actually due to the original publications, not me botching up initials? I don't think I'm allowed to "fix metadata" just because it's prettier... am I?

Google scholar BibTeX is often riddled with errors I find - and sometimes authors are inconsistent with initials when they submit to journals. The ideal situation would be that citations match what is printed on the publication itself, but in this day and age provided that the associated DOI is correct that is the main thing, and any discrepancies in initials are OK (provided the correct format is used - i.e. no extended first names.)

I'll check through your changes this afternoon when I have time set aside for JOSS editorial tasks and then hopefully mode this towards EiC approval.
