### 2026-03-27 #9364 @jatkinson1000
@gtrevisan thank you for making those changes and this is now close.

However, the Author list on the Zenodo archive needs to match the JOSS author list for the paper in terms of both authors (down to spelling and initials) and affiliations (again, in spelling and format). This is not currently the case so please could you edit the entry for v0.14.0 to do so.

Similarly, the version is listed on Zenodo as `0.14.0`, but on GitHub and here as `v0.14.0`. Please update the versioning label on Zenodo to match that of elsewhere.
