### 2026-03-27 #9364 @gtrevisan
could you clarify what's missing? do you mean middle name vs middle name initials, and so on?
I'll add the `v` prefix to all Zenodo releases.
