### 2026-03-27 #9364 @jatkinson1000
Your JOSS paper lists 7 authors as well as the group whilst the v0.14.0 archive lists 2 authors and the group, with the rest under project members.

You do not need to amend every other release for this, but the v0.14.0 archive author list should match the paper author list.

And yes, middle name/middle name initials and affiliations should match the JOSS paper exactly as well please.
