### 2026-03-27 #9364 @gtrevisan
- https://github.com/MIT-PSFC/disruption-py/commit/23e466250bf6e6be59f86a2750cc49b83e68047d
- https://zenodo.org/records/19051387
